@@ -455,13 +455,32 @@ class CompositeUnit(UnitBase):
         unit_name._latex."""
 
         if self._scale != 1:
-            x = ("%.2e" % self._scale).split('e')
-            s = x[0]
-            ex = x[1].lstrip('0+')
-            if len(ex) > 0 and ex[0] == '-':
-                ex = '-' + (ex[1:]).lstrip('0')
-            if ex != '':
-                s += r"\times 10^{" + ex + "}"
+            log10_scale = np.log10(abs(self._scale))
+
+            # if we can compactly express the scale without exponents, do so
+            force_no_exponent = str(self._scale).endswith(".0") and log10_scale < 4
+
+            if (log10_scale < 2.5 and log10_scale > -1.0) or force_no_exponent:
+                if log10_scale > 2.0:
+                    s = "%.1f" % self._scale
+                elif log10_scale > 1.0:
+                    s = "%.2f" % self._scale
+                elif log10_scale > 0.0:
+                    s = "%.3f" % self._scale
+                else:
+                    s = "%.4f" % self._scale
+
+                while "." in s and s.endswith("0") or s.endswith("."):
+                    s = s[:-1]
+            else:
+                x = ("%.2e" % self._scale).split('e')
+                s = x[0]
+                ex = x[1].lstrip('0+')
+                if len(ex) > 0 and ex[0] == '-':
+                    ex = '-' + (ex[1:]).lstrip('0')
+                if ex != '':
+                    s += r"\times 10^{" + ex + "}"
+
         else:
             s = ""
 
@@ -753,9 +772,9 @@ def takes_arg_in_units(*args, **orig_kwargs):
 
         @takes_arg_in_units((2, "Msol"), (1, "kpc"), ("blob", "erg"))
         def my_function(arg0, arg1, arg2, blob=22) :
-           print "Arg 2 is",arg2,"Msol"
-           print "Arg 1 is",arg1,"kpc"
-           print "blob is",blob,"ergs"
+           print("Arg 2 is",arg2,"Msol")
+           print("Arg 1 is",arg1,"kpc")
+           print("blob is",blob,"ergs")
 
 
 
