@@ -13,6 +13,8 @@
 import os
 import sys
 
+import numpy as np
+
 #import sphinx_bootstrap_theme
 
 # If extensions (or modules to document with autodoc) are in another directory,
@@ -32,25 +34,35 @@ sys.path.append(os.path.abspath('sphinxext'))
 extensions = ['sphinx.ext.autodoc',
               'sphinx.ext.doctest',
               'sphinx.ext.todo',
+              'sphinx.ext.napoleon',
               'sphinx.ext.coverage',
-              'sphinx.ext.imgmath',
+              'sphinx.ext.mathjax',
               'sphinx.ext.ifconfig',
               'sphinx.ext.viewcode',
               'matplotlib.sphinxext.mathmpl',
               'matplotlib.sphinxext.plot_directive',
               'sphinx.ext.inheritance_diagram',
-              'numpydoc']
+              'sphinx_copybutton',
+              'numpydoc',
+              'nbsphinx'
+              ]
 
+nbsphinx_input_prompt = 'In [%s]:'
+nbsphinx_output_prompt = 'Out[%s]:'
+nbsphinx_execute = 'never' # the notebook is expensive to evaluate, so we need it pre-evaluated
 
+autosummary_generate = True
 
 ipython_warning_is_error = False
+ipython_savefig_dir = 'plots'
 
-try:
-    import IPython.sphinxext.ipython_console_highlighting
-    extensions+=['IPython.sphinxext.ipython_console_highlighting',
-                 'IPython.sphinxext.ipython_directive']
-except ImportError:
-    extensions+=['ipython_console_highlighting','ipython_directive']
+# for .. plot:: directives
+plot_working_directory = '.'
+
+
+extensions+=['IPython.sphinxext.ipython_console_highlighting',
+             'IPython.sphinxext.ipython_directive']
+
 
 # Add any paths that contain templates here, relative to this directory.
 templates_path = ['_templates']
@@ -64,9 +76,11 @@ source_suffix = '.rst'
 # The master toctree document.
 master_doc = 'index'
 
+globaltoc_collapse = True
+
 # General information about the project.
 project = 'pynbody'
-copyright = '2011-20, pynbody team'
+copyright = '2011-24, pynbody team'
 
 # The version info for the project you're documenting, acts as replacement for
 # |version| and |release|, also used in various other places throughout the
@@ -78,6 +92,9 @@ import pynbody
 version = ".".join(pynbody.__version__.split(".")[:2])
 # The full version, including alpha/beta/rc tags.
 release = pynbody.__version__
+
+# make sure plot module is imported for documenting
+import pynbody.plot
 
 # The language for content autogenerated by Sphinx. Refer to documentation
 # for a list of supported languages.
@@ -114,14 +131,20 @@ pygments_style = 'sphinx'
 #modindex_common_prefix = []
 
 
+
 # -- Options for HTML output ---------------------------------------------------
 
 # The theme to use for HTML and HTML Help pages.  See the documentation for
 # a list of builtin themes.
-html_theme = 'alabaster'
+html_theme = 'sphinx_book_theme'
 
 # set up bootstrap
-html_theme_options = { 'logo': 'logo.svg', 'logo_name': False, 'sidebar_collapse': False}
+html_theme_options = {
+    "repository_url": "https://github.com/pynbody/pynbody",
+    "use_repository_button": True,
+                       }
+
+html_logo = "_static/logo.svg"
 
 """html_theme_path = sphinx_bootstrap_theme.get_html_theme_path()
 
@@ -171,7 +194,16 @@ html_static_path = ['_static']
 #html_use_smartypants = True
 
 # Custom sidebar templates, maps document names to template names.
-#html_sidebars = {}
+# html_sidebars = {'**': ['logo.html', 'fulltoc.html', 'relations.html', 'searchbox.html']}
+
+# html_sidebars = {
+#     '**': [
+#         'about.html',
+#         'navigation.html',
+#         'relations.html',
+#         'searchbox.html',
+#     ]
+# }
 
 # Additional templates that should be rendered to pages, maps page names to
 # template names.
@@ -242,7 +274,6 @@ latex_documents = [
 # Documents to append as an appendix to all manuals.
 #latex_appendices = []
 
-# If false, no module index is generated.
 #latex_domain_indices = True
 
 
@@ -255,4 +286,28 @@ man_pages = [
      ['pynbody team'], 1)
 ]
 
-autodoc_member_order = 'bysource'
+#autodoc_member_order = 'alphabetical'
+#autoclass_content = 'both'
+
+copybutton_copy_empty_lines = False
+copybutton_selector =  "div.highlight > pre"
+copybutton_prompt_text = r">>> |\.\.\. |\$ |In \[\d*\]: | {2,5}\.{3,5}: | {5,8}: "
+copybutton_prompt_is_regexp = True
+copybutton_only_copy_prompt_lines = True
+
+def hide_numpy_methods(app, what, name, obj, skip, options):
+    if hasattr(obj, "__qualname__"):
+        qname = obj.__qualname__
+        if "SimArray" in qname:
+            method_name = qname.split(".")[-1]
+            if method_name in np.ndarray.__dict__:
+                print("Skipping ", qname)
+                return True
+def setup(app):
+    app.connect('autodoc-skip-member', hide_numpy_methods)
+
+import matplotlib
+
+matplotlib.rcParams['savefig.dpi'] = 200
+matplotlib.rcParams['savefig.bbox'] = 'tight'
+matplotlib.rcParams['savefig.pad_inches'] = 0.15
