@@ -3,19 +3,19 @@ import matplotlib.pylab as plt
 import pynbody
 
 # load the snapshot and set to physical units
-s = pynbody.load('testdata/g15784.lr.01024.gz')
+s = pynbody.load('testdata/gasoline_ahf/g15784.lr.01024.gz')
 
 # load the halos
 h = s.halos()
 
 # center on the largest halo and align the disk
-pynbody.analysis.angmom.faceon(h[1])
+pynbody.analysis.angmom.faceon(h[0])
 
 # convert all units to something reasonable (kpc, Msol etc)
 s.physical_units()
 
 # create a profile object for the stars (by default this is a 2D profile)
-p = pynbody.analysis.profile.Profile(h[1].s, vmin =.01, max=50)
+p = pynbody.analysis.profile.Profile(h[0].s, vmin =.01, max=50)
 
 # make the figure and sub plots
 f, axs = plt.subplots(1,2,figsize=(14,6))
@@ -25,11 +25,13 @@ axs[0].plot(p['rbins'],p['density'], 'k')
 axs[0].semilogy()
 axs[0].set_xlabel('R [kpc]')
 axs[0].set_ylabel(r'$\Sigma_{\star}$ [M$_{\odot}$ kpc$^{-2}$]')
+axs[0].set_title("2D stellar density")
 
 # make a 3D density plot of the dark matter (note ndim=3 in the constructor below)
-p = pynbody.analysis.profile.Profile(h[1].d,min=.01,max=50,ndim=3)
+p = pynbody.analysis.profile.Profile(h[0].d,min=.01,max=50,ndim=3)
 
 axs[1].plot(p['rbins'],p['density'], 'k')
 axs[1].semilogy()
 axs[1].set_xlabel('R [kpc]')
 axs[1].set_ylabel(r'$\rho_{DM}$ [M$_{\odot}$ kpc$^{-3}$]')
+axs[1].set_title("3D dark matter density")
