@@ -13,21 +13,10 @@ http://pynbody.github.io/pynbody/
 
 import sys
 
-if sys.version_info < (3,5):
-    raise ImportError("""You are attempting to run pynbody on a version of python prior to 3.5.
-
-Support for python 2 has now ended, and as such pynbody has transitioned to be a fully python-3
-module.
-
-You are strongly recommended to update to python 3.5 or newer. Meanwhile, you can use an older
-version of pynbody (prior to version 1.0), by typing:
-
-pip install 'pynbody<1.0'
-""")
-
-# Note: we need to import the configuration first as it sets up the logging
-#       used in later imports
-from .configuration import config, logger, config_parser # isort:skip
+# We need to import configuration first, so prevent isort from reordering
+# isort: off
+from .configuration import config, logger, config_parser
+# isort: on
 from . import (
     analysis,
     array,
@@ -43,7 +32,6 @@ from . import (
     transformation,
     util,
 )
-from .snapshot import ascii, gadget, gadgethdf, grafic, nchilada, ramses, tipsy
 
 
 # The PlotModuleProxy serves to delay import of pynbody.plot until it's accessed.
@@ -85,10 +73,8 @@ plot = PlotModuleProxy()
 
 from .snapshot import load, new
 
-configuration.configure_snapshot_and_halo_loading_priority()
+derived_array = snapshot.simsnap.SimSnap.derived_array
 
-derived_array = snapshot.SimSnap.derived_quantity
-
-__version__ = '1.4.1'
+__version__ = '2.0.0-beta.11'
 
 __all__ = ['load', 'new', 'derived_array']
