@@ -1,3 +1,12 @@
+"""Tools for scientific analysis with pynbody
+
+This sub-package contains a number of modules that enable scientific analysis beyond the basic capabilities of pynbody.
+The most essential tools are imported into the sub-package itself, so that they can be accessed directly from
+pynbody.analysis. For example, :meth:`pynbody.analysis.halo.center` is accessible as :meth:`pynbody.analysis.center`,
+while :meth:`pynbody.analysis.angmom.faceon` is accessible as :meth:`pynbody.analysis.faceon`.
+
+"""
+
 from . import (
     cosmology,
     halo,
@@ -11,5 +20,7 @@ from . import (
     ramses_util,
     theoretical_profiles,
 )
-from .decomp import decomp
+from .angmom import faceon, sideon
+from .halo import center
 from .hmf import halo_mass_function
+from .profile import Profile
