@@ -1,32 +1,30 @@
 """
+Routines related to halo mass function and other cosmological calculations.
 
-halo mass function (hmf)
-========================
-
-Various halo mass function routines.
+.. warning ::
+    These routines were implemented in 2012 when there were no other python libraries that could do this.
+    Since then, cosmology-focussed libraries such as `hmf <https://hmf.readthedocs.io/en/latest/index.html>`_,
+    `Colossus <https://bdiemer.bitbucket.io/colossus/index.html>`_ and
+    `CCL <https://ccl.readthedocs.io/en/latest/>`_ have been developed.
+    For precision cosmology applications, we recommend using these libraries, since the routines here
+    have not been tested for precision applications (and are known to use approximations).
 
 """
 
+import copy
+import logging
+import math
 import os
+import warnings
 
 import numpy as np
+import scipy
+import scipy.interpolate
 
 import pynbody
 
-from .. import units, util
+from .. import configuration, units, util
 from . import cosmology
-
-try:
-    import scipy
-    import scipy.interpolate
-except ImportError:
-    pass
-
-import logging
-import math
-import subprocess
-import tempfile
-import warnings
 
 logger = logging.getLogger('pynbody.analysis.hmf')
 
@@ -36,9 +34,7 @@ logger = logging.getLogger('pynbody.analysis.hmf')
 #######################################################################
 
 class FieldFilter:
-
-    def __init__(self):
-        raise RuntimeError("Cannot instantiate directly, use a subclass instead")
+    """Represents a filter acting on a field"""
 
     def M_to_R(self, M):
         """Return the mass scale (Msol h^-1) for a given length (Mpc h^-1 comoving)"""
@@ -50,11 +46,12 @@ class FieldFilter:
 
     @staticmethod
     def Wk(kR):
-        raise RuntimeError("Not implemented")
+        """Return the Fourier-space filter function, as a function of kR where R is the scale of the filter"""
+        raise NotImplementedError("Not implemented")
 
 
 class TophatFilter(FieldFilter):
-
+    """A top-hat filter in real space"""
     def __init__(self, context):
         self.gammaF = 4 * math.pi / 3
         self.rho_bar = cosmology.rho_M(context, unit="Msol Mpc^-3 h^2 a^-3")
@@ -65,6 +62,7 @@ class TophatFilter(FieldFilter):
 
 
 class GaussianFilter(FieldFilter):
+    """A Gaussian filter"""
 
     def __init__(self, context):
         self.gammaF = (2 * math.pi) ** 1.5
@@ -76,6 +74,7 @@ class GaussianFilter(FieldFilter):
 
 
 class HarmonicStepFilter(FieldFilter):
+    """A step filter in harmonic space"""
 
     def __init__(self, context):
         self.gammaF = 6 * math.pi ** 2
@@ -90,20 +89,49 @@ class HarmonicStepFilter(FieldFilter):
 #######################################################################
 
 
-class PowerSpectrumCAMB:
+class PowerSpectrum:
+    """A power spectrum object, which can be called to return the power spectrum at a given wavenumber."""
 
-    def __init__(self, context, filename=None, log_interpolation=True):
-        if filename is None:
-            warnings.warn(
-                "Using the default power-spectrum spectrum which assumes ns=0.96 and WMAP7+H0+BAO transfer function.", RuntimeWarning)
-            filename = os.path.join(os.path.dirname(__file__), "CAMB_WMAP7")
+    def __init__(self, context, filename=None, k=None, Pk=None, log_interpolation=True):
+        """Set up a power spectrum object.
 
-        k, Pk = np.loadtxt(filename, unpack=True)
+        Either a tabulated power spectrum or the filename of a tabulated power spectrum can be passed.
+        If neither is provided, a default power spectrum is used which assumes Planck 2018 values.
+
+        Parameters
+        ----------
+
+        context : SimSnap
+            The simulation snapshot from which to pull the cosmological context. This is used for
+            the linear growth factor and sigma8 normalization.
+
+        filename : str, optional
+            The name of a file containing the power spectrum.
+
+        k : array, optional
+            The wavenumbers of the power spectrum in comoving h/Mpc
+
+        Pk : array, optional
+            The power spectrum in Mpc^3 h^-3
+
+        log_interpolation : bool, optional
+            If True, interpolate in log space. This is generally more accurate for power spectra.
+
+        """
+
+        if k is None or Pk is None:
+            if filename is None:
+                warnings.warn(
+                    "Using the default power-spectrum spectrum which assumes Planck 2018 values: "
+                    "Omega_c h^2 = 0.120; Omega_b h^2 = 0.0224; h = 0.676; n_s = 0.9667. Sigma8 will be correctly scaled to the simulation value. ",
+                    RuntimeWarning)
+                filename = os.path.join(os.path.dirname(__file__), "CAMB_Planck18")
+            k, Pk = np.loadtxt(filename, unpack=True)
 
         self._orig_k_min = k.min()
         self._orig_k_max = k.max()
 
-        bot_k = 1.e-5
+        bot_k = 1.e-4
 
         if k[0] > bot_k:
             # extrapolate out
@@ -113,7 +141,7 @@ class PowerSpectrumCAMB:
             k = np.hstack((bot_k, k))
             Pk = np.hstack((Pkinterp, Pk))
 
-        top_k = 1.e7
+        top_k = 1.e4
 
         if k[-1] < top_k:
             # extrapolate out
@@ -127,11 +155,11 @@ class PowerSpectrumCAMB:
         self.k = k.view(pynbody.array.SimArray)
         self.k.units = "Mpc^-1 h a^-1"
 
-        self.Pk = Pk.view(pynbody.array.SimArray)
-        self.Pk.units = "Mpc^3 h^-3"
+        self.Pk_z0_unnormalised = Pk.view(pynbody.array.SimArray)
+        self.Pk_z0_unnormalised.units = "Mpc^3 h^-3"
 
         self.k.sim = context
-        self.Pk.sim = context
+        self.Pk_z0_unnormalised.sim = context
 
         self._lingrowth = 1
 
@@ -139,8 +167,8 @@ class PowerSpectrumCAMB:
             self._lingrowth = cosmology.linear_growth_factor(context) ** 2
 
         self._default_filter = TophatFilter(context)
-        self.min_k = self.k.min()
-        self.max_k = self.k.max()
+        self.min_k = self.k.min()*1.000001
+        self.max_k = self.k.max()*0.999999
         self._norm = 1
 
         self._log_interp = log_interpolation
@@ -152,89 +180,128 @@ class PowerSpectrumCAMB:
     def _init_interpolation(self):
         if self._log_interp:
             self._interp = scipy.interpolate.interp1d(
-                np.log(self.k), np.log(self.Pk))
+                np.log(self.k), np.log(self.Pk_z0_unnormalised))
         else:
-            self._interp = scipy.interpolate.interp1d(np.log(self.k), self.Pk)
+            self._interp = scipy.interpolate.interp1d(np.log(self.k), self.Pk_z0_unnormalised)
 
     def set_sigma8(self, sigma8):
+        """Update the normalisation for a given sigma8 value at z=0"""
         current_sigma8_2 = self.get_sigma8() ** 2
         self._norm *= sigma8 ** 2 / current_sigma8_2
 
     def get_sigma8(self):
+        """Calculate the current sigma8 value at z=0"""
         current_sigma8 = math.sqrt(
             variance(8.0, self._default_filter, self, True) / self._lingrowth)
         return current_sigma8
 
     def __call__(self, k):
-        if np.any(k < self._orig_k_min):
-            warnings.warn(
-                "Power spectrum does not extend to low enough k; using power-law extrapolation (this is likely to be fine)", RuntimeWarning)
-        if np.any(k > self._orig_k_max):
-            warnings.warn(
-                "Power spectrum does not extend to high enough k; using power-law extrapolation. This is bad but your results are unlikely to be sensitive to it unless they relate directly to very small scales or you have run CAMB with inappropriate settings.", RuntimeWarning)
+        """Evaluate the power spectrum at the snapshot redshift"""
         if self._log_interp:
             return self._norm * self._lingrowth * np.exp(self._interp(np.log(k)))
         else:
             return self._norm * self._lingrowth * self._interp(np.log(k))
 
 
-class PowerSpectrumCAMBLive(PowerSpectrumCAMB):
+class PowerSpectrumCAMB(PowerSpectrum):
+    """A power spectrum object that calculates the power spectrum on the fly using CAMB.
+
+    This is slower than using a precomputed power spectrum, but is more flexible."""
 
     def __init__(self, context, use_context=True, camb_params={}, log_interpolation=True):
-        """Run CAMB to get out a power spectrum. The default parameters are in cambtemplate.ini.
-        Any of these can be modified by passing the appropriate kwarg."""
+        """Runs CAMB to calculate a power spectrum for the provided context.
 
-        from .. import config_parser
-        path_to_camb = config_parser.get('camb', 'path')
-        if path_to_camb == '/path/to/camb':
-            raise RuntimeError("You need to compile CAMB and set up the executable path in your pynbody configuration file.")
+        Parameters
+        ----------
+        context : SimSnap
+            The simulation snapshot from which to pull the cosmological context.
 
-        file_in = open(
-            os.path.join(os.path.dirname(__file__), "cambtemplate.ini"))
-        folder_out = tempfile.mkdtemp()
-        file_out = open(os.path.join(folder_out, "camb.ini"), "w")
+        use_context : bool, optional
+            If True, use the context to set the cosmological parameters in CAMB.
+            If False, use only the parameters in ``camb_params``.
+
+        camb_params : dict, optional
+            A dictionary of parameters to pass to CAMB's ``set_cosmology`` method. Note that the
+            scalar spectral index ``ns`` is not included in ``set_cosmology``; if provided here, it is
+            extracted and passed to ``InitPower.set_params``.
+
+        log_interpolation : bool, optional
+            If True, interpolate in log space. This is generally more accurate for power spectra.
+
+        """
+
+        try:
+            import camb
+        except ImportError:
+            warnings.warn("The camb module is not installed. Falling back to using a pre-computed power spectrum"
+                          "which may not be appropriate for your cosmology.")
+            super().__init__(context)
+            return
+
+        camb_params_obj = camb.CAMBparams()
+        camb_params = copy.copy(camb_params)
+
+        # default value for ns, in case use_context is false:
+        ns = configuration.config_parser.get('default-cosmology', 'ns')
 
         if use_context:
             h0 = context.properties['h']
-            omB0 = context.properties['omegaB0'] * h0 ** 2
-            omM0 = context.properties['omegaM0'] * h0 ** 2
-            omC0 = omM0 - omB0
+            omBh2 = context.properties['omegaB0'] * h0 ** 2
+            omMh2 = context.properties['omegaM0'] * h0 ** 2
+            if omBh2 == 0.0 and 'ombh2' not in camb_params_obj:
+                warnings.warn("OmegaB0 is zero, presumably because this is a DMO run. For the power spectrum, setting ombh2 to 0.0224 (Planck 2018)."
+                              "To override, set 'omegaB0' in the simulation parameters, or 'ombh2' in the camb_params argument.")
+                omBh2 = 0.0224
+            OmCh2 = omMh2 - omBh2
+
+            if 'H0' not in camb_params:
+                camb_params['H0'] = h0 * 100
+            if 'ombh2' not in camb_params:
+                camb_params['ombh2'] = omBh2
+            if 'omch2' not in camb_params:
+                camb_params['omch2'] = OmCh2
+
             ns = context.properties['ns']
-            running = context.properties['running']
-            camb_params.update({'ombh2': omB0, 'omch2': omC0, 'hubble': h0 *
-                                100, 'scalar_nrun(1)': running, 'scalar_spectral_index(1)': ns})
 
-        for line in file_in:
-            if "=" in line and "#" not in line:
-                name, val = line.split("=")
-                name = name.strip()
-                if name in camb_params:
-                    val = camb_params[name]
-                print(name, "=", val, file=file_out)
-            else:
-                print(line.strip(), file=file_out)
+        if 'ns' in camb_params:
+            ns = camb_params.pop('ns')
 
-        file_out.close()
+        camb_params_obj.set_cosmology(**camb_params)
 
-        logger.info("Running %s on %s" %
-                    (path_to_camb, os.path.join(folder_out, "camb.ini")))
-        subprocess.check_output("cd %s; %s camb.ini" %
-                                (folder_out, path_to_camb), shell=True)
+        camb_params_obj.InitPower.set_params(ns=ns)
 
-        PowerSpectrumCAMB.__init__(self, context, os.path.join(
-            folder_out, "test_matterpower.dat"), log_interpolation=log_interpolation)
+        camb_params_obj.WantTransfer = True
+
+        # always get the matter power spectrum at z=0; growth factor will be scaled later if needed
+        camb_params_obj.set_matter_power(redshifts=[0.0], kmax=1e2, k_per_logint=0,
+                                     silent=True, nonlinear=False)
+        results = camb.get_results(camb_params_obj)
+
+        k, z, Pk = results.get_matter_power_spectrum(minkh=1.e-4, maxkh=1e2, npoints=1000)
+
+        # filename = os.path.join(os.path.dirname(__file__), "CAMB_Planck18")
+        # np.savetxt(filename, np.array([k, Pk[0]]).T)
+
+        assert len(z)==1
+        assert z[0] == 0.0
+
+        super().__init__(context, k=k, Pk=Pk[0], log_interpolation=log_interpolation)
 
 
-class BiasedPowerSpectrum(PowerSpectrumCAMB):
+
+class BiasedPowerSpectrum(PowerSpectrum):
+    """A power spectrum object with linear bias"""
 
     def __init__(self, bias, pspec):
         """Set up a biased power spectrum.
 
-        **args**
-          bias: either a number for a fixed bias, or a function taking
-                the wavenumber k in Mpc/h and returning the bias
+        Parameters
+        ----------
+        bias : float  or function
+            Either a constant linear bias factor, or a function of wavenumber that returns the bias at that wavenumber.
 
-          pspec: the underlying power spectrum
+        pspec : PowerSpectrum
+            The power spectrum object to wrap with a bias
         """
 
         if not hasattr(bias, '__call__'):
@@ -246,7 +313,7 @@ class BiasedPowerSpectrum(PowerSpectrumCAMB):
         self.min_k = pspec.min_k
         self.max_k = pspec.max_k
         self.k = pspec.k
-        self.Pk = pspec.Pk * self._bias(self.k) ** 2
+        self.Pk_z0_unnormalised = pspec.Pk_z0_unnormalised * self._bias(self.k) ** 2
 
     def __call__(self, k):
         return self._norm * self._pspec(k) * self._bias(k) ** 2
@@ -256,47 +323,52 @@ class BiasedPowerSpectrum(PowerSpectrumCAMB):
 # Variance calculation
 #######################################################################
 
-def variance(M, f_filter=TophatFilter, powspec=PowerSpectrumCAMB, arg_is_R=False):
-    if hasattr(M, '__len__'):
+def variance(M_or_R, f_filter=TophatFilter, powspec=PowerSpectrum, arg_is_R=False):
+    """Calculate the variance of the density field smoothed on a mass scale M, or optionally a length scale R.
+
+    Parameters
+    ----------
+    M_or_R : float or array
+        The mass scale in Msol/h or the radius of the filter in Mpc/h comoving. If an array, the
+        variance is calculated for each element of the array.
+
+    f_filter : FieldFilter, optional
+        The filter to use. Default is a top-hat filter.
+
+    powspec : PowerSpectrum, optional
+        The power spectrum object to use. Default is a Planck 2018 power spectrum at z=0.
+
+    arg_is_R : bool, optional
+        If True, interpret the input as a length scale R rather than a mass scale M.
+
+    Returns
+    -------
+    float or array
+        The variance of the density field smoothed on the given scale(s).
+
+    """
+    if hasattr(M_or_R, '__len__'):
         ax = pynbody.array.SimArray(
-            [variance(Mi, f_filter, powspec, arg_is_R) for Mi in M])
+            [variance(Mi, f_filter, powspec, arg_is_R) for Mi in M_or_R])
         # hopefully dimensionless
-        ax.units = powspec.Pk.units * powspec.k.units ** 3
+        ax.units = powspec.Pk_z0_unnormalised.units * powspec.k.units ** 3
         return ax
 
     if arg_is_R:
-        R = M
+        R = M_or_R
     else:
-        R = f_filter.M_to_R(M)
+        R = f_filter.M_to_R(M_or_R)
 
     integrand = lambda k: k ** 2 * powspec(k) * f_filter.Wk(k * R) ** 2
     integrand_ln_k = lambda k: np.exp(k) * integrand(np.exp(k))
-    v = scipy.integrate.romberg(integrand_ln_k, math.log(powspec.min_k), math.log(
-        1. / R) + 3, divmax=10, rtol=1.e-4) / (2 * math.pi ** 2)
+    v =  scipy.integrate.quad(integrand_ln_k, math.log(powspec.min_k), math.log(
+        1. / R) + 3, epsrel=1.e-6)[0] / (2 * math.pi ** 2)
 
     return v
 
 
-def estimate_neffm(M, f_filter=TophatFilter, powspec=PowerSpectrumCAMB, arg_is_R=False):
-    # this routine is opaque and inefficient
-    dlnm = 0.01  # just needs to be small
-
-# if hasattr(M,'__len__') : ## why is this needed?
-#        ax =  pynbody.array.SimArray([estimate_neffm(Mi, f_filter, powspec) for Mi in M])
-# ax.units = powspec.Pk.units/ powspec.Pk.units * powspec.k.units**3 / powspec.k.units**3  # hopefully dimensionless
-#        return ax
-
-    M2 = np.exp(np.log(M) + dlnm)
-    sig = np.sqrt(variance(M, f_filter, powspec))
-    sig2 = np.sqrt(variance(M2, f_filter, powspec))
-    #  neff = 6 dlnsigmainv / dlnm, eq. 13 reed et al. 2007
-    neff = 6. * (np.log(sig / sig2)) / dlnm - 3.
-
-    return neff
-
-
 def get_neffm(mass, sigma):
-            #  neff = 6 dlnsigmainv / dlnm - 3, eq. 13 reed et al. 2007
+    """Calculate the effective spectral index of the power spectrum at a given mass scale."""
     dlnm = np.diff(np.log(mass))
     dlnsigmainv = np.diff(np.log(1. / sigma))
     neff = 6. * dlnsigmainv / dlnm - 3.
@@ -304,11 +376,12 @@ def get_neffm(mass, sigma):
 
 
 @units.takes_arg_in_units((0, "Mpc h^-1"))
-def correlation(r, powspec=PowerSpectrumCAMB):
+def correlation(r, powspec=PowerSpectrum):
+    """Calculate the correlation function of the density field at a specified radius."""
 
     if hasattr(r, '__len__'):
         ax = pynbody.array.SimArray([correlation(ri,  powspec) for ri in r])
-        ax.units = powspec.Pk.units * powspec.k.units ** 3
+        ax.units = powspec.Pk_z0_unnormalised.units * powspec.k.units ** 3
         return ax
 
     # Because sin kr becomes so highly oscilliatory, normal
@@ -408,36 +481,38 @@ def correlation(r, powspec=PowerSpectrumCAMB):
 def correlation_func(context, log_r_min=-3, log_r_max=2, delta_log_r=0.2,
                      pspec=PowerSpectrumCAMB):
     """
-
     Calculate the linear density field correlation function.
 
-    **Args**:
+    Parameters
+    ----------
 
-    *context* (SimSnap): The snapshot from which to pull the
-        cosmological context (includes sigma8 normalization and growth
-        function integrations, but does not currently affect transfer
-        function)
+    context : SimSnap
+        The snapshot from which to pull the cosmological context
 
-    **Kwargs:**
+    log_r_min : float, optional
+        log10 of the minimum separation (Mpc h^-1) to consider
 
-      *log_r_min:* log10 of the minimum separation (Mpc h^-1) to
-       consider
+    log_r_max : float, optional
+        log10 of the maximum separation (Mpc h^-1) to consider
 
-      *log_r_max:* log10 of the maximum separation (Mpc h^-1) to
-       consider
+    delta_log_r : float, optional
+        The value spacing in dex
 
-      *delta_log_r:* The value spacing in dex
+    pspec : PowerSpectrum, optional
+        The power spectrum class to use, or if an instance is provided, the power spectrum to use.
+        The default is to use PowerSpectrumCAMB which will attempt to calculate a power spectrum
+        from the simulation context; if it fails, it will fall back to using Planck2018 power spectrum
+        and issue a warning.
 
-      *pspec:* A power spectrum object; default is a WMAP7 cosmology
-        calculated by CAMB.
+    Returns
+    -------
 
-    **Returns:**
+    r : array
+        The separation values in Mpc h^-1
 
-      *r:* Array of the r values (Mpc h^-1) for which the correlation
-         function was evaluated.
+    Xi : array
+        The correlation function at each separation
 
-      *Xi:* Array of the dimensionless correlation for each
-       separation.
 
     """
 
@@ -461,7 +536,7 @@ def correlation_func(context, log_r_min=-3, log_r_max=2, delta_log_r=0.2,
 #######################################################################
 
 
-def f_press_schechter(nu):
+def _f_press_schechter(nu):
     """
 
     The Press-Schechter kernel used by halo_mass_function
@@ -472,7 +547,7 @@ def f_press_schechter(nu):
     return f
 
 
-def f_sheth_tormen(nu, Anorm=0.3222, a=0.707, p=0.3):
+def _f_sheth_tormen(nu, Anorm=0.3222, a=0.707, p=0.3):
     """
     Sheth & Tormen (1999) fit (see also Sheth Mo & Tormen 2001)
     """
@@ -486,7 +561,7 @@ def f_sheth_tormen(nu, Anorm=0.3222, a=0.707, p=0.3):
     return f
 
 
-def f_jenkins(nu, deltac=1.68647):
+def _f_jenkins(nu, deltac=1.68647):
     # Jenkins et al (2001) fit   ##  valid for  -1.2 << ln(1/sigma) << 1.05
     sigma = deltac / nu
     lnsigmainv = np.log(1. / sigma)
@@ -497,7 +572,7 @@ def f_jenkins(nu, deltac=1.68647):
     return f
 
 
-def f_warren(nu, deltac=1.68647):
+def _f_warren(nu, deltac=1.68647):
     #  Warren et al. 2006  -- valid for (10**10 - 10**15 Msun/h)
     sigma = deltac / nu
     A = 0.7234
@@ -508,7 +583,7 @@ def f_warren(nu, deltac=1.68647):
     return f
 
 
-def f_reed_no_z(nu, deltac=1.68647):  # universal form
+def _f_reed_no_z(nu, deltac=1.68647):  # universal form
     # Reed et al. (2007) fit, eqn. 9 -- with no redshift depedence (simple
     # universal form)
     """ modified S-T fit  by the G1 gaussian term and c"""
@@ -528,7 +603,7 @@ def f_reed_no_z(nu, deltac=1.68647):  # universal form
     return f
 
 
-def f_reed_z_evo(nu, neff, deltac=1.68647):  # non-universal form
+def _f_reed_z_evo(nu, neff, deltac=1.68647):  # non-universal form
     # Reed et al. (2007) fit, eqn. 11 -- with redshift depedence for accuracy
     # at z >~ z_reion
     """ modified S-T fit  by the n_eff dependence and the G1 and G2 gaussian terms and c
@@ -554,7 +629,7 @@ def f_reed_z_evo(nu, neff, deltac=1.68647):  # non-universal form
     return f
 
 
-def f_bhattacharya(nu, red, deltac=1.68647):
+def _f_bhattacharya(nu, red, deltac=1.68647):
     # Bhattacharya et al. 2010  -- 6x10**11 - 310**15 Msun/h  z=0-2
     sigma = deltac / nu
     A = 0.333 / pow((1. + red), 0.11)
@@ -565,12 +640,7 @@ def f_bhattacharya(nu, red, deltac=1.68647):
     f *= np.power(nu * math.sqrt(a), q) * np.exp(-a * nu * nu / 2.)
     return f
 
-
-#######################################################################
-# Bias functions
-#######################################################################
-
-def cole_kaiser_bias(nu, delta_c):
+def _cole_kaiser_bias(nu, delta_c):
     """
 
     The Cole-Kaiser (1989) bias function. Also in Mo & White 1996.
@@ -579,8 +649,8 @@ def cole_kaiser_bias(nu, delta_c):
     return 1 + (nu ** 2 - 1) / delta_c
 
 
-def sheth_tormen_bias(nu, delta_c,
-                      a=0.707, b=0.5, c=0.6):
+def _sheth_tormen_bias(nu, delta_c,
+                       a=0.707, b=0.5, c=0.6):
     """
 
     The Sheth-Tormen (1999) bias function [eq 8]
@@ -593,78 +663,86 @@ def sheth_tormen_bias(nu, delta_c,
                  - (a * nu ** 2) ** c / ((a * nu ** 2) ** c + b * (1 - c) * (1 - c / 2))) \
         / (root_a * delta_c)
 
-#######################################################################
-# The most useful function: halo_mass_function
-#######################################################################
 
 
 def halo_mass_function(context,
                        log_M_min=8.0, log_M_max=15.0, delta_log_M=0.1,
                        kern="ST",
-                       pspec=PowerSpectrumCAMB,
-                       delta_crit=1.686,
-                       no_h=False):
+                       pspec=PowerSpectrum,
+                       delta_crit=1.686):
     """
+    Returns the halo mass function, dN/d log_{10} M in units of Mpc^-3 h^3.
 
-    Returns the halo mass function, dN/d log_{10} M in units of Mpc^-3
-    h^3.
-
-    **Args:**
-
-    *context (SimSnap):* The snapshot from which to pull the
-    cosmological context (includes sigma8 normalization and growth
-    function integrations, but does not currently affect transfer
-    function)
-
-    **Kwargs:**
-
-    *log_M_min:* The minimum halo mass (Msol h^-1) to consider
-
-    *log_M_max:* The maximum halo mass (Msol h^-1) to consider
-
-    *delta_log_M:* The bin spacing of halo masses (see warning below)
-
-    *kern:* The kernel function which dictates what type of mass
-    function to calculate; or a string ("PS" or "ST") for one
-    of the defaults
-
-    *pspec:* A power spectrum object (which also defines the window
-    function); default is a WMAP7 cosmology calculated by
-    CAMB, and a top hat window
-
-    *delta_crit:* The critical overdensity for collapse
-
-    **Returns:**
-
-    *M:* The centre of the mass bins, in Msol h^-1
-
-    *sigma:* The linear variance of the corresponding sphere
-
-    *N:* The abundance of halos of that mass (Mpc^-3 h^3 comoving,
-    per decade of mass)
-
-    Because numerical derivatives are involved, the value of
-    delta_log_M affects the accuracy. Numerical experiments suggest
-    that delta_log_M=0.1 gives more than enough accuracy, but you
-    should check for your own use case.
+    See :ref:`hmf_tutorial` for an example and more information.
 
 
-    Recommended m.f. for friends-of-friends linking length 0.2 particle sep.:
-    z <~ 2 : bhattacharya
-    z >~ 5 : reed_universal (no redshift dependence)
-    : or reed_evolving (w/redshift dependence for additional accuracy)
+    Parameters
+    ----------
+
+    context : SimSnap
+        The snapshot from which to pull the cosmological context. Sigma8 normalization and growth function
+        redshift dependence is always taken into account. If `camb <https://camb.readthedocs.io/en/latest/>`_ is
+        installed, an appropriate power spectrum is calculated for other cosmological parameters (e.g. baryon
+        density, Hubble constant) from the simulation. Otherwise, a default Planck18 power spectrum is
+        used.
+
+    log_M_min : float, optional
+        The minimum halo mass (Msol h^-1) to consider. Default is 8.
+
+    log_M_max : float, optional
+        The maximum halo mass (Msol h^-1) to consider. Default is 15.
+
+    delta_log_M : float, optional
+        The bin spacing of halo masses. Default is 0.1.
+
+    kern : str or function, optional
+        The kernel function which dictates what type of mass function to calculate. Default is "ST" (Sheth-Tormen).
+        Other options are "PS" (Press-Schechter), "J" (Jenkins), "W" (Warren), "REEDZ" (Reed et al 2007 with redshift
+        dependence), "REEDU" (Reed et al 2007 without redshift dependence), "B" (Bhattacharya).
+
+    pspec : PowerSpectrumCAMB, optional
+        A power spectrum object (which also defines the window function), overriding the cosmological
+        parameters of the simulation.
+
+    delta_crit : float, optional
+        The critical overdensity for collapse. Default is 1.686.
+
+    Returns
+    -------
+
+    M : SimArray
+        The centre of the mass bins, in Msol h^-1.
+
+    sigma : SimArray
+        The linear variance of the corresponding sphere.
+
+    N : SimArray
+        The abundance of halos of that mass (Mpc^-3 h^3 comoving, per decade of mass).
+
+    Notes
+    -----
+
+    Because numerical derivatives are involved, the value of ``delta_log_M`` affects the accuracy.
+
+    The halo mass function code in pynbody was implemented in 2012 when there
+    were no other python libraries that could do this.
+    Since then, cosmology-focussed libraries such as `hmf <https://hmf.readthedocs.io/en/latest/index.html>`_,
+    and `CCL <https://halotools.readthedocs.io/en/latest/>`_ have been developed.
+    For precision cosmology applications, we recommend using these libraries.
+
+    The functionality here is retained for quick cross-checks of simulations.
 
     """
 
     if isinstance(kern, str):
-        kern = {'PS': f_press_schechter,
-                'ST': f_sheth_tormen,
-                'J': f_jenkins,
-                'W': f_warren,
-                'REEDZ': f_reed_z_evo,
+        kern = {'PS': _f_press_schechter,
+                'ST': _f_sheth_tormen,
+                'J': _f_jenkins,
+                'W': _f_warren,
+                'REEDZ': _f_reed_z_evo,
                 # Reed et al 2007 without redshift dependence
-                'REEDU': f_reed_no_z,
-                'B': f_bhattacharya}[kern]
+                'REEDU': _f_reed_no_z,
+                'B': _f_bhattacharya}[kern]
 
     rho_bar = cosmology.rho_M(context, unit="Msol Mpc^-3 h^2")
     red = context.properties['z']  # redshift is always set by simulation
@@ -687,12 +765,11 @@ def halo_mass_function(context,
 
     dM = np.diff(10 ** M)
 
-    if (kern == f_reed_z_evo):
-        ##    neff = estimate_neffm(M)
+    if (kern == _f_reed_z_evo):
         neff = get_neffm(10. ** M, sig ** 0.5)
         out = (rho_bar / (10 ** M_mid)) * kern(nu_mid, neff) * \
             d_ln_nu_d_ln_M * math.log(10.) * context.properties['a'] ** 3
-    elif (kern == f_bhattacharya):
+    elif (kern == _f_bhattacharya):
        # eq 7.46, Mo, van den Bosch and White
         out = (rho_bar / (10 ** M_mid)) * kern(nu_mid, red) * \
             d_ln_nu_d_ln_M * math.log(10.) * context.properties['a'] ** 3
@@ -700,6 +777,7 @@ def halo_mass_function(context,
         # eq 7.46, Mo, van den Bosch and White
         out = (rho_bar / (10 ** M_mid)) * kern(nu_mid) * \
             d_ln_nu_d_ln_M * math.log(10.) * context.properties['a'] ** 3
+    out= out.view(pynbody.array.SimArray)
     out.units = "Mpc^-3 h^3 a^-3"
     out.sim = context
 
@@ -714,36 +792,43 @@ def halo_mass_function(context,
 
 
 @units.takes_arg_in_units((1, "Msol h^-1"), context_arg=0)
-def halo_bias(context, M, kern=cole_kaiser_bias, pspec=PowerSpectrumCAMB,
+def halo_bias(context, M, kern="CK", pspec=PowerSpectrumCAMB,
               delta_crit=1.686):
     """
-
     Return the halo bias for the given halo mass.
 
-    **Args:**
+    Parameters
+    ----------
 
-       *context (SimSnap):* The snapshot from which to pull the
-        cosmological context
+    context : SimSnap
+        The snapshot from which to pull the cosmological context
 
-       *M:* float, unit or string describing the halo mass. If a
-        float, units are Msol h^-1.
+    M : float
+        The halo mass in Msol h^-1
 
-    **Kwargs:**
+    kern : str or function, optional
+        The kernel function describing the halo bias. Default is "CK" (Cole-Kaiser).
+        Other options are "ST" (Sheth-Tormen). Alternatively a function can be provided.
 
-       *kern:* The kernel function describing the halo bias (default
-        Cole-Kaiser).
+    pspec : PowerSpectrum, optional
+        The power spectrum class to use, or if an instance is provided, the power spectrum to use.
+        Default is PowerSpectrumCAMB which will attempt to calculate a power spectrum from the simulation context;
+        if it fails, it will fall back to using Planck2018 power spectrum and issue a warning.
 
-       *pspec:* A power spectrum object (which also defines the window
-              function); default is a WMAP7 cosmology calculated by
-              CAMB, and a top hat window
+    delta_crit : float, optional
+        The critical overdensity for collapse. Default is 1.686.
 
-       *delta_crit:* The critical overdensity for collapse
+    Returns
+    -------
 
-    **Returns:**
-
-       The halo bias (single float)
+    float
+        The halo bias for the given halo mass.
 
     """
+
+    if isinstance(kern, str):
+        kern = {'CK': _cole_kaiser_bias,
+                'ST': _sheth_tormen_bias}[kern]
 
     if isinstance(pspec, type):
         pspec = pspec(context)
@@ -754,42 +839,64 @@ def halo_bias(context, M, kern=cole_kaiser_bias, pspec=PowerSpectrumCAMB,
     return kern(nu, delta_crit)
 
 
-def simulation_halo_mass_function(snapshot,
+def simulation_halo_mass_function(snapshot_or_cat,
                                   log_M_min=8.0, log_M_max=15.0, delta_log_M=0.1,
-                                  masses=None, mass_def="halo_finder", calculate_err=True,
+                                  masses=None, mass_property=None, calculate_err=True,
                                   subsample_catalogue=None):
+    """Estimate a halo mass function from a simulation, via binning haloes in mass.
+
+    Parameters
+    ----------
+
+    snapshot_or_cat : SimSnap or HaloCatalogue
+        The snapshot from which to calculate the halo mass function, or the halo catalogue itself.
+        If a snapshot is provided, ``.halos()`` is called to get the first available catalogue.
+
+    log_M_min : float, optional
+        The minimum halo mass (Msol h^-1) to consider. Default is 8.
+
+    log_M_max : float, optional
+        The maximum halo mass (Msol h^-1) to consider. Default is 15.
+
+    delta_log_M : float, optional
+        The bin spacing of halo masses. Default is 0.1.
+
+    masses : array, optional
+        Provide an array of halo masses in Msol h^-1. If None, this is calculated from the snapshot.
+
+    mass_property : str, optional
+        The property name giving the mass of each halo. If None, sums the mass in particles in each halo
+        according to the particle data in the catalogue. Summing masses can be slow for large catalogues.
+
+    calculate_err : bool, optional
+        If True, estimates error bars according to Poisson statistics.
+
+    subsample_catalogue : int, optional
+        Subsample the halo catalogue by the specified factor. Not recommended except for debugging.
+
+    Returns
+    -------
+
+    bin_centers : SimArray
+        The centre of the mass bins, in Msol h^-1.
+
+    num_den : SimArray
+        The binned number density of haloes in this snapshot in comoving Mpc**-3 h**3 per decade of mass.
+
+    err : SimArray
+        The error on the number density, if ``calculate_err`` is True.
+
     """
 
-    Construct the halo mass function from a halo catalogue by binning haloes in mass and counting them.
-    This allows direct plotting of the simulation HMF against the theory HMF.
+    if isinstance(snapshot_or_cat, pynbody.snapshot.SimSnap):
+        snapshot = snapshot_or_cat
+        halo_catalogue = snapshot.halos()
+    elif isinstance(snapshot_or_cat, pynbody.halo.HaloCatalogue):
+        halo_catalogue = snapshot_or_cat
+        snapshot = halo_catalogue.base
+    else:
+        raise TypeError("snapshot_or_cat must be a SimSnap or HaloCatalogue")
 
-    **Args:**
-
-       *snapshot (SimSnap):* The snapshot from which to calculate halo masses
-
-    **Kwargs:**
-
-        *log_M_min:* The minimum halo mass (Msol h^-1) to consider
-
-        *log_M_max:* The maximum halo mass (Msol h^-1) to consider
-
-        *delta_log_M:* The bin spacing of halo masses
-
-        *masses: Provide array of halo masses in Msol h**-1. If none, this is calculated from the snapshot.
-
-        *mass_def: Definition of the mass of a halo. Possible extensions could be M200_crit, M200_matter, Mvir etc...
-
-        *err: If true, calculates error bars according to Poisson process.
-
-        *subsample_catalogue: Sample the halo catalogue every int value. Mostly for speed and debugging issues.
-        WARNING: If your halo catalogue is ordered, this method does not take responsability in randomising the catalogue so
-        you might not recover the true HMF.
-
-    **Returns:**
-
-       The binned number density of haloes in this snapshot in comoving Mpc**-3 h**3 per decade of mass
-
-    """
 
     #Check that the mass resolution is uniform, this method does not handle otherwise
     if len(set(snapshot.d['mass'])) > 1:
@@ -800,18 +907,20 @@ def simulation_halo_mass_function(snapshot,
     bins = np.logspace(log_M_min, log_M_max, num=nbins, base=10)
     bin_centers = (bins[:-1] + bins[1:]) / 2
 
-    if subsample_catalogue is None:
-        halo_catalogue = snapshot.halos()
-    else:
-        halo_catalogue = snapshot.halos()[::subsample_catalogue]
+    halo_catalogue.load_all()
+
+    if subsample_catalogue is not None:
+        halo_catalogue_underlying = halo_catalogue # keep alive for the lifetime of the function
+        halo_catalogue = halo_catalogue[::subsample_catalogue]
 
     if masses is None:
-        warnings.warn("Halo finder masses not provided. Calculating them (might take a while...)")
-
-        if mass_def=="halo_finder":
+        if mass_property is None:
+            warnings.warn("Halo finder masses not provided. Calculating them (might take a while...)")
             masses = np.array([h['mass'].sum().in_units('1 h**-1 Msol') for h in halo_catalogue])
         else:
-            raise KeyError("Only halo finder mass is supported in this calculation for now")
+            masses = halo_catalogue.get_properties_all_halos(with_units=True)[mass_property]
+            if units.has_unit(masses):
+                masses = masses.in_units('1 h**-1 Msol')
 
     if np.amax(masses) > 10**log_M_max or np.amin(masses) < 10**log_M_min :
         warnings.warn("Your bin range does not encompass the full range of halo masses")
@@ -832,16 +941,16 @@ def simulation_halo_mass_function(snapshot,
     # Make sure units are consistent
     bin_centers = bin_centers.view(pynbody.array.SimArray)
     bin_centers.units = "Msol h**-1"
-    bin_centers.context = snapshot
+    bin_centers.sim = snapshot
 
     num_halos = num_halos.view(pynbody.array.SimArray)
     num_halos.units = "a**-3 Mpc**-3 h**3"
-    num_halos.context = snapshot
+    num_halos.sim = snapshot
 
     if calculate_err:
         err = err.view(pynbody.array.SimArray)
         err.units = "a**-3 Mpc**-3 h**3"
-        err.context = snapshot
+        err.sim = snapshot
         return bin_centers, num_halos, err
     else:
         return bin_centers, num_halos
