@@ -1,22 +1,17 @@
 """
 
-family
-======
+Defines the :class:`Family` class which represents families of particles
+(e.g. dm, gas, star particles).
 
-This module defines the Family class which represents
-families of particles (e.g. dm, gas, star particles).
-New Family objects are automatically registered so that
-snapshots can use them in the normal syntax (snap.dm,
-snap.star, etc).
+Most users will not need to use this module directly. New families of particles
+can, however, be added interactively by creating new instances of the
+:class:`Family` class. See the example in the class documentation.
 
-In practice the easiest way to make use of the flexibility
-this module provides is through adding more families of
-particles in your config.ini.
-
+Another way to add families, more permanently, is to use the configuration file.
+See the :ref:`configuration` section for more information.
 """
 
 import functools
-import sys
 
 from . import config_parser
 
@@ -25,7 +20,20 @@ _registry = []
 
 def family_names(with_aliases=False):
     """Returns a list of the names of all particle families.
-    If with_aliases is True, include aliases in the list."""
+
+    Parameters
+    ----------
+
+    with_aliases : bool
+        If True, the list will include all aliases for each family. Default is False.
+
+    Returns
+    -------
+
+    list of str
+        The list of family names.
+
+    """
 
     global _registry
     l = []
@@ -38,10 +46,19 @@ def family_names(with_aliases=False):
 
 
 def get_family(name, create=False):
-    """Returns a family corresponding to the specified string.  If the
-    family does not exist and create is False, raises ValueError. If
-    the family does not exist and create is True, an appropriate
-    object is instantiated, registered and returned."""
+    """Returns a :class:`Family` object corresponding to the specified string.
+
+    Parameters
+    ----------
+
+    name : str or :class:`Family`
+        The name of the family to retrieve. If a :class:`Family` object is passed, it is returned unchanged.
+
+    create : bool
+        If True, a new family is created if the specified family does not exist. Default is False.
+        If False, a ``ValueError`` is raised if the specified family does not exist.
+
+    """
 
     if isinstance(name, Family):
         return name
@@ -56,13 +73,37 @@ def get_family(name, create=False):
     if create:
         return Family(name)
     else:
-        raise ValueError(name +
-                         " not a family")  # is ValueError the right thing here?
+        raise ValueError(name + " is not a family")
 
-
+@functools.total_ordering
 class Family:
+    """The class representing an abstract family of particles.
+
+    Once a :class:`Family` is defined, it is automatically registered and
+    can be used within :class:`~pynbody.snapshot.simsnap.SimSnap` objects. For example,
+
+    >>> import pynbody
+    >>> tachs = pynbody.family.Family("tachyon", aliases=["t", "tachyons"])
+    >>> f = pynbody.new(dm=100, tachyon=100)
+    >>> f.tachyon['pos'] # <-- get the tachyon positions
+    >>> f.t['pos'] # <-- use an alias to get the tachyon positions
+    >>> f[tachs]['pos'] # <-- another way to get the tachyon positions
+
+    """
 
     def __init__(self, name, aliases=[]):
+        """Create a new family with the specified name and aliases.
+
+        Parameters
+        ----------
+
+        name : str
+            The name of the family. Must be lower case and unique.
+
+        aliases : list of str
+            A list of aliases for the family. Must be lower case. These can be used
+            interchangeably with the family name.
+        """
         if name != name.lower():
             raise ValueError("Family names must be lower case")
         if name in family_names(with_aliases=True):
@@ -101,12 +142,7 @@ class Family:
     def __hash__(self):
         return hash(str(self))
 
-if sys.version_info[0] >= 3:
-    Family = functools.total_ordering(Family)
-
-
-# Instantiate the default families as specified
-# by the configuration file
+# Instantiate the default families as specified by the configuration file
 
 g = globals()
 for f in config_parser.options('families'):
