@@ -1,32 +1,95 @@
 """
+Support for halo and group catalogues.
 
-halo
-====
+Halo catalogues act like a dictionary, mapping from halo numbers to a Halo objects. The halo *number* is typically
+determined by the halo finder, and is often (but not always) the same as the halo *index* which is the zero-based
+offset within the catalogue.
 
-Implements halo catalogue functions. If you have a supported halo
-catalogue on disk or a halo finder installed and correctly configured,
-you can access a halo catalogue through f.halos() where f is a
-SimSnap.
+If you have a supported halo catalogue on disk or a halo finder installed and correctly configured, you can access a
+halo catalogue through ``f.halos()`` where ``f`` is a SimSnap.
 
-See the `halo tutorial
-<http://pynbody.github.io/pynbody/tutorials/halos.html>`_ for some
-examples.
+See the :ref:`halo catalogue tutorial <halo_tutorial>` for introductory information and guidance.
+
+
+.. _v2_0_halo_changes:
+
+.. versionchanged:: 2.0
+
+  Backwards-incompatible changes to the halo catalogue system
+
+  For version 2.0, the halo catalogue loading system was substantially rewritten. The new system is more robust and
+  more consistent across different halo finders. However, this means that some defaults have changed, most significantly
+  in the AHF halo numbering. Backward-compatibility can be achieved by passing ``halo_numbers='v1'`` to the
+  :class:`~pynbody.halo.ahf.AHFCatalogue` constructor. For more information, read the documentation for that class.
+
+  Furthermore, older versions of pynbody (i.e. v1.x) could be configured to create a halo catalogue if one was not
+  found, using AHF. This is no longer the case, as creating a halo catalogue requires choosing a halo finder and its
+  parameters carefully for the task in hand and it was not possible to provide a one-size-fits-all solution.
+
+  Finally, options to write ``.stat`` files and ``.grp`` files have been removed. However it is still possible to
+  generate a ``.grp`` file by  calling :meth:`~HaloCatalogue.get_group_array` and writing out the resulting
+  array of integers using a tool like ``numpy.savetxt``.
+
+  By paring back the less-used functionality of the halo catalogue system, the remaining functionality is more
+  consistent, robust, and extensible to new halo finders.
+
+
+.. _supported_halo_finders:
+
+Supported halo-finder formats
+-----------------------------
+
+The currently-supported formats are:
+
+- Adaptahop (:class:`~pynbody.halo.adaptahop.AdaptaHOPCatalogue`);
+- AHF (:class:`~pynbody.halo.ahf.AHFCatalogue`);
+- HBT+ (:class:`~pynbody.halo.hbtplus.HBTPlusCatalogue`);
+- HOP (:class:`~pynbody.halo.hop.HOPCatalogue`);
+- Rockstar (:class:`~pynbody.halo.rockstar.RockstarCatalogue`);
+- Subfind (old format :class:`~pynbody.halo.subfind.SubfindCatalogue`, or various HDF5 variants
+  as :class:`~pynbody.halo.subfindhdf.SubfindHDFCatalogue`);
+- VELOCIraptor (:class:`~pynbody.halo.velociraptor.VelociraptorCatalogue`).
+
+In addition, generic halo finders which output a list of halo numbers for each particle are supported via
+:class:`~pynbody.halo.number_array.HaloNumberCatalogue`.
+
+
+
+.. note::
+
+    The principal development of ``pynbody`` took place in the UK, and the spelling of "catalogue" is British English.
+    However, since much code is written in American English, v2.0.0 introduced aliases such that all
+    classes can be accessed with the American spelling ``HaloCatalog``, ``AdaptaHOPCatalog`` etc.
+
 
 """
+from __future__ import annotations
 
 import copy
 import logging
 import warnings
 import weakref
+from typing import TYPE_CHECKING, Iterable
 
 import numpy as np
+from numpy.typing import NDArray
 
-from .. import snapshot, util
+from .. import array, snapshot, units, util
+from ..util import iter_subclasses
+from .details.iord_mapping import make_iord_to_offset_mapper
+from .details.number_mapping import (
+    HaloNumberMapper,
+    MonotonicHaloNumberMapper,
+    create_halo_number_mapper,
+)
+from .details.particle_indices import HaloParticleIndices
+
+if TYPE_CHECKING:
+    from .subhalo_catalogue import SubhaloCatalogue
 
 logger = logging.getLogger("pynbody.halo")
 
-
-class DummyHalo(snapshot.ContainerWithPhysicalUnitsOption):
+class DummyHalo(snapshot.util.ContainerWithPhysicalUnitsOption):
 
     def __init__(self):
         self.properties = {}
@@ -35,33 +98,38 @@ class DummyHalo(snapshot.ContainerWithPhysicalUnitsOption):
         pass
 
 
-class Halo(snapshot.IndexedSubSnap):
+class Halo(snapshot.subsnap.IndexedSubSnap):
 
     """
-    Generic class representing a halo.
+    Represents a single halo from a halo catalogue.
+
+    Note that pynbody refers to groups, halos and subhalos interchangably, with the term "halo" being used to cover
+    all of these.
     """
 
-    def __init__(self, halo_id, halo_catalogue, *args, **kwa):
+    def __init__(self, halo_number, properties, halo_catalogue, *args, **kwa):
         super().__init__(*args, **kwa)
         self._halo_catalogue = halo_catalogue
-        self._halo_id = halo_id
-        self._descriptor = "halo_" + str(halo_id)
+        self._halo_number = halo_number
+        self._descriptor = "halo_" + str(halo_number)
         self.properties = copy.copy(self.properties)
-        self.properties['halo_id'] = halo_id
-        if halo_id in halo_catalogue._halos:
-            for key in list(halo_catalogue._halos[halo_id].properties.keys()):
-                self.properties[key] = halo_catalogue._halos[halo_id].properties[key]
+        self.properties['halo_number'] = halo_number
+        self.properties.update(properties)
 
         # Inherit autoconversion from parent
         self._autoconvert_properties()
 
-    def is_subhalo(self, otherhalo):
-        """
-        Convenience function that calls the corresponding function in
-        a halo catalogue.
-        """
+    @property
+    @util.deprecated("The sub property has been renamed to subhalos")
+    def sub(self):
+        """Deprecated alias for :property:`subhalos`."""
+        return self.subhalos
 
-        return self._halo_catalogue.is_subhalo(self._halo_id, otherhalo._halo_id)
+    @property
+    def subhalos(self) -> SubhaloCatalogue:
+        """A HaloCatalogue object containing only the subhalos of this halo."""
+        return self._halo_catalogue._get_subhalo_catalogue(self._halo_number)
+
 
     def physical_units(self, distance='kpc', velocity='km s^-1',
                        mass='Msol', persistent=True, convert_parent=True):
@@ -107,20 +175,55 @@ class Halo(snapshot.IndexedSubSnap):
             self._autoconvert_properties()
 
 
-# ----------------------------#
-# General HaloCatalogue class #
-# -----------------------------#
+class HaloCatalogue(snapshot.util.ContainerWithPhysicalUnitsOption,
+                    iter_subclasses.IterableSubclasses):
 
-class HaloCatalogue(snapshot.ContainerWithPhysicalUnitsOption):
+    """Generic halo catalogue object.
+
+    To the user, this presents a simple interface where calling ``h[i]`` returns halo ``i``. Properties of halos
+    can be retrieved without loading the halo via :meth:`get_properties_one_halo` or :meth:`get_properties_all_halos`.
+
+    More information for users can be found in the :ref:`halo catalogue tutorial <halo_tutorial>`; see also the
+    :ref:`supported halo finders <supported_halo_finders>`.
+
+    Implementing a new format
+    ^^^^^^^^^^^^^^^^^^^^^^^^^
+
+    To support a new format, subclass :class:`HaloCatalogue` and implement the following methods:
+
+    * :meth:`__init__`
+    * :meth:`_can_load`
+    * :meth:`_get_all_particle_indices`
+    * :meth:`_get_particle_indices_one_halo` [only if it's possible to do this more efficiently than
+      :meth:`_get_all_particle_indices` for users accessing only a few halos]
+    * :meth:`get_properties_all_halos` [only if you have halo finder-provided properties to expose]
+    * :meth:`get_properties_one_halo` [only if you have halo finder-provided properties to expose, and it's efficient
+      to expose them one halo at a time; the default implementation will call get_properties_all_halos and extract]
+    * :meth:`get_group_array` [only if it's possible to do this more efficiently than the default implementation]
+
+    Nomenclature/conventions are worth being aware of if you are implementing a new format:
+
+    * The halo number is the user-exposed identifier for a halo. It is typically assigned by the halo finder, although
+      subclasses are free to assign their own (e.g. some have a `halo_number` option that can be passed to the
+      constructor to override the halo finder's numbering). The halo numbers are used to access individual halos via
+      the [] operator.
+    * The halo *index* is the zero-based offset within the catalogue, which may be different from the halo number.
+      Internally, *pynbody* converts between these using a :class:`details.number_mapping.HaloNumberMapper` object,
+      which is set up in the :meth:`__init__` method.
+    * Particle indices should be returned from methods like :meth:`_get_particle_indices_one_halo` as zero-relative
+      offsets within the snapshot, not particle IDs or 'iord's. Many halo finders output particle IDs which must
+      therefore be mapped. To aid this, call :meth:`_init_iord_to_fpos` in your :meth:`__init__` method, which creates
+      a mapper as :attr:`_iord_to_fpos`. See :mod:`details.iord_mapping` for more information.
 
     """
-    Generic halo catalogue object.
-    """
 
-    def __init__(self, sim):
-        self._base = weakref.ref(sim)
-        self._halos = {}
-        self.lazy_off = util.ExecutionControl()
+    def __init__(self, sim, number_mapper):
+        self._base: weakref[snapshot.SimSnap] = weakref.ref(sim)
+        self.number_mapper: HaloNumberMapper = number_mapper
+        self._index_lists: HaloParticleIndices | None = None
+        self._properties: dict | None = None
+        self._cached_halos: dict[int, Halo] = {}
+        self._persistent_units = None
 
     def calc_item(self, i):
         if i in self._halos:  # and self._halos[i]() is not None :
@@ -132,64 +235,84 @@ class HaloCatalogue(snapshot.ContainerWithPhysicalUnitsOption):
             else:
                 return self._halos[i]
         else:
-            h = self._get_halo(i)
-            self._halos[i] = h  # weakref.ref(h)
-            return h
+            return {k: units.get_item_with_unit(self._properties[k],halo_index)
+                    for k in self._properties}
 
-    def __len__(self):
-        return len(self._halos)
+    def _get_particle_indices_one_halo(self, halo_number) -> NDArray[int]:
+        """Get the index list for a single halo, given a halo_number.
 
-    def __iter__(self):
-        return self._halo_generator()
+        A generic implementation is provided that fetches index lists for all halos and then extracts the one"""
+        self.load_all()
+        return self._index_lists.get_particle_index_list_for_halo(
+            self.number_mapper.number_to_index(halo_number)
+        )
 
-    def __getitem__(self, item):
-        if isinstance(item, slice):
-            for x in self._halo_generator(item.start, item.stop):
-                pass
-            indices = item.indices(len(self._halos))
-            res = [self.calc_item(i) for i in range(*indices)]
-            return res
+    def _get_particle_indices_one_halo_using_list_if_available(self, halo_number, halo_index) -> NDArray[int]:
+        if self._index_lists:
+            return self._index_lists.get_particle_index_list_for_halo(halo_index)
         else:
-            return self.calc_item(item)
+            if len(self._cached_halos) == 5:
+                warnings.warn("Accessing multiple halos may be more efficient if you call load_all() on the "
+                              "halo catalogue", RuntimeWarning)
+            return self._get_particle_indices_one_halo(halo_number)
+            # NB subclasses may implement loading one halo direct from disk in the above
+            # if not, the default implementation will populate _cached_index_lists
+
+    def _get_halo_cached(self, halo_number) -> Halo:
+        if halo_number not in self._cached_halos:
+            self._cached_halos[halo_number] = self._get_halo(halo_number)
+        return self._cached_halos[halo_number]
+
+    def _get_halo(self, halo_number) -> Halo:
+        halo_index = self.number_mapper.number_to_index(halo_number)
+        return Halo(halo_number,
+                    self._get_properties_one_halo_using_cache_if_available(halo_number, halo_index),
+                    self, self.base,
+                    self._get_particle_indices_one_halo_using_list_if_available(halo_number, halo_index))
+
+    def get_dummy_halo(self, halo_number) -> DummyHalo:
+        """Return a DummyHalo object containing only the halo properties, no particle information"""
+        h = DummyHalo()
+        h.properties.update(self.get_properties_one_halo(halo_number))
+        return h
+
+    def __len__(self) -> int:
+        return self._get_num_halos()
+
+    def __iter__(self) -> Iterable[Halo]:
+        self.load_all()
+        for i in self.number_mapper:
+            yield self[i]
+
+    def __repr__(self):
+        return f"<{type(self).__name__}, length {len(self)}>"
+
+    def keys(self):
+        """Return an iterable of all halo numbers in the catalogue."""
+        return self.number_mapper.all_numbers
+
+    def __getitem__(self, item) -> Halo | SubhaloCatalogue:
+        from .subhalo_catalogue import SubhaloCatalogue
+        if isinstance(item, slice):
+            return SubhaloCatalogue(self, np.arange(*item.indices(len(self))))
+        elif hasattr(item, "__len__"):
+            return SubhaloCatalogue(self, item)
+        else:
+            return self._get_halo_cached(item)
 
     @property
-    def base(self):
+    def base(self) -> snapshot.SimSnap:
+        """The snapshot object that this halo catalogue is based on."""
         return self._base()
 
-    def _halo_generator(self, i_start=None, i_stop=None):
-        if len(self) == 0:
-            return
-        if i_start is None or i_stop is None:
-            try:
-                self[0]
-                one_indexed = False
-            except KeyError:
-                one_indexed = True
-
-        if i_start is None:
-            i = 1 if one_indexed else 0
-        else:
-            i = i_start
-
-        if i_stop is None:
-            i_stop = len(self) + 1 if one_indexed else len(self)
-
-        while True:
-            try:
-                yield self[i]
-                i += 1
-                if i != i_stop and len(self[i]) == 0:
-                    continue
-            except RuntimeError:
-                break
-            if i == i_stop:
-                return
-
     def _init_iord_to_fpos(self):
+        """Create a member array, _iord_to_fpos, that maps particle IDs to file positions.
+
+        This is a convenience function for subclasses to use."""
         if not hasattr(self, "_iord_to_fpos"):
-            if 'iord' in self.base.loadable_keys():
-                self._iord_to_fpos = np.empty(self.base['iord'].max() + 1, dtype=np.int64)
-                self._iord_to_fpos[self.base['iord']] = np.arange(len(self.base))
+            if 'iord' in self.base.loadable_keys() or 'iord' in self.base.keys():
+                self._iord_to_fpos = make_iord_to_offset_mapper(self.base['iord'])
+
             else:
                 warnings.warn("No iord array available; assuming halo catalogue is using sequential particle IDs",
                               RuntimeWarning)
@@ -200,68 +323,66 @@ class HaloCatalogue(snapshot.ContainerWithPhysicalUnitsOption):
 
                 self._iord_to_fpos = OneToOneIndex()
 
-    def is_subhalo(self, childid, parentid):
-        """Checks whether the specified 'childid' halo is a subhalo
-        of 'parentid' halo.
-        """
-        if (childid in self._halos[parentid].properties['children']):
-            return True
+    def _get_subhalo_catalogue(self, parent_halo_number: int) -> SubhaloCatalogue:
+        from .subhalo_catalogue import SubhaloCatalogue
+        props = self.get_properties_one_halo(parent_halo_number)
+        if 'children' in props:
+            return SubhaloCatalogue(self, props['children'])
         else:
-            return False
+            raise ValueError(f"This halo catalogue does not support subhalos")
 
-    def contains(self, haloid):
-        if (haloid in self._halos):
-            return True
+    @util.deprecated("This method is deprecated and will be removed in a future release. Use python `in` syntax instead.")
+    def contains(self, halo_number: int) -> bool:
+        """Deprecated alias; instead of ``h.contains(number)`` use ``number in h``."""
+        return halo_number in self
+
+    def __contains__(self, halo_number) -> bool:
+        """Returns True if the halo catalogue contains the specified halo number."""
+        return halo_number in self.number_mapper
+
+    def get_group_array(self, family=None, use_index=False, fill_value=-1):
+        """Return an array with an integer for each particle in the simulation, indicating the halo of that particle.
+
+        If there are multiple levels (i.e. subhalos), the number returned corresponds to the lowest level, i.e.
+        the smallest subhalo.
+
+        Parameters
+        ----------
+
+        family : str, optional
+            If specified, return only the group array for the specified family.
+
+        use_index: bool, optional
+            If True, return the halo index rather than the halo number. (See the class documentation for the
+            distinction between halo numbers and indices.)
+
+        fill_value : int, optional
+            The value to fill for particles not in any halo.
+
+        """
+        self.load_all()
+        number_per_particle = self._index_lists.get_halo_number_per_particle(len(self.base),
+                                                                             None if use_index else self.number_mapper,
+                                                                             fill_value = fill_value)
+        if family is not None:
+            return number_per_particle[self.base._get_family_slice(family)]
         else:
-            return False
+            return number_per_particle
 
-    def __contains__(self, haloid):
-        return self.contains(haloid)
+    def load_copy(self, halo_number):
+        """Load a fresh SimSnap with only the particles in specified halo
 
-    def get_group_array(self):
-        """Return an array with an integer for each particle in the simulation
-        indicating which halo that particle is associated with. If there are multiple
-        levels (i.e. subhalos), the number returned corresponds to the lowest level, i.e.
-        the smallest subhalo."""
-        raise NotImplementedError
+        This relies on the underlying SimSnap being capable of partial loading."""
+        from .. import load
+        halo_index = self.number_mapper.number_to_index(halo_number)
+        return load(self.base.filename,
+                    take=self._get_particle_indices_one_halo_using_list_if_available(halo_number, halo_index))
 
-    @staticmethod
-    def _can_load(self):
-        return False
-
-    @staticmethod
-    def _can_run(self):
-        return False
-
-    def physical_units(self, distance='kpc', velocity='km s^-1',
-                       mass='Msol', persistent=True, convert_parent=False):
-        """
-        Converts all array's units to be consistent with the
-        distance, velocity, mass basis units specified.
-
-        Base units can be specified using keywords.
-
-        **Optional Keywords**:
-
-           *distance*: string (default = 'kpc')
-
-           *velocity*: string (default = 'km s^-1')
-
-           *mass*: string (default = 'Msol')
-
-           *persistent*: boolean (default = True); apply units change to future lazy-loaded arrays if True
-
-           *convert_parent*: boolean (default = None); ignored for HaloCatalogue objects
-
-        """
-        self.base.physical_units(
-            distance=distance,
-            velocity=velocity,
-            mass=mass,
-            persistent=persistent)
+    def physical_units(self, distance='kpc', velocity='km s^-1', mass='Msol', persistent=True, convert_parent=False):
+        self.base.physical_units(distance=distance, velocity=velocity, mass=mass, persistent=persistent)
 
         # Convert all instantiated subhalos
-        for halo in self._halos.values():
+        for halo in self._cached_halos.values():
             halo.physical_units(
                 distance,
                 velocity,
@@ -270,138 +391,55 @@ class HaloCatalogue(snapshot.ContainerWithPhysicalUnitsOption):
                 convert_parent=False
             )
 
+        all_units = [units.Unit(x) for x in (distance, velocity, mass, 'a', 'h', 'K')]
 
-class GrpCatalogue(HaloCatalogue):
-    """
-    A generic catalogue using a .grp file to specify which particles
-    belong to which group.
-    """
+        if persistent:
+            self._persistent_units = all_units
 
-    def __init__(self, sim, array='grp', ignore=None, **kwargs):
-        """Construct a GrpCatalogue, extracting halos based on a simulation-wide integer array with their IDs
+        self._cached_properties_to_physical_units(all_units)
 
-        *sim* - the SimSnap for which the halos will be constructed
-        *array* - the name of the array which should be present, loadable or derivable across the simulation
-        *ignore* - a special value indicating "no halo", or None if no such special value is defined
-        """
-        sim[array]  # trigger lazy-loading and/or kick up a fuss if unavailable
-        self._halos = {}
-        self._array = array
-        self._sorted = None
-        self._ignore = ignore
-        HaloCatalogue.__init__(self, sim)
-
-    def __len__(self):
-        if self._ignore is None:
-            N = self.base[self._array].max()
-        else:
-            N = self.base[self._array]
-            N = N[N != self._ignore]
-            N = N.max()
-        if N < 0:
-            N = 0
-        return N
-
-    def precalculate(self):
-        """Speed up future operations by precalculating the indices
-        for all halos in one operation. This is slow compared to
-        getting a single halo, however."""
-        self._sorted = np.argsort(
-            self.base[self._array], kind='mergesort')  # mergesort for stability
-        self._unique_i = np.unique(self.base[self._array])
-        self._boundaries = np.searchsorted(
-            self.base[self._array][self._sorted], self._unique_i)
-
-    def get_group_array(self, family=None):
-        if family is not None:
-            return self.base[family][self._array]
-        else:
-            return self.base[self._array]
-
-    def _get_halo_indices(self, i):
-        if self.base is None:
-            raise RuntimeError("Parent SimSnap has been deleted")
-
-        no_exist = ValueError("Halo %s does not exist" % (str(i)))
-
-        if self._sorted is None:
-            # one-off selection
-            index = np.where(self.base[self._array] == i)
-            return index
-        else:
-            # pre-calculated
-            if not np.isin(i, self._unique_i):
-                raise no_exist
-
-            match = np.where(self._unique_i == i)[0]
-
-            start = self._boundaries[match][0]
-
-            if start == self._boundaries[-1]:
-                # This is the final halo
-                end = None
-            else:
-                end = self._boundaries[match+1][0]
-
-            return self._sorted[start:end]
-
-    def _get_halo(self, i):
-        x = Halo(i, self, self.base, self._get_halo_indices(i))
-        if len(x) == 0:
-            raise ValueError("Halo %s does not exist" % (str(i)))
-        x._descriptor = "halo_" + str(i)
-        return x
-
-    def load_copy(self, i):
-        """Load the a fresh SimSnap with only the particle in halo i"""
-        from .. import load
-        return load(self.base.filename, take=self._get_halo_indices(i))
-
-    @property
-    def base(self):
-        return self._base()
-
-    @staticmethod
-    def _can_load(sim, arr_name='grp'):
-        if (arr_name in sim.loadable_keys()) or (arr_name in list(sim.keys())):
-            return True
-        else:
-            return False
+    def _cached_properties_to_physical_units(self, all_units):
+        if self._properties is not None:
+            for k in self._properties:
+                if isinstance(self._properties[k], array.SimArray) and units.has_unit(self._properties[k]):
+                    self.base._autoconvert_array_unit(self._properties[k], all_units)
 
 
-class AmigaGrpCatalogue(GrpCatalogue):
-    def __init__(self, sim, arr_name='amiga.grp', **kwargs):
-        GrpCatalogue.__init__(self, sim, arr_name)
+    @classmethod
+    def _can_load(cls, sim):
+        return False
 
-    @staticmethod
-    def _can_load(sim, arr_name='amiga.grp'):
-        return GrpCatalogue._can_load(sim, arr_name)
-
-
-from pynbody.halo.adaptahop import AdaptaHOPCatalogue, NewAdaptaHOPCatalogue
-from pynbody.halo.ahf import AHFCatalogue
-from pynbody.halo.hop import HOPCatalogue
-from pynbody.halo.legacy import RockstarIntermediateCatalogue
-from pynbody.halo.rockstar import RockstarCatalogue
-from pynbody.halo.subfind import SubfindCatalogue
-from pynbody.halo.subfindhdf import (
-    ArepoSubfindHDFCatalogue,
-    Gadget4SubfindHDFCatalogue,
-    SubFindHDFHaloCatalogue,
-    TNGSubfindHDFCatalogue,
+from . import (
+    adaptahop,
+    ahf,
+    hbtplus,
+    hop,
+    number_array,
+    rockstar,
+    subfind,
+    subfindhdf,
+    velociraptor,
 )
 from pynbody.zooms.auriga import AurigaSubfindHDFCatalogue
 
 
-def _get_halo_classes():
-    # AmigaGrpCatalogue MUST be scanned first, because if it exists we probably
-    # want to use it, but an AHFCatalogue will probably be on-disk too.
-    _halo_classes = [
-        GrpCatalogue, AmigaGrpCatalogue, AHFCatalogue,
-        RockstarCatalogue, SubfindCatalogue, SubFindHDFHaloCatalogue,
-        NewAdaptaHOPCatalogue, AdaptaHOPCatalogue,
-        RockstarIntermediateCatalogue, HOPCatalogue, Gadget4SubfindHDFCatalogue,
-        ArepoSubfindHDFCatalogue, TNGSubfindHDFCatalogue, AurigaSubfindHDFCatalogue
-    ]
+def _fix_american_spelling(p):
+    """Map American to British spelling (used by SimSnap.halos to allow flexible spelling)"""
+    if isinstance(p, str) and p.endswith('Catalog'):
+        return p.replace('Catalog', 'Catalogue')
+    else:
+        return p
+def _alias_american_spelling():
+    """Create American spelling aliases for all HaloCatalogue subclasses."""
+    for c in HaloCatalogue.iter_subclasses():
+        american_name = c.__name__.replace("Catalogue", "Catalog")
+        # put american_name into the same module as c (not this module)
 
-    return _halo_classes
+        if c.__module__.startswith('pynbody.halo.'):
+            module = eval(c.__module__.replace('pynbody.halo.', ''))
+
+        setattr(module, american_name, c)
+
+    globals()['HaloCatalog'] = HaloCatalogue
+
+_alias_american_spelling()
