@@ -1,3 +1,7 @@
+"""Support for the Rockstar halo finder"""
+
+from __future__ import annotations
+
 import glob
 import os.path
 import sys
@@ -5,46 +9,44 @@ import sys
 import numpy as np
 
 from .. import util
-from . import DummyHalo, Halo, HaloCatalogue
+from . import HaloCatalogue
+from .details import number_mapping
+
+
+class RockstarFormatRevisionError(RuntimeError):
+    pass
 
 
 class RockstarCatalogue(HaloCatalogue):
-    def __init__(self, sim, dummy=False, pathname=None, format_revision=None,
-                 filenames=None, sort=False, **kwargs):
+    def __init__(self, sim, filename=None, format_revision=None):
         """Initialize a RockstarCatalogue.
 
-        **kwargs** :
+        Parameters
+        ----------
 
+        sim : :class:`pynbody.snapshot.simsnap.SimSnap`
+            The simulation snapshot to which this catalogue applies.
 
-        *dummy*: if True, the particle file is not loaded, and all
-                 halos returned are just dummies (with the correct
-                 properties dictionary loaded). Use load_copy to get
-                 the actual data in this case.
+        filename : str, optional
+            The path to the Rockstar outputs. This can either be the first binary file in the series (e.g.
+            ``path/to/halos_15.0.bin``) or the stem for the series (e.g. ``path/to/halos_15``). If not specified,
+            the code will attempt to find the Rockstar outputs in the same directory as the simulation snapshot.
 
-        *sort*: if True, resort the halos into descending order of
-                particle number. Otherwise, leave in RockStar output order.
-
-        *filenames*: a list of filenames of each of the RockStar outputs.
-                     You probably want to use pathname instead, which specifies
-                     the path to the output folder.
-
-        *pathname*: the path of the output folder with the individual RockStar outputs
-
-        *format_revision*: Override the header's format revision information. Specify
-                    1, 2, or 'caterpillar' for Rockstar prior to 2014, post 2014, and
-                    customized for the caterpillar project respectively
+        format_revision : int, str, optional
+            Override the header's format revision information. Specify 1, 2, 'caterpillar', 'galaxies' for Rockstar
+            prior to 2014, post 2014, customized for the caterpillar project and for rockstar galaxies respectively.
+            If not specified, the code will attempt to detect the format revision automatically.
 
         """
-        HaloCatalogue.__init__(self, sim)
 
-        self._dummy = dummy
-
-        if filenames is not None:
-            self._files = filenames
+        if filename is not None:
+            filestem = str(filename)
+            if filestem.endswith(".0.bin"):
+                filestem = filestem[:-6]
+            self._files = glob.glob(filestem + "*.bin") + glob.glob(filestem + "*.boundbin")
         else:
-            if pathname is None:
-                pathname = os.path.dirname(sim.filename)
-            self._files = glob.glob(os.path.join(pathname,'halos*.bin'))
+            pathname = os.path.dirname(sim.filename)
+            self._files = glob.glob(os.path.join(pathname, 'halos*.bin'))
             if len(self._files)==0 :
                 self._files = glob.glob(os.path.join(pathname, 'halos*.boundbin'))
             self._files.sort()
@@ -52,97 +54,99 @@ class RockstarCatalogue(HaloCatalogue):
         if len(self._files)==0:
             raise OSError("Could not find any Rockstar output. Try specifying pathname='/path/to/rockstar/outputfolder'")
 
-        self._cpus = [RockstarCatalogueOneCpu(sim,dummy,file_i, format_revision=format_revision) for file_i in self._files]
-        self._prune_files_from_wrong_scalefactor()
-        self._cpus[0]._init_iord_to_fpos()
-        for cpu in self._cpus:
-            cpu._iord_to_fpos = self._cpus[0]._iord_to_fpos
+        self._cpus = [_RockstarCatalogueOneCpu(file_i, format_revision=format_revision) for file_i in self._files]
+        self._prune_files_from_wrong_scalefactor(sim)
 
-        self._init_index_ar()
-        if sort:
-            self._sort_index_ar()
+        halomins = [cpu.halo_min_inclusive for cpu in self._cpus]
+        # order self._cpus and self._files by halomins
+        # this isn't technically necessary but it means that the internal halo indexing will be
+        # in the same order as the halo numbering in the files. In fact, it probably means that
+        # the internal halo indexing will be identical to the halo numbering in the files, but we
+        # won't take that for granted anywhere.
+        self._cpus = [x for _, x in sorted(zip(halomins, self._cpus))]
+        self._files = [x for _, x in sorted(zip(halomins, self._files))]
 
-    def _prune_files_from_wrong_scalefactor(self):
+        halo_numbers = np.empty(sum(len(x) for x in self._cpus), dtype=int)
+        self._cpu_per_halo = np.empty(len(halo_numbers), dtype=int)
+        i = 0
+        for j, cat in enumerate(self._cpus):
+            halo_numbers[i:i+len(cat)] = cat.get_halo_numbers()
+            self._cpu_per_halo[i:i+len(cat)] = j
+            i+=len(cat)
+        assert i == len(halo_numbers)
+
+        super().__init__(sim, number_mapping.create_halo_number_mapper(halo_numbers))
+
+
+
+    def _prune_files_from_wrong_scalefactor(self, sim):
         new_cpus = []
         new_files = []
         for file,cpu in zip(self._files,self._cpus):
-            if abs(self.base.properties['a']-cpu._head['scale'][0])<1e-6:
+            if abs(sim.properties['a']-cpu._head['scale'][0])<1e-6:
                 new_cpus.append(cpu)
                 new_files.append(file)
         self._cpus = new_cpus
         self._files = new_files
 
-    def _pass_on(self, function, *args, **kwargs):
-        if self._index_ar is not None:
-            cpui, hi = self._cpus[self._index_ar[args[0],0]], self._index_ar[args[0],1]
-            return function(cpui,hi,*args[1:],**kwargs)
-        for i in self._cpus:
-            try:
-                return function(i,*args,**kwargs)
-            except KeyError:
-                pass
+    def _get_particle_indices_one_halo(self, halo_number):
+        halo_index = self.number_mapper.number_to_index(halo_number)
+        cpu = self._cpu_per_halo[halo_index]
+        iords = self._cpus[cpu].read_iords_for_halo(halo_number)
+        self._init_iord_to_fpos()
+        return self._iord_to_fpos.map_ignoring_order(iords)
 
-    def __getitem__(self, k):
-        return self._pass_on(RockstarCatalogueOneCpu.__getitem__, k)
+    def _get_all_particle_indices(self):
+        iords = np.empty(0, dtype=int)
+        boundaries = np.empty((0, 2), dtype=int)
+        for cpu in self._cpus:
+            iords_this_cpu, boundaries_this_cpu = cpu.read_iords_for_all_halos()
+            boundaries = np.append(boundaries, boundaries_this_cpu + len(iords), axis=0)
+            iords = np.append(iords, iords_this_cpu)
 
-    def load_copy(self, k):
-        return self._pass_on(RockstarCatalogueOneCpu.load_copy, k)
+        self._init_iord_to_fpos()
+        fpos = iords # nb this doesn't copy! but we won't use the iords again after this, so it's ok
+        for a, b in boundaries:
+            # iord_to_fpos may not retain ordering, so have to do this per halo
+            fpos[a:b] = self._iord_to_fpos.map_ignoring_order(iords[a:b])
+        return fpos, boundaries
 
-    def get_group_array(self):
-        ar = np.zeros(len(self.base), dtype=int)-1
-        for cpu_i in self._cpus:
-            cpu_i._update_grp_array(ar)
-        return ar
+    def get_properties_one_halo(self, halo_number):
+        halo_index = self.number_mapper.number_to_index(halo_number)
+        cpu = self._cpu_per_halo[halo_index]
+        return self._cpus[cpu].read_properties_for_halo(halo_number)
 
-    def make_grp(self, name='grp'):
-        """
-        Creates a 'grp' array which labels each particle according to
-        its parent halo.
-        """
-        self.base[name]= self.get_group_array()
+    def get_properties_all_halos(self, with_units=True) -> dict:
+        props = {}
+        for cpu in self._cpus:
+            props_this_cpu = cpu.read_properties_all_halos()
+            # concatenate all the properties
+            for k, v in props_this_cpu.items():
+                if k in props:
+                    props[k] = np.concatenate((props[k], v))
+                else:
+                    props[k] = v
 
-    def __len__(self):
-        return sum(len(x) for x in self._cpus)
+        return props
 
-    def _init_index_ar(self):
-        index_ar = np.empty((len(self),2),dtype=np.int32)
-
-        for cpu_id, cpu in enumerate(self._cpus):
-            i0 = cpu._halo_min
-            i1 = cpu._halo_max
-            index_ar[i0:i1,0]=cpu_id
-            index_ar[i0:i1,1]=np.arange(i0,i1,dtype=np.int32)
-
-        self._index_ar = index_ar
-
-
-    def _sort_index_ar(self):
-        num_ar = np.empty(len(self))
-
-        for cpu_id, cpu in enumerate(self._cpus):
-            i0 = cpu._halo_min
-            i1 = cpu._halo_max
-            num_ar[i0:i1]=self._cpus[cpu_id]._halo_lens
-
-        num_ar = np.argsort(num_ar)[::-1]
-        self._index_ar = self._index_ar[num_ar]
-
-
-    @staticmethod
-    def _can_run(sim):
-        return False
-
-    @staticmethod
-    def _can_load(sim, **kwargs):
-        for file in glob.glob(os.path.join(os.path.dirname(sim.filename), 'halos*.bin')):
-            if os.path.exists(file):
+    @classmethod
+    def _can_load(cls, sim, filename=None, format_revision=None):
+        if filename is None:
+            return len(
+                glob.glob(os.path.join(os.path.dirname(sim.filename), 'halos*.bin'))
+            ) > 0
+        else:
+            if str(filename).endswith(".0.bin") and os.path.exists(filename):
                 return True
-        return False
+            if os.path.exists(str(filename)+".0.bin"):
+                return True
+            return False
 
 
-class RockstarCatalogueOneCpu(HaloCatalogue):
+
+class _RockstarCatalogueOneCpu:
     """
-    Class to handle sub-catalogues produced by Rockstar. Users should normally not use this class,
+    Low-level reader for single CPU output from Rockstar. Users should normally not use this class,
     rather using RockstarCatalogue which collates the multiple sub-files that Rockstar produces.
     """
 
@@ -211,271 +215,181 @@ class RockstarCatalogueOneCpu(HaloCatalogue):
                           ('min_pos_err','f'),('min_vel_err','f'),
                           ('min_bulkvel_err','f'),
                           ('num_bound', 'i8'), ('num_iter', 'i8')]
-                                , align=True) # Hacked rockstar from caterpillar project
+                                , align=True), # Hacked rockstar from caterpillar project
+        'galaxies':  np.dtype(
+        [
+                ("id", np.int64),
+                ("pos", np.float32, 3),
+                ("vel", np.float32, 3),
+                ("corevel", np.float32, 3),
+                ("bulkvel", np.float32, 3),
+                ("m", np.float32),
+                ("r", np.float32),
+                ("child_r", np.float32),
+                ("vmax_r", np.float32),
+                ("mgrav", np.float32),
+                ("vmax", np.float32),
+                ("rvmax", np.float32),
+                ("rs", np.float32),
+                ("klypin_rs", np.float32),
+                ("vrms", np.float32),
+                ("J", np.float32, 3),
+                ("energy", np.float32),
+                ("spin", np.float32),
+                ("alt_m", np.float32, 4),
+                ("Xoff", np.float32),
+                ("Voff", np.float32),
+                ("b_to_a", np.float32),
+                ("c_to_a", np.float32),
+                ("A", np.float32, 3),
+                ("b_to_a2", np.float32),
+                ("c_to_a2", np.float32),
+                ("A2", np.float32, 3),
+                ("bullock_spin", np.float32),
+                ("kin_to_pot", np.float32),
+                ("m_pe_b", np.float32),
+                ("m_pe_d", np.float32),
+                ("num_p", np.int64),
+                ("num_child_particles", np.int64),
+                ("p_start", np.int64),
+                ("desc", np.int64),
+                ("flags", np.int64),
+                ("n_core", np.int64),
+                ("min_pos_err", np.float32),
+                ("min_vel_err", np.float32),
+                ("min_bulkvel_err", np.float32),
+                ("type", np.int32),
+                ("sm", np.float32),
+                ("gas", np.float32),
+                ("bh", np.float32),
+                ("peak_density", np.float32),
+                ("av_density", np.float32),
+            ],
+            align=True
+        ), # Galaxy format from Rockstar
     }
 
 
-    def __init__(self, sim, dummy=False, filename=None, format_revision=None, **kwargs):
-        """Initialize a RockstarCatalogue.
+    def __init__(self, filename=None, format_revision=None):
 
-        **kwargs** :
-
-
-        *dummy*: if True, the particle file is not loaded, and all
-                 halos returned are just dummies (with the correct
-                 properties dictionary loaded). Use load_copy to get
-                 the actual data in this case.
+        self._rsFilename = filename
 
 
-        """
+        if not os.path.exists(self._rsFilename):
+            raise OSError(
+                "Halo catalogue not found -- check the file name of catalogue data"
+                " or try specifying a catalogue using the filename keyword"
+            )
 
-        HaloCatalogue.__init__(self,sim)
-
-        self._dummy = dummy
-
-        if filename is not None: self._rsFilename = filename
-        else:
-            self._rsFilename = util.cutgz(glob.glob('halos*.bin')[0])
-
-        try :
-            f = util.open_(self._rsFilename, 'rb')
-        except OSError:
-            raise OSError("Halo catalogue not found -- check the file name of catalogue data or try specifying a catalogue using the filename keyword")
-
-        with f:
-            self._head = np.fromstring(f.read(self.head_type.itemsize),
+        with util.open_(self._rsFilename, 'rb') as f:
+            self._head = np.frombuffer(f.read(self.head_type.itemsize),
                                        dtype=self.head_type)
-            if format_revision is None:
-                format_revision = self._head['format_revision'][0]
 
-            self.halo_type = self.halo_types[format_revision]
-            unused = f.read(256 - self._head.itemsize)
+            # Seek to absolute position
+            f.seek(256)
 
             self._nhalos = self._head['num_halos'][0]
 
-            self._load_rs_halos(f,sim)
+            self._load_rs_halos_with_format_detection(f, format_revision)
 
+    def _load_rs_halos_with_format_detection(self, f, format_revision):
+        if format_revision is None:
+            # The 'galaxies' format can be either 1 or 2, so we need to try it
+            # in both cases.
+            format_revision_to_try = [self._head['format_revision'][0], "galaxies"]
+        else:
+            format_revision_to_try = [format_revision]
 
+        current_pos = f.tell()
+        for format_revision in format_revision_to_try:
+            f.seek(current_pos)
+            try:
+                self.halo_type = self.halo_types[format_revision]
+                self._load_rs_halos(f)
+                return
+            except RockstarFormatRevisionError:
+                pass
+
+        raise RockstarFormatRevisionError(
+            "Could not detect the format revision of the Rockstar catalogue."
+        )
 
 
     def __len__(self):
         return len(self._halo_lens)
 
-    def _load_all(self):
-        for i in range(self._halo_min, self._halo_max):
-            self._load_rs_halo_if_required(i)
-            self._load_rs_particles_for_halo_if_required(i)
-
-    def calc_item(self, i):
-        if self.base is None:
-            raise RuntimeError("Parent SimSnap has been deleted")
-
-        self._load_rs_halo_if_required(i)
-        self._load_rs_particles_for_halo_if_required(i)
-
-        return self._halos[i]
-
-    def load_copy(self, i):
-        """Load a fresh SimSnap with only the particles in halo i"""
-        if i<self._halo_min or i>=self._halo_max:
-            raise KeyError("No such halo")
-
-        from . import load
-        return load(self.base.filename, take=self._get_particles_for_halo(i))
-
-    def _update_grp_array(self, ar):
-        """Insert into an existing grp array halos from this CPU"""
-        self._load_all()
-        for halo in list(self._halos.values()):
-            ar[halo.get_index_list(self.base)] = halo._halo_id
 
 
-    def _load_rs_halo_if_required(self, i):
-        if i not in self._halos:
-            self._halos[i] = self._get_dummy_for_halo(i)
-
-    def _get_dummy_for_halo(self, n):
-        if n<self._halo_min or n>=self._halo_max:
+    def read_properties_for_halo(self, n):
+        if n<self.halo_min_inclusive or n>=self.halo_max_exclusive:
             raise KeyError("No such halo")
 
         with util.open_(self._rsFilename, 'rb') as f:
-            f.seek(self._haloprops_offset+(n-self._halo_min)*self.halo_type.itemsize)
+            f.seek(self._haloprops_offset + (n - self.halo_min_inclusive) * self.halo_type.itemsize)
             halo_data = np.fromfile(f, dtype=self.halo_type, count=1)
 
-        hn = DummyHalo()
         # TODO: properties are in Msun / h, Mpc / h
-        hn.properties = dict(list(zip(halo_data.dtype.names,halo_data[0])))
-        return hn
+        return dict(list(zip(halo_data.dtype.names,halo_data[0])))
 
+    def read_properties_all_halos(self):
+        with util.open_(self._rsFilename, 'rb') as f:
+            f.seek(self._haloprops_offset)
+            data = np.fromfile(f, dtype=self.halo_type, count=self.halo_max_exclusive - self.halo_min_inclusive)
 
-    def _load_rs_halos(self, f, sim):
+        data_dict = {name: data[name] for name in data.dtype.names}
+
+        return data_dict
+
+    def _load_rs_halos(self, f):
         self._haloprops_offset = f.tell()
         self._halo_offsets = np.empty(self._head['num_halos'][0],dtype=np.int64)
         self._halo_lens = np.empty(self._head['num_halos'][0],dtype=np.int64)
+
         offset = self._haloprops_offset+self.halo_type.itemsize*self._head['num_halos'][0]
 
-
-        self._halo_min = int(np.fromfile(f, dtype=self.halo_type, count=1)['id'])
-        self._halo_max = int(self._halo_min+self._head['num_halos'][0])
+        self.halo_min_inclusive = int(np.fromfile(f, dtype=self.halo_type, count=1)['id'][0])
+        self.halo_max_exclusive = int(self.halo_min_inclusive + self._head['num_halos'][0])
 
         f.seek(self._haloprops_offset)
 
-        this_id = self._halo_min
-
-        for h in range(self._head['num_halos'][0]):
-            halo_data =np.fromfile(f, dtype=self.halo_type, count=1)
-            assert halo_data['id']==this_id
-            self._halo_offsets[this_id-self._halo_min] = int(offset)
+        for this_id in range(self.halo_min_inclusive, self.halo_max_exclusive):
+            halo_data = np.fromfile(f, dtype=self.halo_type, count=1)
+            if halo_data['id'] != this_id:
+                raise RockstarFormatRevisionError(
+                    "Error while reading halo catalogue. Expected "
+                    "halo ID %d, but got %d" % (this_id, halo_data['id'])
+                )
+            self._halo_offsets[this_id - self.halo_min_inclusive] = offset
             if 'num_bound' in self.halo_type.names:
-                num_ptcls = int(halo_data['num_bound'])
+                num_ptcls = int(halo_data['num_bound'][0])
             else:
-                num_ptcls = int(halo_data['num_p'])
-            self._halo_lens[this_id-self._halo_min] = num_ptcls
+                num_ptcls = int(halo_data['num_p'][0])
+            self._halo_lens[this_id - self.halo_min_inclusive] = num_ptcls
             offset+=num_ptcls*np.dtype('int64').itemsize
-            this_id+=1
 
 
-    def _get_particles_for_halo(self, num):
-        self._init_iord_to_fpos()
+    def get_halo_numbers(self):
+        return np.arange(self.halo_min_inclusive, self.halo_max_exclusive)
+
+    def read_iords_for_halo(self, num):
+        if num<self.halo_min_inclusive or num>=self.halo_max_exclusive:
+            raise KeyError("No such halo")
+
         with util.open_(self._rsFilename, 'rb') as f:
-            f.seek(self._halo_offsets[num-self._halo_min])
-            halo_ptcls=np.fromfile(f,dtype=np.int64,count=self._halo_lens[num-self._halo_min])
-            halo_ptcls = self._iord_to_fpos[halo_ptcls]
-            halo_ptcls.sort()
+            f.seek(self._halo_offsets[num - self.halo_min_inclusive])
+            return np.fromfile(f, dtype=np.int64, count=self._halo_lens[num - self.halo_min_inclusive])
 
-        return halo_ptcls
+    def read_iords_for_all_halos(self):
+        """Returns an array with all halo iords, and the boundaries of each halo in the array."""
+        # check the halo ids are contiguous as expected
+        assert (np.diff(self._halo_offsets) == 8 * self._halo_lens[:-1]).all()
+        with util.open_(self._rsFilename, 'rb') as f:
+            f.seek(self._halo_offsets[0])
+            iords = np.fromfile(f, dtype=np.int64, count=self._halo_lens.sum())
 
-    def _load_rs_particles_for_halo(self, num):
-        halo_ptcls = self._get_particles_for_halo(num)
+        boundaries = np.empty((len(self._halo_lens), 2), dtype=np.int64)
+        boundaries[:,1] = np.cumsum(self._halo_lens)
+        boundaries[0,0] = 0
+        boundaries[1:,0] = boundaries[:-1,1]
 
-        properties_from_proxy = self._halos[num].properties
-
-        self._halos[num]=Halo(num, self, self.base, halo_ptcls)
-        self._halos[num]._descriptor = "halo_"+str(num)
-
-        self._halos[num].properties.update(properties_from_proxy)
-
-    def _load_rs_particles_for_halo_if_required(self, num):
-        if isinstance(self._halos[num], DummyHalo) and not self._dummy:
-            self._load_rs_particles_for_halo(num)
-
-    def writestat(self, outfile=None, hubble=None):
-        """
-        write a condensed skid.stat style ascii file from ahf_halos
-        file.  header + 1 halo per line. should reproduce `Alyson's
-        idl script' except does not do last 2 columns (Is it a
-        satellite?) and (Is central halo is `false'ly split?).  output
-        units are set to Mpc Msun, km/s.
-
-        user can specify own hubble constant hubble=(H0/(100
-        km/s/Mpc)), ignoring the snaphot arg for hubble constant
-        (which sometimes has a large roundoff error).
-        """
-        s = self._base()
-        mindarkmass = min(s.dark['mass'])
-
-        if hubble is None:
-            hubble = s.properties['h']
-
-        if outfile is None: outfile = self._base().filename+'.stat'
-        print("write stat file to ", outfile)
-        fpout = open(outfile, "w")
-        header = "#Grp  N_tot     N_gas      N_star    N_dark    Mvir(M_sol)       Rvir(kpc)       GasMass(M_sol) StarMass(M_sol)  DarkMass(M_sol)  V_max  R@V_max  VelDisp    Xc   Yc   Zc   VXc   VYc   VZc   Contam   Satellite?   False?   ID_A"
-        print(header, file=fpout)
-        for ii in np.arange(self._nhalos)+1:
-            print('%d '%ii, end=' ')
-            sys.stdout.flush()
-            h = self[ii].properties  # halo index starts with 1 not 0
-##  'Contaminated'? means multiple dark matter particle masses in halo)"
-            icontam = np.where(self[ii].dark['mass'] > mindarkmass)
-            if (len(icontam[0]) > 0):
-                contam = "contam"
-            else:
-                contam = "clean"
-## may want to add implement satellite test and false central breakup test.
-            ss = "     "  # can adjust column spacing
-            outstring = str(ii)+ss
-            outstring += str(len(self[ii]))+ss+str(len(self[ii].g))+ss
-            outstring += str(len(self[ii].s)) + ss+str(len(self[ii].dark))+ss
-            outstring += str(h['m']/hubble)+ss+str(h['r']/hubble)+ss
-            outstring += str(self[ii].g['mass'].in_units('Msol').sum())+ss
-            outstring += str(self[ii].s['mass'].in_units('Msol').sum())+ss
-            outstring += str(self[ii].d['mass'].in_units('Msol').sum())+ss
-            outstring += str(h['vmax'])+ss+str(h['vmax_r']/hubble)+ss
-            outstring += str(h['vrms'])+ss
-        ## pos: convert kpc/h to mpc (no h).
-            outstring += str(h['pos'][0][0]/hubble)+ss
-            outstring += str(h['pos'][0][1]/hubble)+ss
-            outstring += str(h['pos'][0][2]/hubble)+ss
-            outstring += str(h['vel'][0][0])+ss+str(h['vel'][0][1])+ss
-            outstring += str(h['vel'][0][2])+ss
-            outstring += contam+ss
-            outstring += "unknown" + \
-                ss  # unknown means sat. test not implemented.
-            outstring += "unknown"+ss  # false central breakup.
-            print(outstring, file=fpout)
-        fpout.close()
-
-    def writetipsy(self, outfile=None, hubble=None):
-        """
-        write halos to tipsy file (write as stars) from ahf_halos
-        file.  returns a shapshot where each halo is a star particle.
-
-        user can specify own hubble constant hubble=(H0/(100
-        km/s/Mpc)), ignoring the snaphot arg for hubble constant
-        (which sometimes has a large roundoff error).
-        """
-        import math
-
-        from snapshot import _new as new
-
-        from . import tipsy
-        from .analysis import cosmology
-        s = self._base()
-        if outfile is None: outfile = s.filename+'.gtp'
-        print("write tipsy file to ", outfile)
-        sout = new(star=self._nhalos)  # create new tipsy snapshot written as halos.
-        sout.properties['a'] = s.properties['a']
-        sout.properties['z'] = s.properties['z']
-        sout.properties['boxsize'] = s.properties['boxsize']
-        if hubble is None: hubble = s.properties['h']
-        sout.properties['h'] = hubble
-    ### ! dangerous -- rho_crit function and unit conversions needs simplifying
-        rhocrithhco = cosmology.rho_crit(s, z=0, unit="Msol Mpc^-3 h^2")
-        lboxkpc = sout.properties['boxsize'].ratio("kpc a")
-        lboxkpch = lboxkpc*sout.properties['h']
-        lboxmpch = lboxkpc*sout.properties['h']/1000.
-        tipsyvunitkms = lboxmpch * 100. / (math.pi * 8./3.)**.5
-        tipsymunitmsun = rhocrithhco * lboxmpch**3 / sout.properties['h']
-
-        print("transforming ", self._nhalos, " halos into tipsy star particles")
-        for ii in range(self._nhalos):
-            h = self[ii+1].properties
-            sout.star[ii]['mass'] = h['m']/hubble / tipsymunitmsun
-            ## tipsy units: box centered at 0. (assume 0<=x<=1)
-            sout.star[ii]['x'] = h['pos'][0][0]/lboxmpch - 0.5
-            sout.star[ii]['y'] = h['pos'][0][1]/lboxmpch - 0.5
-            sout.star[ii]['z'] = h['pos'][0][2]/lboxmpch - 0.5
-            sout.star[ii]['vx'] = h['vel'][0][0]/tipsyvunitkms
-            sout.star[ii]['vy'] = h['vel'][0][1]/tipsyvunitkms
-            sout.star[ii]['vz'] = h['vel'][0][2]/tipsyvunitkms
-            sout.star[ii]['eps'] = h['r']/lboxkpch
-            sout.star[ii]['metals'] = 0.
-            sout.star[ii]['phi'] = 0.
-            sout.star[ii]['tform'] = 0.
-        print("writing tipsy outfile %s"%outfile)
-        sout.write(fmt=tipsy.TipsySnap, filename=outfile)
-        return sout
-
-    def writehalos(self, hubble=None, outfile=None):
-        s = self._base()
-        if outfile is None:
-            statoutfile = s.filename+".rockstar.stat"
-            tipsyoutfile = s.filename+".rockstar.gtp"
-        else:
-            statoutfile = outfile+'.stat'
-            gtpoutfile = outfile+'.gtp'
-        self.make_grp()
-        self.writestat(statoutfile, hubble=hubble)
-        shalos = self.writetipsy(gtpoutfile, hubble=hubble)
-        return shalos
+        return iords, boundaries
