@@ -1,147 +1,210 @@
+"""Support for the SubFind halo finder"""
+
+from __future__ import annotations
+
 import os.path
 import warnings
-import weakref
 
 import numpy as np
 
 from .. import units
 from ..array import SimArray
-from . import Halo, HaloCatalogue
+from . import HaloCatalogue
+from .details import number_mapping, particle_indices
+from .subhalo_catalogue import SubhaloCatalogue
 
 
 class SubfindCatalogue(HaloCatalogue):
+    """Handles catalogues produced by the SubFind halo finder (old versions that do not use HDF5 outputs)."""
 
-    """Class to handle catalogues produced by the SubFind halo finder.
-
-    By default, the FoF groups are imported, but the subhalos can also be imported by setting subs=True
-    when constructing.
-
-    """
-
-
-    def __init__(self, sim, subs=False, ordered=None, make_grp=None, verbose=False, **kwargs):
+    def __init__(self, sim, filename=None, subs=None, subhalos=False,  ordered=None, _inherit_data_from=None):
         """Initialise a SubFind catalogue
 
-        **kwargs** :
+        By default, the FoF groups are imported, and subhalos are available via the 'subhalos' attribute of each
+        halo object, e.g.
 
-        *subs*: if True, load the subhalos; otherwise, load FoF groups (default)
-        *ordered*: if True, the snapshot must have sequential iords. If False, the iords might be out-of-sequence, and
-                 therefore reordering will take place. If None (default), the iords are examined to check whether
-                 re-ordering is required or not. Note that re-ordering can be undesirable because it also destroys
-                 subfind's order-by-binding-energy
-        *make_grp*: if True, create a 'grp' array with the halo id to which each particle is assigned
-        *verbose*: if True, produces printed output to track progress while making the 'grp' array
+        >>> f = pynbody.load('path/to/snapshot')
+        >>> h = f.halos()
+        >>> h[1].subhalos[2] # returns the third subhalo of FoF group 1
+
+        However by setting ``subs=True``, the FoF groups are ignored and the catalogue is of all subhalos.
+
+        Parameters
+        ----------
+
+        sim : ~pynbody.snapshot.simsnap.SimSnap
+            The simulation snapshot to which this catalogue applies.
+
+        filename : str, optional
+            The path to the SubFind output(s). This is expected to be a folder named ``groups_XXX`` where XXX is the
+            snapshot number. The code extracts the snapshot number from the folder name and uses it to construct
+            the filename of the catalogue files, for example ``groups_XXX/subhalo_tab_XXX.0``. If no filename is
+            provided, the code will attempt to find a suitable catalogue starting from the simulation's directory.
+
+        subhalos : bool, optional
+            If False (default), catalogue represents the FoF groups and subhalos are available through the
+            :meth:`~pynbody.halo.Halo.subhalos` attribute of each group (see note above). If True, the catalogue
+            represents the subhalos directly and FoF groups are not available.
+
+        ordered : bool, optional
+            If True, the snapshot must have sequential iords. If False, the iords might be out-of-sequence, and
+            therefore reordering will take place. If None (default), the iords are examined to check whether
+            re-ordering is required or not. Note that re-ordering can be undesirable because it also destroys
+            subfind's order-by-binding-energy
+
+        _inherit_data_from : SubfindCatalogue, optional
+            For internal use only; allows subhalo catalogue to share data with its parent FOF catalogue
+
+
         """
 
-        if 'order' in kwargs:
-            warnings.warn("The keyword 'order' has been renamed to 'ordered'", DeprecationWarning)
-            ordered = kwargs.pop('order')
+        if subs is not None:
+            warnings.warn("The 'subs' argument to SubfindCatalogue is deprecated; use 'subhalos' instead",
+                          DeprecationWarning)
+            subhalos = subs
 
-        self._base = weakref.ref(sim)
-        self._subs=subs
+        self._subs=subhalos
 
         if ordered is not None:
             self._ordered = ordered
         else:
-            self._ordered = bool((self.base['iord']==np.arange(len(self.base))).all())
+            self._ordered = bool((sim['iord']==np.arange(len(sim))).all())
 
         self._halos = {}
-        HaloCatalogue.__init__(self,sim)
+
         self.dtype_int = sim['iord'].dtype
-        self.dtype_flt='float32' #SUBFIND data is apparently always single precision???
-        self.halodir = self._name_of_catalogue(sim)
-        self.header = self._readheader()
-        if subs is True:
+        self.dtype_flt = 'float32' #SUBFIND data is apparently always single precision???
+        self._subfind_dir = filename or self._name_of_catalogue(sim)
+
+        self.header = self._read_header()
+        self._tasks = self.header[4]
+
+        if _inherit_data_from:
+            self._inherit_data(_inherit_data_from)
+        else:
+            self._read_data(sim)
+
+
+
+        if subhalos:
             if self.header[6]==0:
                 raise ValueError("This file does not contain subhalos")
-            if make_grp is True:
-                raise ValueError("subs=True and make_grp=True are not compatible.")
-        self._tasks = self.header[4]
-        self.ids = self._read_ids()
-        self._keys={}
-        self._halodat, self._subhalodat=self._read_groups()
-        #self.data_len, self.data_off = self._read_groups()
-        if make_grp:
-            self.make_grp(v=verbose)
-
-    def make_grp(self, name='grp', v=False):
-        """Creates a 'grp' array which labels each particle according to its parent halo.
-
-        This can take quite some time! Option: v=True prints out 'progress' in terms of total number of groups.
-        """
-        self.base[name] = self.get_group_array(v=v)
-
-    def get_group_array(self, v=False):
-        ar = np.zeros(len(self.base), dtype=int)-1
-        for i in range(0, self.header[1]): #total number of groups
-            if v:
-                print("Halo #", i , "of", self.header[1])
-            halo=self[i]
-            ar[halo.get_index_list(self.base)] = halo._halo_id
-        return ar
-
-    def get_halo_properties(self, i, with_unit=True):
-        if with_unit:
-            extract = units.get_item_with_unit
+            length = len(self._subhalodat['sub_off'])
         else:
-            extract = lambda array, element: array[element]
+            length = len(self._halodat['group_off'])
+
+        super().__init__(sim, number_mapping.SimpleHaloNumberMapper(0, length))
+
+        if not subhalos:
+            self._subhalo_catalogue = SubfindCatalogue(sim, subhalos=True, filename=filename,
+                                                       ordered=self._ordered, _inherit_data_from=self)
+
+    def _inherit_data(self, from_):
+        inherit = ['_halodat', '_subhalodat', '_keys_halo', '_keys_subhalo']
+        for k in inherit:
+            setattr(self, k, getattr(from_, k))
+
+    def _get_all_particle_indices(self):
+        ids = self._read_ids()
+        boundaries = np.empty((len(self), 2), dtype=self.dtype_int)
+        if self._subs:
+            boundaries[:, 0] = self._subhalodat['sub_off']
+            boundaries[:, 1] = self._subhalodat['sub_off'] + self._subhalodat['sub_len']
+        else:
+            boundaries[:, 0] = self._halodat['group_off']
+            boundaries[:, 1] = self._halodat['group_off'] + self._halodat['group_len']
+
+        if not self._ordered:
+            self._init_iord_to_fpos()
+            for a, b in boundaries:
+                ids[a:b] = self._iord_to_fpos.map_ignoring_order(ids[a:b])
+                # must be done segmented in case iord_to_fpos doesn't preserve input order
+
+        return particle_indices.HaloParticleIndices(ids, boundaries)
+
+    def get_properties_one_halo(self, i):
+
+        extract = units.get_item_with_unit
+
         properties = {}
-        halo_number_within_group = (self._subhalodat['sub_groupNr'][:i] == self._subhalodat['sub_groupNr'][i]).sum()
+
         if self._subs is False:
-            for key in self._keys:
+            for key in self._keys_halo:
                 properties[key] = extract(self._halodat[key], i)
-            if self.header[6] > 0:
-                properties['children'], = np.where(self._subhalodat['sub_groupNr'] == i)
-                # this is the FIRST level of substructure, sub-subhalos (etc) can be accessed via the
-                # subs=True output (below)
+            properties['children'] = self._get_children_of_group(i)
         else:
-            for key in self._keys:
+            for key in self._keys_subhalo:
                 properties[key] = extract(self._subhalodat[key], i)
-            properties['children'], = np.where((self._subhalodat['sub_parent'] == halo_number_within_group) \
-                                               & (self._subhalodat['sub_groupNr'] == properties['sub_groupNr']))
-            # this goes down one level in the hierarchy, i.e. a subhalo will have all its sub-subhalos listed,
-            # but not its sub-sub-subhalos (those will be listed in each sub-subhalo)
+            properties['children'] = self._get_children_of_sub(i)
         return properties
 
-    def _get_halo(self,i):
-        if self._ordered is False:
-            if self._subs is True:
-                #this needs to be tested again on a snapshot that is not ordered!
-                x = Halo(i, self, self.base, np.where(np.in1d(self.base['iord'], self.ids[self._subhalodat['sub_off'][i]:self._subhalodat['sub_off'][i]+self._subhalodat['sub_len'][i]]  )))
-            else:
-                x = Halo(i, self, self.base, np.where(np.in1d(self.base['iord'], self.ids[self._halodat['group_off'][i]:self._halodat['group_off'][i]+self._halodat['group_len'][i]]  )))
-
+    def _get_children_of_group(self, group_nr):
+        if self.header[6] > 0:
+            return np.where(self._subhalodat['sub_groupNr'] == group_nr)[0]
         else:
-            if self._subs is False: #to use groups as halos:
-                x = Halo(i, self, self.base, self.ids[self._halodat['group_off'][i]:self._halodat['group_off'][i]+self._halodat['group_len'][i]] )
+            return []
+
+    def _get_children_of_sub(self, subhalo_nr):
+        # this goes down one level in the hierarchy, i.e. a subhalo will have all its sub-subhalos listed,
+        # but not its sub-sub-subhalos (those will be listed in each sub-subhalo)
+        halo_number_within_group = (self._subhalodat['sub_groupNr'][:subhalo_nr]
+                                    == self._subhalodat['sub_groupNr'][subhalo_nr]).sum()
+        return np.where((self._subhalodat['sub_parent'] == halo_number_within_group) &
+                        (self._subhalodat['sub_groupNr'] == self._subhalodat['sub_groupNr'][subhalo_nr]))[0]
+
+
+    def get_properties_all_halos(self, with_units=True) -> dict:
+        properties = {}
+        if self._subs:
+            data = self._subhalodat
+            properties['children'] = [self._get_children_of_sub(subhalo_nr)
+                                      for subhalo_nr in self.number_mapper.all_numbers]
+        else:
+            data = self._halodat
+            if self.header[6] > 0:
+                properties['children'] = [self._get_children_of_group(group_nr)
+                                          for group_nr in self.number_mapper.all_numbers]
             else:
-                x=Halo(i, self, self.base, self.ids[self._subhalodat['sub_off'][i]:self._subhalodat['sub_off'][i]+self._subhalodat['sub_len'][i]] )
+                properties['children'] = []
 
-        x._descriptor = "halo_"+str(i)
-        x.properties.update(self.get_halo_properties(i))
-        return x
+        keys = self._keys_halo if not self._subs else self._keys_subhalo
 
-    def _readheader(self):
-        header = np.array([], dtype='int32')
-        iout = self.halodir.split("_")[-1]
+        for key in keys:
+            if with_units:
+                properties[key] = data[key]
+            else:
+                properties[key] = data[key].view(np.ndarray)
+
+
+        return properties
+
+    def _read_header(self):
+        iout = self._subfind_dir.split("_")[-1]
         filename = os.path.join(
-            self.halodir,
+            self._subfind_dir,
             f"subhalo_tab_{iout}.0"
         )
-        fd = open(filename, "rb")
-        # read header: this is strange but it works: there is an extra value in
-        # header which we delete in the next step
-        header1 = np.fromfile(fd, dtype='int32', sep="", count=8)
-        header = np.delete(header1, 4)
-        fd.close()
-        return header  # [4]
+        with open(filename, "rb") as fd:
+            # read header: this is strange but it works: there is an extra value in
+            # header which we delete in the next step
+            header = np.fromfile(fd, dtype='int32', sep="", count=8)
 
+        header = np.delete(header, 4)
+
+        return header
+
+    def _get_subhalo_catalogue(self, parent_halo_number):
+        if self._subs:
+            return SubhaloCatalogue(self, self._get_children_of_sub(parent_halo_number))
+        else:
+            return SubhaloCatalogue(self._subhalo_catalogue,
+                                    self._get_children_of_group(parent_halo_number))
     def _read_ids(self):
         data_ids = np.array([], dtype=self.dtype_int)
-        iout = self.halodir.split("_")[-1]
+        iout = self._subfind_dir.split("_")[-1]
         for n in range(0, self._tasks):
             filename = os.path.join(
-                self.halodir,
+                self._subfind_dir,
                 f"subhalo_ids_{iout}.{n}"
             )
             fd = open(filename, "rb")
@@ -155,7 +218,7 @@ class SubfindCatalogue(HaloCatalogue):
             data_ids = np.append(data_ids, ids)
         return data_ids
 
-    def _read_groups(self):
+    def _read_data(self, sim):
         halodat={}
         keys_flt=['mass', 'pos', 'mmean_200', 'rmean_200', 'mcrit_200', 'rcrit_200', 'mtop_200', 'rtop_200', 'contmass']
         keys_int=['group_len', 'group_off',  'first_sub', 'Nsubs', 'cont_count', 'mostboundID']
@@ -174,14 +237,13 @@ class SubfindCatalogue(HaloCatalogue):
         for key in subkeys_flt:
             subhalodat[key]=np.array([], dtype=self.dtype_flt)
 
-        self._keys=keys_flt+keys_int
-        if self._subs is True:
-            self._keys=subkeys_flt+subkeys_int
+        self._keys_subhalo = subkeys_flt+subkeys_int
+        self._keys_halo = keys_flt + keys_int
 
         for n in range(0,self._tasks):
-            iout = self.halodir.split("_")[-1]
+            iout = self._subfind_dir.split("_")[-1]
             filename = os.path.join(
-                self.halodir,
+                self._subfind_dir,
                 f"subhalo_tab_{iout}.{n}"
             )
             fd=open(filename, "rb")
@@ -241,17 +303,13 @@ class SubfindCatalogue(HaloCatalogue):
 
         for name, dimension in zip(ar_names, ar_dimensions):
             if name in halodat:
-                halodat[name] = SimArray(halodat[name], self.base.infer_original_units(dimension))
+                halodat[name] = SimArray(halodat[name], sim.infer_original_units(dimension))
+                halodat[name].sim = sim
             if name in subhalodat:
-                subhalodat[name] = SimArray(subhalodat[name], self.base.infer_original_units(dimension))
+                subhalodat[name] = SimArray(subhalodat[name], sim.infer_original_units(dimension))
+                subhalodat[name].sim = sim
 
-        return halodat, subhalodat
-
-    def __len__(self):
-        if self._subs:
-            return len(self._subhalodat['sub_pos'])
-        else:
-            return len(self._halodat['pos'])
+        self._halodat, self._subhalodat = halodat, subhalodat
 
     @staticmethod
     def _name_of_catalogue(sim):
@@ -269,18 +327,17 @@ class SubfindCatalogue(HaloCatalogue):
             parent_dir = os.path.dirname(sim.filename)
             return os.path.join(parent_dir,"groups_" + snapnum)
 
-    @property
-    def base(self):
-        return self._base()
-
     @staticmethod
-    def _can_load(sim, **kwargs):
+    def _can_load(sim, filename=None, **kwargs):
+        if filename is not None:
+            if str(filename)[:-3].endswith("groups_"):
+                return True
         file = SubfindCatalogue._name_of_catalogue(sim)
         if os.path.exists(file):
             if os.path.exists(file):
                 if file.endswith(".hdf5"):
                     return False
-                elif os.listdir(file)[0].endswith(".hdf5"):
+                elif os.path.isdir(file) and os.listdir(file)[0].endswith(".hdf5"):
                     return False
                 else:
                     return True
