@@ -12,7 +12,7 @@ import logging
 import matplotlib.pyplot as plt
 import numpy as np
 
-from ..analysis import angmom, halo, profile
+from ..analysis import angmom, profile
 from ..units import Unit
 from .generic import hist2d
 
