@@ -1,27 +1,33 @@
 """
-
-sph
-===
-
-routines for plotting smoothed quantities
+Routines for plotting smoothed quantities
 
 """
+
+import warnings
 
 import matplotlib
 import numpy as np
 import pylab as p
 
 from .. import config, sph, units as _units
+from ..sph import renderers
 
+
+def _width_in_sim_units(sim, width):
+	if isinstance(width, str) or issubclass(width.__class__, _units.UnitBase):
+		if isinstance(width, str):
+			width = _units.Unit(width)
+		width = width.in_units(sim['pos'].units, **sim.conversion_context())
+	return width
 
 def sideon_image(sim, *args, **kwargs):
-    """
+    """Create an image of the given simulation, side-on to the disc of the passed halo.
 
-    Rotate the simulation so that the disc of the passed halo is
-    side-on, then make an SPH image by passing the parameters into
-    the function image
+    This routine is a small wrapper around :func:`~pynbody.plot.sph.image` that rotates the simulation so that the disc
+    of the passed halo is side-on, makes the SPH image, then rotates the simulation back to its original orientation.
 
-    For a description of keyword arguments see :func:`~pynbody.plot.sph.image`.
+    More flexible control over the orientation of the simulation can be achieved by using
+    :func:`~pynbody.analysis.angmom.sideon` in combination with :func:`~pynbody.plot.sph.image`.
 
     """
 
@@ -32,13 +38,13 @@ def sideon_image(sim, *args, **kwargs):
 
 
 def faceon_image(sim, *args, **kwargs):
-    """
+    """Create an image of the given simulation, face-on to the disc of the passed halo.
 
-    Rotate the simulation so that the disc of the passed halo is
-    face-on, then make an SPH image by passing the parameters into
-    the function image
+    This routine is a small wrapper around :func:`~pynbody.plot.sph.image` that rotates the simulation so that the disc
+    of the passed halo is face-on, makes the SPH image, then rotates the simulation back to its original orientation.
 
-    For a description of keyword arguments see :func:`~pynbody.plot.sph.image`.
+    More flexible control over the orientation of the simulation can be achieved by using
+    :func:`~pynbody.analysis.angmom.faceon` in combination with :func:`~pynbody.plot.sph.image`.
 
     """
 
@@ -48,62 +54,207 @@ def faceon_image(sim, *args, **kwargs):
         return image(sim, *args, **kwargs)
 
 
-def velocity_image(sim, width="10 kpc", vector_color='black', edgecolor='black', quiverkey_bg_color=None,
-                   vector_resolution=40, scale=None, mode='quiver', key_x=0.3, key_y=0.9,
-                   key_color='white', key_length="100 km s**-1", quiverkey=True, density=1.0,
-                   vector_qty='vel', **kwargs):
+def contour(sim, qty, width="10 kpc", resolution=None, units=None, axes=None, label=True, log=True, weight=None,
+			contour_kwargs=None, smooth_floor=0.0, _transform=None):
+    """Create an image of the given quantity then turn it into contours.
+
+    Parameters
+    ----------
+
+    sim : pynbody.snapshot.simsnap.SimSnap
+        The simulation snapshot to plot. The image is generated in the plane z=0, or
+        projected along the z axis.
+
+    qty : str | pynbody.array.SimArray
+        The name of the array to interpolate. Default is 'rho', which gives a density
+        image. Alternatively, an array can be passed in.
+
+    width : str or float, optional
+        The overall width and height of the plot. If a float, it is assumed
+        to be in units of sim['pos']. It can also be passed in as a string indicating
+        the units, e.g. '10 kpc'. (Default is '10 kpc')
+
+    resolution : int, optional
+        The number of pixels wide and tall. (Default is determined by the
+        :ref:`configuration file <configuration>`.)
+
+    units : str or pynbody.units.Unit, optional
+        The units of the output. Default is None, in which case the units of the input
+        quantity are used. If the units correspond to integrating the quantity along
+        a spatial dimension, the output is a projected image. For example, if the units
+        are 'Msol kpc^-2', and the quantity is 'rho', the output is a projected image of
+        the surface density.
+
+    axes : matplotlib.axes.Axes, optional
+        Axes instance on which the image will be shown; if None, the current pyplot figure is
+        used. (Default is False)
+
+    label : bool, optional
+        Whether to label the contours. (Default is True)
+
+    log : bool, optional
+        If True, the image is log-scaled before being contoured. (Default is True)
+
+    weight : str, optional
+        If set, the requested quantity is volume-averaged down the line of sight, weighted either
+        by volume (if weight is True) or by a specified quantity. (Default is None)
+
+    contour_kwargs : dict, optional
+        Additional keyword arguments to pass to the matplotlib contour function. (Default is None)
+
+    smooth_floor : float, optional
+        The minimum size of the smoothing kernel, either as a float or a unit string.
+		Setting this to a non-zero value makes smoother, clearer contours but loses fine detail.
+		Default is 0.0.
+
+    _transform : function, optional
+        A function to apply to the image before contouring. (Default is None)
+
     """
 
+    if resolution is None:
+        resolution = config['image-default-resolution']
+
+    if axes is None:
+        axes = p.gca()
+
+    if contour_kwargs is None:
+        contour_kwargs = {}
+
+    width = _width_in_sim_units(sim, width)
+    pixel_size = width / resolution
+
+    # width of image must be a pixel wider than the width of the final contour field, since the contours
+    # are based on the centres of the pixels not their edges
+
+    pipeline = renderers.make_render_pipeline(sim.s, quantity=qty, width=width + pixel_size,
+                                              weight = weight, out_units = units, resolution = resolution,
+                                              smooth_floor = smooth_floor)
+
+    im = pipeline.render()
+
+    if log:
+        im = np.log10(im)
+
+    if _transform:
+        im = _transform(im)
+
+    # width of image was expanded above, so this is now the positions of the centres
+    X = np.linspace(-width/2, width/2, resolution)
+
+    CS = axes.contour(X, X, im, **contour_kwargs)
+    if label:
+        axes.clabel(CS, fontsize=12, inline=True)
+
+
+def velocity_image(sim, qty='rho', vector_qty='vel', width="10 kpc", mode='quiver',
+                   vector_color='black', vector_edgecolor='black', vector_resolution=40,
+                   vector_scale=None, key=True, key_x=0.5, key_y=0.88,
+                   key_color='k', key_edge_color='k', key_bg_color='w',
+                   key_length="100 km s**-1",
+                   stream_density=1.0, stream_linewidth = 1.0,
+                   weight=None, restrict_depth=False,
+                   **kwargs):
+    """
     Make an SPH image of the given simulation with velocity vectors overlaid on top.
 
-    For a description of additional keyword arguments see :func:`~pynbody.plot.sph.image`,
-    or see the `tutorial <http://pynbody.github.io/pynbody/tutorials/pictures.html#velocity-vectors>`_.
+    Any keyword argument that can be passed to :func:`~pynbody.plot.sph.image` can also be passed
+    to this function. See that function for a full list of options.
 
-    **Keyword arguments:**
+    Parameters
+    ----------
 
-    *vector_color* (black): The color for the velocity vectors
+    sim : pynbody.snapshot.simsnap.SimSnap
+        The simulation snapshot to plot. The image is generated in the plane z=0, or
+        projected along the z axis.
 
-    *edgecolor* (black): edge color used for the lines - using a color
-     other than black for the *vector_color* and a black *edgecolor*
-     can result in poor readability in pdfs
+    qty : str | pynbody.array.SimArray
+        The name of the array to interpolate. Default is 'rho', which gives a density
+        image. Alternatively, an array can be passed in.
 
-    *vector_resolution* (40): How many vectors in each dimension (default is 40x40)
+    vector_qty : str
+        The name of the array to use for the vectors. Default is 'vel'.
 
-    *quiverkey_bg_color* (none): The color for the legend (scale) background
+    width : str or float, optional
+        The overall width and height of the plot. If a float, it is assumed
+        to be in units of sim['pos']. It can also be passed in as a string indicating
+        the units, e.g. '10 kpc'. (Default is '10 kpc')
 
-    *scale* (None): The length of a vector that would result in a displayed length of the
-    figure width/height.
+    mode : str, optional
+        The type of plot to make. Options are 'quiver' or 'stream'. (Default is 'quiver')
 
-    *mode* ('quiver'): make a 'quiver' or 'stream' plot
+    vector_color : str, optional
+        The color of the velocity vectors. (Default is 'black')
 
-    *key_x* (0.3): Display x (width) position for the vector key (quiver mode only)
+    vector_edgecolor : str, optional
+        The color for the edges of the velocity vectors. (Default is 'black')
 
-    *key_y* (0.9): Display y (height) position for the vector key (quiver mode only)
+    vector_resolution : int, optional
+        The number of velocity vectors to generate in each dimension. (Default is 40)
 
-    *key_color* (white): Color for the vector key (quiver mode only)
+    vector_scale : str or float, optional
+        The length of a vector that would result in a displayed length of the
+        figure width/height. This can be provided as a unit string or a float, in which
+        case it is interpreted as having the same units as the vector array being plotted.
+        Default is None, in which case the vectors are scaled by matplotlib's quiver function.
+        This option is only used in 'quiver' mode.
 
-    *key_length* (100 km/s): Velocity to use for the vector key (quiver mode only)
+    key : bool, optional
+        Whether or not to inset a key showing the scale of the vectors.
+        Only used if in 'quiver' mode. (Default is True)
 
-    *density* (1.0): Density of stream lines (stream mode only)
+    key_x : float, optional
+        The x position of the key in 'quiver' mode, if key is True. (Default is 0.5)
 
-    *quiverkey* (True): Whether or not to inset the key
+    key_y : float, optional
+        The y position of the key in 'quiver' mode, if key is True. (Default is 0.88)
 
-    *vector_qty* ('vel'): The name of the vector field to plot
+    key_color : str, optional
+        The color of the key arrow/text in 'quiver' mode, if key is True. (Default is 'white')
+
+    key_edge_color : str, optional
+        The color of the border around the key in 'quiver' mode, if key is True. (Default is 'black')
+
+    key_bg_color : str, optional
+        The color of the background of the key in 'quiver' mode, if key is True. (Default is 'white')
+
+    key_length : str, optional
+        The velocity to use for the key in 'quiver' mode, if key is True. (Default is '100 km s**-1')
+
+    stream_density : float, optional
+        The density of stream lines in 'stream' mode. (Default is 1.0)
+
+    stream_linewidth : float, optional
+        The width of stream lines in 'stream' mode. (Default is 1.0)
+
+    **kwargs :
+        Any additional keyword arguments to pass to :func:`~pynbody.plot.sph.image`.
+
     """
 
-    subplot = kwargs.get('subplot', False)
-    av_z = kwargs.get('av_z',None)
-    if subplot:
-        p = subplot
-    else:
-        import matplotlib.pylab as p
+
+
+    if 'av_z' in kwargs:
+        weight = kwargs.pop('av_z')
+        warnings.warn("av_z is deprecated; use weight instead", DeprecationWarning)
+
+
 
     vx_name, vy_name, _ = sim._array_name_ND_to_1D(vector_qty)
 
-    vx = image(sim, qty=vx_name, width=width, log=False,
-               resolution=vector_resolution, noplot=True,av_z=av_z)
-    vy = image(sim, qty=vy_name, width=width, log=False,
-               resolution=vector_resolution, noplot=True,av_z=av_z)
+    weight_for_vector = weight
+
+    if 'units' in kwargs and _units_imply_projection(sim, qty, _units.Unit(kwargs['units'])) and weight is None:
+        weight_for_vector = 'rho'
+
+    vel_pipeline = renderers.make_render_pipeline(sim, quantity=vx_name, width=width,
+                                                  resolution=vector_resolution, weight=weight_for_vector,
+                                                  restrict_depth=restrict_depth)
+
+    vx = vel_pipeline.render()
+    vel_pipeline.set_quantity(vy_name)
+    vy = vel_pipeline.render()
+
     key_unit = _units.Unit(key_length)
 
     if isinstance(width, str) or issubclass(width.__class__, _units.UnitBase):
@@ -117,29 +268,38 @@ def velocity_image(sim, width="10 kpc", vector_color='black', edgecolor='black',
     X, Y = np.meshgrid(np.linspace(-width / 2 + pixel_size/2, width / 2 - pixel_size/2, vector_resolution),
                        np.linspace(-width / 2 + pixel_size/2, width / 2 - pixel_size/2, vector_resolution))
 
-    im = image(sim, width=width, **kwargs)
+    im = image(sim, qty=qty, width=width, weight=weight, restrict_depth=restrict_depth,
+               **kwargs)
+
+    axes = kwargs.get('axes', None)
+    if axes is None:
+        axes = p.gca()
 
     if mode == 'quiver':
-        if scale is None:
-            Q = p.quiver(X, Y, vx, vy, color=vector_color, edgecolor=edgecolor)
+        if vector_scale is None:
+            Q = axes.quiver(X, Y, vx, vy, color=vector_color, edgecolor=vector_edgecolor)
         else:
-            Q = p.quiver(X, Y, vx, vy, scale=_units.Unit(scale).in_units(
-                sim['vel'].units), color=vector_color, edgecolor=edgecolor)
-        if quiverkey:
-        	qk = p.quiverkey(Q, key_x, key_y, key_unit.in_units(sim['vel'].units, **sim.conversion_context()),
-                    r"$\mathbf{" + key_unit.latex() + "}$", labelcolor=key_color, color=key_color, fontproperties={'size': 24})
-        if  quiverkey_bg_color is not None:
-            qk.text.set_backgroundcolor(quiverkey_bg_color)
+            if isinstance(vector_scale, str):
+                vector_scale = _units.Unit(vector_scale)
+            if _units.is_unit(vector_scale):
+                vector_scale = vector_scale.in_units(sim['vel'].units)
+            Q = axes.quiver(X, Y, vx, vy, scale=vector_scale, color=vector_color,
+                         edgecolor=vector_edgecolor)
+        if key:
+            from . import quiverkey
+            qk = quiverkey.PynbodyQuiverKey(Q, key_x, key_y,
+                                      key_unit.in_units(sim['vel'].units, **sim.conversion_context()),
+                                            "$" + key_unit.latex() + "$",
+                                      color=key_color, labelcolor=key_color,
+                                      boxedgecolor=key_edge_color, boxfacecolor=key_bg_color)
+            qk.set_zorder(6)
+            p.gca().add_artist(qk)
     elif mode == 'stream' :
-        Q = p.streamplot(X, Y, vx, vy, color=vector_color, density=density)
+        Q = axes.streamplot(X, Y, vx, vy, color=vector_color, density=stream_density, linewidth=stream_linewidth)
 
-	# RS - if a axis object is passed in, use the right limit call
-    if subplot:
-        p.set_xlim(-width/2, width/2)
-        p.set_ylim(-width/2, width/2)
-    else:
-        p.xlim(-width/2, width/2)
-        p.ylim(-width/2, width/2)
+    axes.set_xlim(-width/2, width/2)
+    axes.set_ylim(-width/2, width/2)
+
 
     return im
 
@@ -150,27 +310,46 @@ def volume(sim, qty='rho', width=None, resolution=200,
            create_figure=True):
     """Create a volume rendering of the given simulation using mayavi.
 
-    **Keyword arguments:**
+    .. warning ::
+        This function requires mayavi to be installed. However, mayavi does not seem to be under
+        active development and is not compatible with the latest versions of python.
+        As a result, this function will probably be removed in future versions of pynbody.
+        For a more modern alternative, consider using `topsy <https://github.com/pynbody/topsy/>`_.
 
-    *qty* (rho): The name of the array to interpolate
+    Parameters
+    ----------
+    sim : pynbody.snapshot.simsnap.SimSnap
+        The simulation snapshot to visualize
 
-    *width* (None): The width of the cube to generate, centered on the origin
+    qty : str, optional
+        The name of the array to interpolate. Default is 'rho', which gives a density image.
 
-    *resolution* (200): The number of elements along each side of the cube
+    width : str or float, optional
+        The width of the cube to generate, centered on the origin. If None, the width is determined
+        by the extent of the simulation snapshot.
 
-    *color* (white): The color of the volume rendering. The value of each voxel
-       is used to set the opacity.
+    resolution : int, optional
+        The number of elements along each side of the cube. (Default is 200)
 
-    *vmin* (None): The value for zero opacity (calculated using dynamic_range if None)
+    color : tuple, optional
+        The color of the volume rendering. The value of each voxel is used to set the opacity.
 
-    *vmax* (None): The value for full opacity (calculated from the maximum
-       value in the region if None)
+    vmin : float, optional
+        The value for zero opacity. If None, this is inferred from vmax and dynamic_range.
 
-    *dynamic_range*: The dynamic range to use if vmin and vmax are not specified
+    vmax : float, optional
+        The value for full opacity. If None, the maximum value of the image is used.
 
-    *log* (True): log-scale the image before passing to mayavi
+    dynamic_range : float, optional
+        The dynamic range in dex to use if vmin and vmax are not specified.
+        Default is 4.0
 
-    *create_figure* (True): create a new mayavi figure before rendering
+    log : bool, optional
+        If True, the image is log-scaled before passing to mayavi. (Default is True)
+
+    create_figure : bool, optional
+        If True, create a new mayavi figure before rendering. (Default is True)
+
     """
 
     import mayavi
@@ -231,205 +410,192 @@ def volume(sim, qty='rho', width=None, resolution=200,
 
     return V
 
-def contour(*args, **kwargs):
-    """
-    Make an SPH image of the given simulation and render it as contours.
-    nlevels and levels are passed to pyplot's contour command.
-
-    Other arguments are as for *image*.
-    """
-
-    import copy
-    kwargs_image = copy.copy(kwargs)
-    nlevels = kwargs_image.pop('nlevels',None)
-    levels = kwargs_image.pop('levels',None)
-    width = kwargs_image.get('width','10 kpc')
-    kwargs_image['noplot']=True
-    im = image(*args, **kwargs_image)
-    res = im.shape
-
-    units = kwargs_image.get('units',None)
-
-    if isinstance(width, str) or issubclass(width.__class__, _units.UnitBase):
-        if isinstance(width, str):
-            width = _units.Unit(width)
-        sim = args[0]
-        width = width.in_units(sim['pos'].units, **sim.conversion_context())
-
-    width = float(width)
-    x,y = np.meshgrid(np.linspace(-width/2,width/2,res[0]),np.linspace(-width/2,width/2,res[0]))
-
-    p.contour(x,y,im,nlevels=nlevels,levels=levels)
-
 
 def _units_imply_projection(sim, qty, units):
+    if isinstance(qty, str):
+        qty = sim[qty]
+
     try:
-        sim[qty].units.ratio(units, **sim[qty].conversion_context())
+        qty.units.ratio(units, **sim.conversion_context())
         # if this fails, perhaps we're requesting a projected image?
         return False
     except _units.UnitsException:
         # if the following fails, there's no interpretation this routine
         # can cope with. The error will be allowed to propagate.
-        sim[qty].units.ratio(
-            units / (sim['x'].units), **sim[qty].conversion_context())
+        qty.units.ratio(
+            units / (sim['x'].units), **sim.conversion_context())
         return True
 
 
-def image(sim, qty='rho', width="10 kpc", resolution=500, units=None, log=True,
-          vmin=None, vmax=None, av_z=False, filename=None,
-          z_camera=None, clear=True, cmap=None,
-          title=None, qtytitle=None, show_cbar=True, subplot=False,
-          noplot=False, ret_im=False, fill_nan=True, fill_val=0.0, linthresh=None,
-          kernel_type='spline',
+def image(sim, qty='rho', width="10 kpc", resolution=None, units=None, log=True,
+          vmin=None, vmax=None, weight=False, z_camera=None, clear=True, cmap=None,
+          title=None, colorbar_label=None, qtytitle=None, show_cbar=True, axes=None,
+          noplot=False, filename=None,
+          return_image=False, return_array=False,
+          fill_nan=True, fill_val=0.0, linthresh=None,
+          restrict_depth = False, threaded=True, approximate_fast=None, denoise=None,
+          kernel=None,
           **kwargs):
     """
+    Make an image of the given simulation, using SPH or denoised-SPH interpolation.
 
-    Make an SPH image of the given simulation.
+    Parameters
+    ----------
+    sim : pynbody.snapshot.simsnap.SimSnap
+        The simulation snapshot to plot. The image is generated in the plane z=0, or
+        projected along the z axis.
 
-    **Keyword arguments:**
+    qty : str | pynbody.array.SimArray
+        The name of the array to interpolate. Default is 'rho', which gives a density
+        image. Alternatively, an array can be passed in.
 
-    *qty* (rho): The name of the array to interpolate
+    width : str or float, optional
+        The overall width and height of the plot. If a float, it is assumed
+        to be in units of sim['pos']. It can also be passed in as a string indicating
+        the units, e.g. '10 kpc'. (Default is '10 kpc')
 
-    *width* (10 kpc): The overall width and height of the plot. If
-     ``width`` is a float or an int, then it is assumed to be in units
-     of ``sim['pos']``. It can also be passed in as a string
-     indicating the units, i.e. '10 kpc', in which case it is
-     converted to units of ``sim['pos']``.
+    resolution : int, optional
+        The number of pixels wide and tall. (Default is determined by the
+        :ref:`configuration file <configuration>`.)
 
-    *resolution* (500): The number of pixels wide and tall
+    units : str or pynbody.units.Unit, optional
+        The units of the output. Default is None, in which case the units of the input
+        quantity are used. If the units correspond to integrating the quantity along
+        a spatial dimension, the output is a projected image. For example, if the units
+        are 'Msol kpc^-2', and the quantity is 'rho', the output is a projected image of
+        the surface density.
 
-    *units* (None): The units of the output
+    log : bool, optional
+        If True, the image is log-scaled. (Default is True)
 
-    *av_z* (False): If True, the requested quantity is averaged down
-            the line of sight (default False: image is generated in
-            the thin plane z=0, unless output units imply an integral
-            down the line of sight). If a string, the requested quantity
-            is averaged down the line of sight weighted by the av_z
-            array (e.g. use 'rho' for density-weighted quantity;
-            the default results when av_z=True are volume-weighted).
+    vmin : float, optional
+        The minimum value for the color scale. If None, the minimum value of the image.
 
-    *z_camera* (None): If set, a perspective image is rendered. See
-                :func:`pynbody.sph.image` for more details.
+    vmax : float, optional
+        The maximum value for the color scale. If None, the maximum value of the image.
 
-    *filename* (None): if set, the image will be saved in a file
+    weight : bool or str, optional
+        If True, the requested quantity is volume-averaged down the line of sight. If a string, the
+        requested quantity is averaged down the line of sight weighted by the named array
+        (e.g. use 'rho' for density-weighted quantity).
 
-    *clear* (True): whether to call clf() on the axes first
+    restrict_depth : bool, optional
+        If True, restrict the depth of the image to the width of the image. (Default is False)
 
-    *cmap* (None): user-supplied colormap instance
+    z_camera : float, optional
+        If set, a perspective image is rendered, as though the camera is a pinhole camera
+        at (0,0,z_camera). The frustrum is defined by the width of the image in the plane z=0.
 
-    *title* (None): plot title
+    clear : bool, optional
+        Whether to clear the axes before plotting. (Default is True)
 
-    *qtytitle* (None): colorbar quantity title
+    cmap : matplotlib.colors.Colormap or str, optional
+        Colormap to use. If None, the default colormap is used.
 
-    *show_cbar* (True): whether to plot the colorbar
+    title : str, optional
+        Plot title.
 
-    *subplot* (False): the user can supply a AxesSubPlot instance on
-    which the image will be shown
+    colorbar_label : str, optional
+        Colorbar label. If not provided, one will be generated from the quantity name and units.
 
-    *noplot* (False): do not display the image, just return the image array
+    show_cbar : bool, optional
+        Whether to automatically plot the colorbar. (Default is True)
 
-    *ret_im* (False): return the image instance returned by imshow
+    axes : matplotlib.axes.Axes, optional
+        Axes instance on which the image will be shown; if None, the current pyplot figure is
+        used. (Default is False)
 
-    *num_threads* (None) : if set, specify the number of threads for
-    the multi-threaded routines; otherwise the pynbody.config default is used
+    noplot : bool, optional
+        If True, the image is not displayed, only the image array is returned. This option therefore
+        implies return_array = True.
 
-    *fill_nan* (True): if any of the image values are NaN, replace with fill_val
+    filename : str, optional
+        If set, the image will be saved in a file.
 
-    *fill_val* (0.0): the fill value to use when replacing NaNs
+    return_image : bool, optional
+        If True, the image instance returned by imshow is returned. (Default is False)
 
-    *linthresh* (None): if the image has negative and positive values
-     and a log scaling is requested, the part between `-linthresh` and
-     `linthresh` is shown on a linear scale to avoid divergence at 0
+    return_array : bool, optional
+        If True, the numpy array of the image is returned. (Default is False)
 
-    *kernel_type* ('spline'): SPH kernel to use for smoothing. Defaults to a
-     cubic spline, but can also be set to 'wendlandC2'
+    fill_nan : bool, optional
+        If any of the image values are NaN, replace with fill_val. (Default is True)
+
+    fill_val : float, optional
+        The fill value to use when replacing NaNs. (Default is 0.0)
+
+    linthresh : float, optional
+        If the image has negative and positive values and a log scaling is requested, the part
+        between -linthresh and linthresh is shown on a linear scale to avoid divergence at 0.
+
+    kernel : str, optional
+        SPH kernel to use for smoothing; see :func:`~pynbody.sph.kernels.create_kernel` for options.
+
+    approximate_fast : bool, optional
+        If True, speed up the image-making by rendering large kernels onto a lower-resolution image
+        first, which is then interpolated to the final resolution. (Default is set in the config file).
+
+    threaded : bool, optional
+        If True, use threads to parallelise the rendering. (Default is set in the config file).
+
+    qtytitle : str, optional
+        Deprecated alias for colorbar_label.
+
+    av_z : bool or str, optional
+        Deprecated alias for weight.
+
+    Returns
+    -------
+
+    matplotlib.image.AxesImage
+        The image instance returned by imshow, if return_image is True.
+
+    numpy.ndarray
+        The image array, if return_array is True or noplot is True.
+
     """
+
+    global config
 
     if not noplot:
         import matplotlib.pylab as plt
-
-    global config
-    if not noplot:
-        if subplot:
-            p = subplot
+        if axes:
+            p = axes
         else:
             p = plt
 
-    if qtytitle is None:
+    if qtytitle is not None:
+        warnings.warn("qtytitle is deprecated; use colorbar_label instead", DeprecationWarning)
+        colorbar_label = qtytitle
+
+    if kwargs.get('av_z', None) is not None:
+        weight = kwargs.pop('av_z')
+        warnings.warn("av_z is deprecated; use weight instead", DeprecationWarning)
+
+    if kwargs.get('ret_im', None) is not None:
+        return_image = kwargs.pop('ret_im')
+        warnings.warn("ret_im is deprecated; use return_image instead", DeprecationWarning)
+
+    if colorbar_label is None and isinstance(qty, str):
         qtytitle = qty
-
-    if isinstance(units, str):
-        units = _units.Unit(units)
-
-    if isinstance(width, str) or issubclass(width.__class__, _units.UnitBase):
-        if isinstance(width, str):
-            width = _units.Unit(width)
-        width = width.in_units(sim['pos'].units, **sim.conversion_context())
-
-    width = float(width)
-
-    if kernel_type == 'wendlandC2':
-        kernel = sph.WendlandC2Kernel()
-    elif kernel_type == 'spline':
-        kernel = sph.Kernel()
     else:
-        raise ValueError('Invalid kernel_type specified. Options are "spline" (default) and "wendlandC2".')
+        qtytitle = None
 
-    perspective = z_camera is not None
-    if perspective and not av_z:
-        kernel = sph.Kernel2D(kernel)
+    if weight and qtytitle:
+        qtytitle = f"$\\langle${qtytitle}$\\rangle$"
 
-    is_projected = False
-    if units is not None:
-        is_projected = _units_imply_projection(sim, qty, units)
+    renderer = renderers.make_render_pipeline(sim, quantity=qty, width=width, resolution=resolution,
+                                              out_units=units, weight=weight, restrict_depth=restrict_depth,
+                                              kernel=kernel, z_camera=z_camera, threaded=threaded,
+                                              approximate_fast=approximate_fast, denoise=denoise)
 
-    if is_projected:
-        kernel = sph.Kernel2D(kernel)
+    # if width was provided e.g. as string, we'll need it as a float
+    width = renderer.geometry.width
 
-    if av_z:
-        if isinstance(kernel, sph.Kernel2D):
-            raise _units.UnitsException(
-                "Units already imply projected image; can't also average over line-of-sight!")
-        else:
-            kernel = sph.Kernel2D(kernel)
-            if units is not None:
-                aunits = units * sim['z'].units
-            else:
-                aunits = None
+    if renderer.is_projected and qtytitle is not None:
+        qtytitle = f"$\\int\\,${qtytitle}$\\,\\mathrm{{d}}z$"
 
-            if isinstance(av_z, str):
-                if units is not None:
-                    aunits = units * sim[av_z].units * sim['z'].units
-                sim["__prod"] = sim[av_z] * sim[qty]
-                qty = "__prod"
-
-            else:
-                av_z = "__one"
-                sim["__one"] = np.ones_like(sim[qty])
-                sim["__one"].units = "1"
-
-            im = sph.render_image(sim, qty, width / 2, resolution, out_units=aunits, kernel=kernel,
-                                  z_camera=z_camera, **kwargs)
-            im2 = sph.render_image(sim, av_z, width / 2, resolution, kernel=kernel,
-                                   z_camera=z_camera, **kwargs)
-
-            top = sim.ancestor
-
-            try:
-                del top["__one"]
-            except KeyError:
-                pass
-
-            try:
-                del top["__prod"]
-            except KeyError:
-                pass
-
-            im = im / im2
-
-    else:
-        im = sph.render_image(sim, qty, width / 2, resolution, out_units=units,
-                              kernel=kernel,  z_camera=z_camera, **kwargs)
+    im = renderer.render()
 
     if fill_nan:
         im[np.isnan(im)] = fill_val
@@ -462,18 +628,14 @@ def image(sim, qty='rho', width="10 kpc", resolution=500, units=None, log=True,
         #
         # do the actual plotting
         #
-        if clear and not subplot:
+        if clear and not axes:
             p.clf()
-
-        if ret_im:
-            return p.imshow(im[::-1, :].view(np.ndarray), extent=(-width / 2, width / 2, -width / 2, width / 2),
-                            cmap=cmap, norm = norm)
 
         ims = p.imshow(im[::-1, :].view(np.ndarray), extent=(-width / 2, width / 2, -width / 2, width / 2),
                        cmap=cmap, norm = norm)
 
         u_st = sim['pos'].units.latex()
-        if not subplot:
+        if not axes:
             plt.xlabel("$x/%s$" % u_st)
             plt.ylabel("$y/%s$" % u_st)
         else:
@@ -483,7 +645,8 @@ def image(sim, qty='rho', width="10 kpc", resolution=500, units=None, log=True,
         if units is None:
             units = im.units
 
-
+        if not isinstance(units, _units.UnitBase):
+            units = _units.Unit(units)
 
         if units.latex() == "":
             units=""
@@ -491,16 +654,15 @@ def image(sim, qty='rho', width="10 kpc", resolution=500, units=None, log=True,
             units = "$"+units.latex()+"$"
 
         if show_cbar:
-             plt.colorbar(ims).set_label(qtytitle+"/"+units)
+            colorbar = plt.colorbar(ims)
+            if colorbar_label is not None:
+                colorbar.set_label(colorbar_label)
+            elif qtytitle is not None:
+                colorbar.set_label(qtytitle+"/"+units)
 
-        # colorbar doesn't work wtih subplot:  mappable is NoneType
-        # elif show_cbar:
-        #    import matplotlib.pyplot as mpl
-        #    if qtytitle: mpl.colorbar().set_label(qtytitle)
-        #    else:        mpl.colorbar().set_label(units)
 
         if title is not None:
-            if not subplot:
+            if not axes:
                 p.title(title)
             else:
                 p.set_title(title)
@@ -508,34 +670,10 @@ def image(sim, qty='rho', width="10 kpc", resolution=500, units=None, log=True,
         if filename is not None:
             p.savefig(filename)
 
-        plt.draw()
-        # plt.show() - removed by AP on 30/01/2013 - this should not be here as
-        # for some systems you don't get back to the command prompt
 
-    return im
-
-
-def image_radial_profile(im, bins=100):
-
-    xsize, ysize = np.shape(im)
-    x = np.arange(-xsize / 2, xsize / 2)
-    y = np.arange(-ysize / 2, ysize / 2)
-    xs, ys = np.meshgrid(x, y)
-    rs = np.sqrt(xs ** 2 + ys ** 2)
-    hist, bin_edges = np.histogram(rs, bins=bins)
-    inds = np.digitize(rs.flatten(), bin_edges)
-    ave_vals = np.zeros(bin_edges.size)
-    max_vals = np.zeros(bin_edges.size)
-    min_vals = np.zeros(bin_edges.size)
-    for i in np.arange(bin_edges.size):
-        try:
-            min_vals[i] = np.min(10 ** (im.flatten()[np.where(inds == i)]))
-        except ValueError:
-            min_vals[i] = float('nan')
-        ave_vals[i] = np.mean(10 ** (im.flatten()[np.where(inds == i)]))
-        try:
-            max_vals[i] = np.max(10 ** (im.flatten()[np.where(inds == i)]))
-        except ValueError:
-            max_vals[i] = float('nan')
-
-    return ave_vals, min_vals, max_vals, bin_edges
+    if return_image and return_array:
+        return ims, im
+    elif return_image:
+        return ims
+    elif return_array or noplot:
+        return im
