@@ -1,11 +1,13 @@
 """
+Implements reading old-style gadget binary files (format 1 and 2, but not HDF5).
 
-gadget
-======
+The gadget array names are mapped into pynbody array names according to the mappings given by the `config.ini`
+section `[gadget-name-mapping]`.
 
-Implements classes and functions for handling gadget files; you rarely
-need to access this module directly as it will be invoked
-automatically via pynbody.load.
+Very old gadget-1 style files have a fixed block order, which is specified in the config.ini section
+`[gadget-1-blocks]`.
+
+
 
 """
 
@@ -14,13 +16,15 @@ import configparser
 import copy
 import errno
 import itertools
+import os
+import pathlib
 import struct
 import sys
 import warnings
 
 import numpy as np
 
-from .. import array, config, config_parser, family, units
+from .. import array, config_parser, family, units
 from . import SimSnap, namemapper
 
 # This is set here and not in a config file because too many things break
@@ -65,7 +69,7 @@ def gadget_type(fam):
         return _type_map[fam]
 
 
-class GadgetBlock:
+class _GadgetBlock:
 
     """Class to describe each block.
     Each block has a start, a length, and a length-per-particle"""
@@ -123,8 +127,8 @@ def _construct_gadget_header(data, endian='='):
      NallHW[0], NallHW[1], NallHW[2], NallHW[3], NallHW[4], NallHW[5],
      flag_entropy_instead_u, flag_doubleprecision, flag_ic_info, lpt_scalingfactor, fill) = struct.unpack(fmt, data)
 
-    header = GadgetHeader(npart, mass, time, redshift,
-                          BoxSize, Omega0, OmegaLambda, HubbleParam, num_files)
+    header = _GadgetHeader(npart, mass, time, redshift,
+                           BoxSize, Omega0, OmegaLambda, HubbleParam, num_files)
     header.flag_sfr = flag_sfr
     header.flag_feedback = flag_feedback
     header.npartTotal = npartTotal
@@ -141,7 +145,7 @@ def _construct_gadget_header(data, endian='='):
     return header
 
 
-class GadgetHeader:
+class _GadgetHeader:
 
     """Describes the header of gadget class files; this is all our metadata, so we are going to store it inline"""
 
@@ -248,10 +252,10 @@ class GadgetHeader:
         return data
 
 
-class GadgetFile:
+class _GadgetFile:
 
     """Gadget file management class. Users should access gadget files through
-    :class:`~pynbody.gadget.GadgetSnap`."""
+    :class:`~pynbody.snapshot.gadget.GadgetSnap`."""
 
     def __init__(self, filename):
         self._filename = filename
@@ -259,110 +263,107 @@ class GadgetFile:
         self.endian = ''
         self.format2 = True
         t_part = 0
-        fd = open(filename, "rb")
-        self.check_format(fd)
-        # If format 1, load the block definitions.
-        if not self.format2:
-            self.block_names = config_parser.get(
-                'gadget-1-blocks', "blocks").split(",")
-            self.block_names = [q.upper().ljust(4) for q in self.block_names]
-            if sys.version_info[0] > 2:
-                self.block_names = [str.encode(x, 'utf-8') for x in self.block_names]
-            # This is a counter for the fallback
-            self.extra = 0
-        while True:
-            block = GadgetBlock()
-            (name, block.length) = self.read_block_head(fd)
-            if block.length == 0:
-                break
-            # Do special things for the HEAD block
-            if name[0:4] == b"HEAD":
-                if block.length != 256:
-                    raise OSError("Mis-sized HEAD block in " + filename)
-                self.header = fd.read(256)
-                if len(self.header) != 256:
-                    raise OSError("Could not read HEAD block in " + filename)
-                self.header = _construct_gadget_header(
-                    self.header, self.endian)
+        with open(filename, "rb") as fd:
+            self.check_format(fd)
+            # If format 1, load the block definitions.
+            if not self.format2:
+                self.block_names = config_parser.get(
+                    'gadget-1-blocks', "blocks").split(",")
+                self.block_names = [q.upper().ljust(4) for q in self.block_names]
+                if sys.version_info[0] > 2:
+                    self.block_names = [str.encode(x, 'utf-8') for x in self.block_names]
+                # This is a counter for the fallback
+                self.extra = 0
+            while True:
+                block = _GadgetBlock()
+                (name, block.length) = self.read_block_head(fd)
+                if block.length == 0:
+                    break
+                # Do special things for the HEAD block
+                if name[0:4] == b"HEAD":
+                    if block.length != 256:
+                        raise OSError("Mis-sized HEAD block in " + filename)
+                    self.header = fd.read(256)
+                    if len(self.header) != 256:
+                        raise OSError("Could not read HEAD block in " + filename)
+                    self.header = _construct_gadget_header(
+                        self.header, self.endian)
+                    record_size = self.read_block_foot(fd)
+                    if record_size != 256:
+                        raise OSError("Bad record size for HEAD in " + filename)
+                    t_part = self.header.npart.sum()
+                    if  ((not self.format2) and
+                        ((self.header.npart != 0) * (self.header.mass == 0)).sum()==0):
+                        # The "Spec" says that if all the existing particle masses
+                        # are in the header, we shouldn't have a MASS block
+                        self.block_names.remove(b"MASS")
+                    continue
+                # Set the partlen, using our amazing heuristics
+                success = False
+                if name[0:4] == b"POS " or name[0:4] == b"VEL ":
+                    if block.length == t_part * 24:
+                        block.partlen = 24
+                        block.data_type = np.float64
+                    else:
+                        block.partlen = 12
+                        block.data_type = np.float32
+                    block.p_types = self.header.npart != 0
+                    success = True
+                elif name[0:4] == b"ID  ":
+                    # Heuristic for long (64-bit) IDs
+                    if block.length == t_part * 4:
+                        block.partlen = 4
+                        block.data_type = np.int32
+                    else:
+                        block.partlen = 8
+                        block.data_type = np.int64
+                    block.p_types = self.header.npart != 0
+                    success = True
+
+                block.start = fd.tell()
+                # Check for the case where the record size overflows an int.
+                # If this is true, we can't get record size from the length and we just have to guess
+                # At least the record sizes at either end should be consistently wrong.
+                # Better hope this only happens for blocks where all particles are
+                # present.
+                extra_len = int(t_part) * block.partlen
+                if extra_len >= 2 ** 32:
+                    fd.seek(extra_len, 1)
+                else:
+                    fd.seek(block.length, 1)
                 record_size = self.read_block_foot(fd)
-                if record_size != 256:
-                    raise OSError("Bad record size for HEAD in " + filename)
-                t_part = self.header.npart.sum()
-                if  ((not self.format2) and
-                	((self.header.npart != 0) * (self.header.mass == 0)).sum()==0):
-                    # The "Spec" says that if all the existing particle masses
-                    # are in the header, we shouldn't have a MASS block
-                    self.block_names.remove(b"MASS")
-                continue
-            # Set the partlen, using our amazing heuristics
-            success = False
-            if name[0:4] == b"POS " or name[0:4] == b"VEL ":
-                if block.length == t_part * 24:
-                    block.partlen = 24
-                    block.data_type = np.float64
+                if record_size != block.length:
+                    raise OSError("Corrupt record in " +
+                                filename + " footer for block " + name + "dtype" + str(block.data_type))
+                if extra_len >= 2 ** 32:
+                    block.length = extra_len
+
+                if not success:
+                    # Figure out what particles are here and what types
+                    # they have. This also is a heuristic, which assumes
+                    # that blocks are either fully present or not for a
+                    # given particle. It also has to try all
+                    # possibilities of dimensions of array and data type.
+                    for dim, tp in (1, np.float32), (1, np.float64), (3, np.float32), (3, np.float64), (11, np.float32):
+                        try:
+                            block.data_type = tp
+                            block.partlen = np.dtype(tp).itemsize * dim
+                            block.p_types = self.get_block_types(
+                                block, self.header.npart)
+                            success = True
+                            break
+                        except ValueError:
+                            continue
+
+                if not success:
+                    warnings.warn("Encountered a gadget block %r which could not be interpreted - is it a strange length or data type (length=%d)?" %
+                                (name, block.length), RuntimeWarning)
                 else:
-                    block.partlen = 12
-                    block.data_type = np.float32
-                block.p_types = self.header.npart != 0
-                success = True
-            elif name[0:4] == b"ID  ":
-                # Heuristic for long (64-bit) IDs
-                if block.length == t_part * 4:
-                    block.partlen = 4
-                    block.data_type = np.int32
-                else:
-                    block.partlen = 8
-                    block.data_type = np.int64
-                block.p_types = self.header.npart != 0
-                success = True
-
-            block.start = fd.tell()
-            # Check for the case where the record size overflows an int.
-            # If this is true, we can't get record size from the length and we just have to guess
-            # At least the record sizes at either end should be consistently wrong.
-            # Better hope this only happens for blocks where all particles are
-            # present.
-            extra_len = int(t_part) * block.partlen
-            if extra_len >= 2 ** 32:
-                fd.seek(extra_len, 1)
-            else:
-                fd.seek(block.length, 1)
-            record_size = self.read_block_foot(fd)
-            if record_size != block.length:
-                raise OSError("Corrupt record in " +
-                              filename + " footer for block " + name + "dtype" + str(block.data_type))
-            if extra_len >= 2 ** 32:
-                block.length = extra_len
-
-            if not success:
-                # Figure out what particles are here and what types
-                # they have. This also is a heuristic, which assumes
-                # that blocks are either fully present or not for a
-                # given particle. It also has to try all
-                # possibilities of dimensions of array and data type.
-                for dim, tp in (1, np.float32), (1, np.float64), (3, np.float32), (3, np.float64), (11, np.float32):
-                    try:
-                        block.data_type = tp
-                        block.partlen = np.dtype(tp).itemsize * dim
-                        block.p_types = self.get_block_types(
-                            block, self.header.npart)
-                        success = True
-                        break
-                    except ValueError:
-                        continue
-
-            if not success:
-                warnings.warn("Encountered a gadget block %r which could not be interpreted - is it a strange length or data type (length=%d)?" %
-                              (name, block.length), RuntimeWarning)
-            else:
-                self.blocks[name[0:4]] = block
-
-        # and we're done.
-        fd.close()
+                    self.blocks[name[0:4]] = block
 
         # Make a mass block if one isn't found.
         if b'MASS' not in self.blocks:
-            block = GadgetBlock()
+            block = _GadgetBlock()
             block.length = 0
             block.start = 0
 
@@ -471,14 +472,14 @@ class GadgetFile:
         p_start = self.get_start_part(name, p_type)
         if p_toread > parts:
             p_toread = parts
-        fd = open(self._filename, 'rb')
-        fd.seek(cur_block.start + int(cur_block.partlen * p_start), 0)
-        # This is just so that we can get a size for the type
-        dt = np.dtype(cur_block.data_type)
-        n_type = p_toread * cur_block.partlen // dt.itemsize
-        data = np.fromfile(
-            fd, dtype=cur_block.data_type, count=n_type, sep='')
-        fd.close()
+        with open(self._filename, 'rb') as fd:
+            fd.seek(cur_block.start + int(cur_block.partlen * p_start), 0)
+            # This is just so that we can get a size for the type
+            dt = np.dtype(cur_block.data_type)
+            n_type = p_toread * cur_block.partlen // dt.itemsize
+            data = np.fromfile(
+                fd, dtype=cur_block.data_type, count=n_type, sep='')
+
         if self.endian != '=':
             data = data.byteswap(True)
         return (p_toread, data)
@@ -538,33 +539,30 @@ class GadgetFile:
         if bt.kind != dt.kind:
             raise ValueError("Data of incorrect type passed to write_block")
         # Open the file
-        if filename == None:
-            fd = open(self._filename, "r+b")
-        else:
-            fd = open(filename, "r+b")
-        # Seek to the start of the block
-        fd.seek(cur_block.start + cur_block.partlen * p_start, 0)
-        # Add the block header if we are at the start of a block
-        if p_type == MinType or p_type < 0:
-            data = self.write_block_header(name, cur_block.length)
-            # Better seek back a bit first.
-            fd.seek(-len(data), 1)
-            fd.write(data)
+        fn = self._filename if filename is None else filename
 
-        if self.endian != '=':
-            big_data = big_data.byteswap(False)
+        with open(fn, 'r+b') as fd:
+            # Seek to the start of the block
+            fd.seek(cur_block.start + cur_block.partlen * p_start, 0)
+            # Add the block header if we are at the start of a block
+            if p_type == MinType or p_type < 0:
+                data = self.write_block_header(name, cur_block.length)
+                # Better seek back a bit first.
+                fd.seek(-len(data), 1)
+                fd.write(data)
 
-        # Actually write the data
-        # Make sure to ravel it, otherwise the wrong amount will be written,
-        # because it will also write nulls every time the first array dimension
-        # changes.
-        d = np.ravel(big_data.astype(dt)).tostring()
-        fd.write(d)
-        if p_type == MaxType or p_type < 0:
-            data = self.write_block_footer(name, cur_block.length)
-            fd.write(data)
+            if self.endian != '=':
+                big_data = big_data.byteswap(False)
 
-        fd.close()
+            # Actually write the data
+            # Make sure to ravel it, otherwise the wrong amount will be written,
+            # because it will also write nulls every time the first array dimension
+            # changes.
+            d = np.ravel(big_data.astype(dt)).tobytes()
+            fd.write(d)
+            if p_type == MaxType or p_type < 0:
+                data = self.write_block_footer(name, cur_block.length)
+                fd.write(data)
 
     def add_file_block(self, name, blocksize, partlen=4, dtype=np.float32, p_types=None):
         """Add a block to the block table at the end of the file. Do not actually write anything"""
@@ -581,7 +579,7 @@ class GadgetFile:
             dtype = np.float32  # coerce to single precision
 
         # Make new block
-        block = GadgetBlock(length=blocksize, partlen=partlen, dtype=dtype)
+        block = _GadgetBlock(length=blocksize, partlen=partlen, dtype=dtype)
         block.start = lb.start + lb.length + 6 * \
             4  # For the block header, and footer of the previous block
         if p_types is None:
@@ -618,15 +616,15 @@ class GadgetFile:
         head.npart = np.array(self.header.npart)
         data = self.write_block_header(b"HEAD", 256)
         data += head.serialize()
-        if filename == None:
+        if filename is None:
             filename = self._filename
-        # a mode will ignore the file position, and w truncates the file.
+
         try:
             fd = open(filename, "r+b")
-        except OSError as xxx_todo_changeme:
+        except OSError as error:
             # If we couldn't open it because it doesn't exist open it for
             # writing.
-            (err, strerror) = xxx_todo_changeme.args
+            (err, strerror) = error.args
             # If we couldn't open it because it doesn't exist open it for
             # writing.
             if err == errno.ENOENT:
@@ -634,18 +632,21 @@ class GadgetFile:
             # If we couldn't open it for any other reason, reraise exception
             else:
                 raise OSError(err, strerror)
-        fd.seek(0)  # Header always at start of file
-        # Write header
-        fd.write(data)
-        # Seek 48 bytes forward, to skip the padding (which may contain extra
-        # data)
-        fd.seek(48, 1)
-        data = self.write_block_footer(b"HEAD", 256)
-        fd.write(data)
-        fd.close()
+
+        try:
+            fd.seek(0)  # Header always at start of file
+            # Write header
+            fd.write(data)
+            # Seek 48 bytes forward, to skip the padding (which may contain extra
+            # data)
+            fd.seek(48, 1)
+            data = self.write_block_footer(b"HEAD", 256)
+            fd.write(data)
+        finally:
+            fd.close()
 
 
-class GadgetWriteFile (GadgetFile):
+class _GadgetWriteFile(_GadgetFile):
 
     """Class for write-only snapshots, as when we are creating a new set of files from, eg, a TipsySnap.
         Should not be used directly. block_names is a list so we can specify an on-disc ordering."""
@@ -668,14 +669,14 @@ class GadgetWriteFile (GadgetFile):
             # Add block if present for some types
             if block.types.sum():
                 b_part = npart * block.types
-                b = GadgetBlock(
+                b = _GadgetBlock(
                     start=cur_pos + header_size, partlen=block.partlen,
                     length=block.partlen * b_part.sum(), dtype=block.dtype, p_types=block.types)
                 cur_pos += b.length + header_size + footer_size
                 self.blocks[_to_raw(block.name)] = b
 
 
-class WriteBlock:
+class _WriteBlock:
 
     """Internal structure for passing data around between file and snapshot"""
 
@@ -694,36 +695,33 @@ class WriteBlock:
 
 
 class GadgetSnap(SimSnap):
+    """Class for reading Gadget-1 and Gadget-2 old-style (i.e. pre-HDF5) snapshots."""
 
-    """Main class for reading Gadget-2 snapshots. The constructor makes a map of the locations
-    of the blocks, which are then read by _load_array"""
+    def __init__(self, filename: pathlib.Path, only_header=False, must_have_paramfile=False, ignore_cosmo=False):
 
-    def __init__(self, filename, only_header=False, must_have_paramfile=False, ignore_cosmo=False):
+        filename = str(filename)
 
         global config
         super().__init__()
         self._files = []
         self._filename = filename
         self._ignore_cosmo = ignore_cosmo
-        npart = np.empty(N_TYPE)
+
         # Check whether the file exists, and get the ".0" right
-        try:
-            fd = open(filename, 'rb')
+        if os.path.exists(filename):
             files = [filename]
-        except OSError:
-            fd = open(filename + ".0", 'rb')
-            # The second time if there is an exception we let it go through
+        elif os.path.exists(filename + ".0"):
             filename = filename + ".0"
             files = None
-        fd.close()
+
         if filename[-2:] == ".0":
             self._filename = filename[:-2]
         # Read the first file and use it to get an idea of how many files we
         # are expecting.
-        first_file = GadgetFile(filename)
+        first_file = _GadgetFile(filename)
         self._files.append(first_file)
         files_expected = self._files[0].header.num_files
-        npart = np.array(self._files[0].header.npart)
+        npart = np.array(self._files[0].header.npart,dtype=np.uint64) # 64-bit necessary in numpy 2.0 because of changes to data type promotion rules in the 2 * 32 calc below
 
         if files is None:
             # we want to load all files
@@ -732,7 +730,7 @@ class GadgetSnap(SimSnap):
                      for i in range(files_expected)]
 
         for filename in files[1:]:
-            tmp_file = GadgetFile(filename)
+            tmp_file = _GadgetFile(filename)
             if not self.check_headers(tmp_file.header, self._files[0].header):
                 warnings.warn("file " + str(
                     i) + " is not part of this snapshot set!", RuntimeWarning)
@@ -745,7 +743,7 @@ class GadgetSnap(SimSnap):
         self.header = copy.deepcopy(self._files[0].header)
         self.header.npart = npart
         # Check and fix npartTotal and NallHW if they are wrong.
-        if npart is not self.header.npartTotal + 2 ** 32 * self.header.NallHW:
+        if npart is not self.header.npartTotal.astype(np.uint64) + 2 ** 32 * self.header.NallHW.astype(np.uint64):
             self.header.NallHW = npart // 2 ** 32
             self.header.npartTotal = npart - 2 ** 32 * self.header.NallHW
             for f in self._files:
@@ -781,9 +779,9 @@ class GadgetSnap(SimSnap):
                 mm = nn.lower().strip()
             else:
                 mm = nn.lower().strip().decode('utf-8')
-            if not nn in _rev_name_map:
+            if nn not in _rev_name_map:
                 _rev_name_map[nn] = mm
-            if not mm in _name_map:
+            if mm not in _name_map:
                 _name_map[mm] = nn
 
         # Use translated keys only
@@ -891,7 +889,7 @@ class GadgetSnap(SimSnap):
 
     def _load_array(self, name, fam=None):
         """Read in data from a Gadget file.
-        If fam != None, loads only data for that particle family"""
+        If fam is not None, loads only data for that particle family"""
         # g_name is the internal name
         g_name = _translate_array_name(name)
 
@@ -905,9 +903,9 @@ class GadgetSnap(SimSnap):
         ndim = self._get_array_dims(name)
 
         if ndim == 1:
-            dims = [self.get_block_parts(g_name, fam), ]
+            dims = [int(self.get_block_parts(g_name, fam)), ]
         else:
-            dims = [self.get_block_parts(g_name, fam), ndim]
+            dims = [int(self.get_block_parts(g_name, fam)), ndim]
 
         if fam is not None:
             p_types = gadget_type(fam)
@@ -920,8 +918,8 @@ class GadgetSnap(SimSnap):
         for p in p_types:
             # Special-case mass
             if g_name == b"MASS" and self.header.mass[p] != 0.:
-                data = np.append(data, self.header.mass[
-                                 p] * np.ones(self.header.npart[p], dtype=data.dtype))
+                mass_as_correct_type = self.header.mass[p].astype(self._get_array_type(name))
+                data = np.append(data, np.repeat(mass_as_correct_type, self.header.npart[p]))
             else:
                 data = np.append(data, self.__load_array(g_name, p))
 
@@ -936,7 +934,8 @@ class GadgetSnap(SimSnap):
     def __load_array(self, g_name, p_type):
        """Internal helper function for _load_array that takes a g_name and a gadget type,
        gets the data from each file and returns it as one long array."""
-       data = np.array(np.zeros(self._get_array_dims(g_name) * self.header.npart[p_type]), dtype=self._get_array_type_g(g_name))
+       # int cast necessary because numpy makes int * uint64 a float!
+       data = np.zeros(int(self._get_array_dims(g_name) * self.header.npart[p_type]), dtype=self._get_array_type_g(g_name))
        # Get a type from each file
        ipos = 0
        for f in self._files:
@@ -952,23 +951,22 @@ class GadgetSnap(SimSnap):
            ipos += iread
        return data
 
-    @staticmethod
-    def _can_load(f):
+    @classmethod
+    def _can_load(cls, f: pathlib.Path):
         """Check whether we can load the file as Gadget format by reading
         the first 4 bytes"""
-        try:
-            fd = open(f, 'rb')
-        except OSError:
-            try:
-                fd = open(f + ".0", 'rb')
-            except:
+        fname = f
+        if not f.exists():
+            fname = f.parent / (f.name + ".0")
+            if not fname.exists():
                 return False
-                # If we can't open the file, we certainly can't load it...
-        (r,) = struct.unpack('=I', fd.read(4))
-        fd.close()
+
+        with open(fname, "br") as fd:
+            r, = struct.unpack('=I', fd.read(4))
+
         # First int32 is 8 for a Gadget 2 file, or 256 for Gadget 1, or the
         # byte swapped equivalent.
-        if r == 8 or r == 134217728 or r == 65536 or r == 256:
+        if r in (8, 134217728, 65536, 256):
             return True
         else:
             return False
@@ -983,11 +981,11 @@ class GadgetSnap(SimSnap):
             all_keys = set(self.loadable_keys()).union(
                 list(self.keys())).union(self.family_keys())
             all_keys = [
-                k for k in all_keys if not k in ["x", "y", "z", "vx", "vy", "vz"]]
+                k for k in all_keys if k not in ["x", "y", "z", "vx", "vy", "vz"]]
             # This code supports (limited) format conversions
             if self.__class__ is not GadgetSnap:
                 # We need a filename if we are writing to a new type
-                if filename == None:
+                if filename is None:
                     raise Exception(
                         "Please specify a filename to write a new file.")
             # Splitting the files correctly is hard; the particles need to be reordered, and
@@ -1011,7 +1009,7 @@ class GadgetSnap(SimSnap):
                 # Construct a header
                 # npart, mass, time, redshift, BoxSize,Omega0, OmegaLambda,
                 # HubbleParam, num_files=1
-                gheader = GadgetHeader(
+                gheader = _GadgetHeader(
                     npart, np.zeros(N_TYPE, float), self.properties[
                         "a"], self.properties["z"],
                     self.properties["boxsize"].in_units(self[
@@ -1042,11 +1040,11 @@ class GadgetSnap(SimSnap):
                                 partlen = 1  # dtype.itemsize
                         except KeyError:
                             pass
-                    bb = WriteBlock(partlen, dtype=dtype, types=types, name=_translate_array_name(
+                    bb = _WriteBlock(partlen, dtype=dtype, types=types, name=_translate_array_name(
                         k).upper().ljust(4)[0:4])
                     block_names.append(bb)
                 # Create an output file
-                out_file = GadgetWriteFile(
+                out_file = _GadgetWriteFile(
                     filename, npart, block_names, gheader)
                 # Write the header
                 out_file.write_header(gheader, filename)
@@ -1066,7 +1064,7 @@ class GadgetSnap(SimSnap):
                 return
 
             # Write headers
-            if filename != None:
+            if filename is not None:
                 if np.size(self._files) > 1:
                     for i in np.arange(0, np.size(self._files)):
                         ffile = filename + "." + str(i)
@@ -1132,7 +1130,7 @@ class GadgetSnap(SimSnap):
                         if f_parts[i]==0:
                             continue
                         # Set up filename
-                        if filename != None:
+                        if filename is not None:
                             ffile = filename + "." + str(i)
                             if nfiles == 1:
                                 ffile = filename
@@ -1165,10 +1163,7 @@ def _header_suggests_cosmological(gadget_header):
     return (gadget_header.HubbleParam != 0.) and (gadget_header.Omega0 != 0.) and (gadget_header.BoxSize != 0.)
 
 @GadgetSnap.decorator
-def do_units(sim):
-    # cosmo =
-    # (sim._hdf_files['Parameters']['NumericalParameters'].attrs['ComovingIntegrationOn'])!=0
-
+def _do_units(sim):
     vel_unit = config_parser.get('gadget-units', 'vel')
     dist_unit = config_parser.get('gadget-units', 'pos')
     mass_unit = config_parser.get('gadget-units', 'mass')
@@ -1188,7 +1183,7 @@ def do_units(sim):
 
 
 @GadgetSnap.decorator
-def do_properties(sim):
+def _do_properties(sim):
     h = sim.header
 
     if not(sim._ignore_cosmo) and _header_suggests_cosmological(h):
