@@ -1,27 +1,18 @@
 """
+Implements reading HDF5 Gadget files, in various variants.
 
-gadgethdf
-=========
-
-
-Implementation of backend reader for GadgetHDF files by Andrew Pontzen.
-
-The gadget array names are mapped into pynbody array mes according
-to the mappings given by the config.ini section [gadgethdf-name-mapping].
+The gadget array names are mapped into pynbody array names according
+to the mappings given by the config.ini section ``[gadgethdf-name-mapping]``.
 
 The gadget particle groups are mapped into pynbody families according
-to the mappings specified by the config.ini section [gadgethdf-type-mapping].
+to the mappings specified by the config.ini section ``[gadgethdf-type-mapping]``.
 This can be many-to-one (many gadget particle types mapping into one
 pynbody family), but only datasets which are common to all gadget types
 will be available from pynbody.
 
-Spanned files are supported. To load a range of files snap.0, snap.1, ... snap.n,
-pass the filename 'snap'. If you pass snap.0, only that particular file will
-be loaded.
+Spanned files are supported. To load a range of files ``snap.0.hdf5``, ``snap.1.hdf5``, ... ``snap.n.hdf5``,
+pass the filename ``snap``. If you pass e.g. ``snap.2.hdf5``, only file 2 will be loaded.
 """
-
-
-  # for py2.5
 
 import configparser
 import functools
@@ -31,7 +22,7 @@ import warnings
 
 import numpy as np
 
-from .. import config_parser, family, halo, units, util
+from .. import config_parser, family, units, util
 from . import SimSnap, namemapper
 
 logger = logging.getLogger('pynbody.snapshot.gadgethdf')
@@ -57,7 +48,7 @@ for hdf_groups in _default_type_map.values():
 
 
 
-class DummyHDFData:
+class _DummyHDFData:
 
     """A stupid class to allow emulation of mass arrays for particles
     whose mass is in the header"""
@@ -76,22 +67,33 @@ class DummyHDFData:
         target[:] = self.value
 
 
-class GadgetHdfMultiFileManager:
+class _GadgetHdfMultiFileManager:
     _nfiles_groupname = "Header"
     _nfiles_attrname = "NumFilesPerSnapshot"
+    _size_from_hdf5_key = "ParticleIDs"
     _subgroup_name = None
 
     def __init__(self, filename, mode='r') :
+        filename = str(filename)
         self._mode = mode
         if h5py.is_hdf5(filename):
             self._filenames = [filename]
             self._numfiles = 1
         else:
             h1 = h5py.File(filename + ".0.hdf5", mode)
-            self._numfiles = h1[self._nfiles_groupname].attrs[self._nfiles_attrname]
+            self._numfiles = self._get_num_files(h1)
+            if hasattr(self._numfiles, "__len__"):
+                assert len(self._numfiles) == 1
+                self._numfiles = self._numfiles[0]
             self._filenames = [filename+"."+str(i)+".hdf5" for i in range(self._numfiles)]
 
         self._open_files = {}
+
+    def _get_num_files(self, first_file):
+        return first_file[self._nfiles_groupname].attrs[self._nfiles_attrname]
+
+    def __len__(self):
+        return self._numfiles
 
     def __iter__(self) :
         for i in range(self._numfiles) :
@@ -108,6 +110,12 @@ class GadgetHdfMultiFileManager:
         except KeyError :
             self._open_files[i] = next(itertools.islice(self,i,i+1))
             return self._open_files[i]
+
+    def iter_particle_groups_with_name(self, hdf_family_name):
+        for hdf in self:
+            if hdf_family_name in hdf:
+                if self._size_from_hdf5_key in hdf[hdf_family_name]:
+                    yield hdf[hdf_family_name]
 
     def get_header_attrs(self):
         return self[0].parent['Header'].attrs
@@ -141,7 +149,7 @@ class GadgetHdfMultiFileManager:
 
 
 
-class SubfindHdfMultiFileManager(GadgetHdfMultiFileManager):
+class _SubfindHdfMultiFileManager(_GadgetHdfMultiFileManager):
     _nfiles_groupname = "FOF"
     _nfiles_attrname = "NTask"
     _subgroup_name = "FOF"
@@ -149,27 +157,35 @@ class SubfindHdfMultiFileManager(GadgetHdfMultiFileManager):
 
 class GadgetHDFSnap(SimSnap):
     """
-    Class that reads HDF Gadget data
+    Class that reads HDF Gadget snapshots.
     """
 
-    _multifile_manager_class = GadgetHdfMultiFileManager
+    _multifile_manager_class = _GadgetHdfMultiFileManager
     _readable_hdf5_test_key = "PartType?"
     _size_from_hdf5_key = "ParticleIDs"
+    _namemapper_config_section = "gadgethdf-name-mapping"
 
-    def __init__(self, filename, ):
+    def __init__(self, filename):
+        """Initialise a Gadget HDF snapshot.
+
+        Spanned files are supported. To load a range of files ``snap.0.hdf5``, ``snap.1.hdf5``, ... ``snap.n.hdf5``,
+        pass the filename ``snap``. If you pass e.g. ``snap.2.hdf5``, only file 2 will be loaded.
+        """
+
         super().__init__()
 
         self._filename = filename
 
         self._init_hdf_filemanager(filename)
 
-        self._translate_array_name = namemapper.AdaptiveNameMapper('gadgethdf-name-mapping')
-        self.__init_unit_information()
+        self._translate_array_name = namemapper.AdaptiveNameMapper(self._namemapper_config_section,
+                                                                   return_all_format_names=True) # required for swift
+        self._init_unit_information()
         self.__init_family_map()
         self.__init_file_map()
         self.__init_loadable_keys()
         self.__infer_mass_dtype()
-
+        self._init_properties()
         self._decorate()
 
     def _get_hdf_header_attrs(self):
@@ -206,17 +222,13 @@ class GadgetHDFSnap(SimSnap):
         self._loadable_keys = list(self._loadable_keys)
 
     def _all_hdf_groups(self):
-        for hdf in self._hdf_files:
-            for hdf_family_name in _all_hdf_particle_groups:
-                if hdf_family_name in hdf:
-                    yield hdf[hdf_family_name]
+        for hdf_family_name in _all_hdf_particle_groups:
+            yield from self._hdf_files.iter_particle_groups_with_name(hdf_family_name)
 
     def _all_hdf_groups_in_family(self, fam):
         for hdf_family_name in self._family_to_group_map[fam]:
-            for hdf in self._hdf_files:
-                if hdf_family_name in hdf:
-                    if self._size_from_hdf5_key in hdf[hdf_family_name]:
-                        yield hdf[hdf_family_name]
+            yield from self._hdf_files.iter_particle_groups_with_name(hdf_family_name)
+
 
     def __init_file_map(self):
         family_slice_start = 0
@@ -291,7 +303,7 @@ class GadgetHDFSnap(SimSnap):
         raise RuntimeError("Not implemented")
 
     def write_array(self, array_name, fam=None, overwrite=False):
-        translated_name = self._translate_array_name(array_name)
+        translated_name = self._translate_array_name(array_name)[0]
 
         self._hdf_files.reopen_in_mode('r+')
 
@@ -351,8 +363,8 @@ class GadgetHDFSnap(SimSnap):
                 pgid = int(particle_group.name[-1])
                 mtab = particle_group.parent['Header'].attrs['MassTable'][pgid]
                 if mtab > 0:
-                    return DummyHDFData(mtab, particle_group[self._size_from_hdf5_key].size,
-                                        self._mass_dtype)
+                    return _DummyHDFData(mtab, particle_group[self._size_from_hdf5_key].size,
+                                         self._mass_dtype)
             except (IndexError, KeyError):
                 pass
 
@@ -460,8 +472,8 @@ class GadgetHDFSnap(SimSnap):
             raise OSError("No such array on disk")
         else:
 
-            translated_name = self._translate_array_name(array_name)
-            dtype, dy, units = self.__get_dtype_dims_and_units(fam, translated_name)
+            translated_names = self._translate_array_name(array_name)
+            dtype, dy, units = self.__get_dtype_dims_and_units(fam, translated_names)
 
             if array_name=='mass':
                 dtype = self._mass_dtype
@@ -488,10 +500,15 @@ class GadgetHDFSnap(SimSnap):
                 i0 = 0
                 for hdf in self._all_hdf_groups_in_family(loading_fam):
                     npart = hdf['ParticleIDs'].size
+                    if npart == 0:
+                        continue
                     i1 = i0+npart
 
-                    dataset = self._get_hdf_dataset(hdf, translated_name)
-
+                    for translated_name in translated_names:
+                        try:
+                            dataset = self._get_hdf_dataset(hdf, translated_name)
+                        except KeyError:
+                            continue
                     target_array = self[loading_fam][array_name][i0:i1]
                     assert target_array.size == dataset.size
 
@@ -499,7 +516,7 @@ class GadgetHDFSnap(SimSnap):
 
                     i0 = i1
 
-    def __get_dtype_dims_and_units(self, fam, translated_name):
+    def __get_dtype_dims_and_units(self, fam, translated_names):
         if fam is None:
             fam = self.families()[0]
 
@@ -509,18 +526,25 @@ class GadgetHDFSnap(SimSnap):
         # not all arrays are present in all hdfs so need to loop
         # until we find one
         for hdf0 in self._hdf_files:
-            try:
-                representative_dset = self._get_hdf_dataset(hdf0[
-                                                  self._family_to_group_map[fam][0]], translated_name)
-                representative_hdf = hdf0
-                if hasattr(representative_dset, "attrs"):
-                    inferred_units = self._get_units_from_hdf_attr(representative_dset.attrs)
-
-                if len(representative_dset)!=0:
-                    # suitable for figuring out everything we need to know about this array
+            for translated_name in translated_names:
+                try:
+                    representative_dset = self._get_hdf_dataset(hdf0[
+                                                      self._family_to_group_map[fam][0]], translated_name)
                     break
-            except KeyError:
+                except KeyError:
+                    continue
+
+            if representative_dset is None:
                 continue
+
+            representative_hdf = hdf0
+            if hasattr(representative_dset, "attrs"):
+                inferred_units = self._get_units_from_hdf_attr(representative_dset.attrs)
+
+            if len(representative_dset)!=0:
+                # suitable for figuring out everything we need to know about this array
+                break
+
         if representative_dset is None:
             raise KeyError("Array is not present in HDF file")
 
@@ -545,13 +569,18 @@ class GadgetHDFSnap(SimSnap):
             dtype = self._mass_dtype
         return dtype, dy, inferred_units
 
-    def __init_unit_information(self):
+    _velocity_unit_key = 'UnitVelocity_in_cm_per_s'
+    _length_unit_key = 'UnitLength_in_cm'
+    _mass_unit_key = 'UnitMass_in_g'
+    _time_unit_key = 'UnitTime_in_s'
+
+    def _init_unit_information(self):
         try:
             atr = self._hdf_files.get_unit_attrs()
         except KeyError:
             # Gadget 4 stores unit information in Parameters attr <sigh>
             atr = self._hdf_files.get_parameter_attrs()
-            if 'UnitVelocity_in_cm_per_s' not in atr.keys():
+            if self._velocity_unit_key not in atr.keys():
                 warnings.warn("No unit information found in GadgetHDF file. Using gadget default units.", RuntimeWarning)
                 vel_unit = config_parser.get('gadget-units', 'vel')
                 dist_unit = config_parser.get('gadget-units', 'pos')
@@ -561,21 +590,35 @@ class GadgetHDFSnap(SimSnap):
                 return
 
         # Define the SubFind units, we will parse the attribute VarDescriptions for these
-        vel_unit = atr['UnitVelocity_in_cm_per_s'] * units.cm / units.s
-        dist_unit = atr['UnitLength_in_cm'] * units.cm
-        mass_unit = atr['UnitMass_in_g'] * units.g
+        if self._velocity_unit_key is not None:
+            vel_unit = atr[self._velocity_unit_key]
+        else:
+            vel_unit = None
+
+        dist_unit = atr[self._length_unit_key]
+        mass_unit = atr[self._mass_unit_key]
         try:
-            time_unit = atr['UnitTime_in_s'] * units.s
+            time_unit = atr[self._time_unit_key] * units.s
         except KeyError:
             # Gadget 4 seems not to store time units explicitly <sigh>
             time_unit = dist_unit/vel_unit
 
+        if vel_unit is None:
+            # Swift files don't store the velocity explicitly
+            vel_unit = dist_unit / time_unit
+
+        temp_unit = 1.0
+
         # Create a dictionary for the units, this will come in handy later
-        unitvar = {'U_V': vel_unit, 'U_L': dist_unit, 'U_M': mass_unit,
-                   'U_T': time_unit, '[K]': units.K,
-                   'SEC_PER_YEAR': units.yr, 'SOLAR_MASS': units.Msol,
-                   'solar masses / yr': units.Msol/units.yr, 'BH smoothing': dist_unit}
-        # Some arrays like StarFormationRate don't follow the patter of U_ units
+        unitvar = {'U_V': vel_unit * units.cm/units.s, 'U_L': dist_unit * units.cm,
+                   'U_M': mass_unit * units.g,
+                   'U_T': time_unit,
+                   '[K]': temp_unit * units.K,
+                   'SEC_PER_YEAR': units.yr,
+                   'SOLAR_MASS': units.Msol,
+                   'solar masses / yr': units.Msol/units.yr,
+                   'BH smoothing': dist_unit}
+        # Some arrays like StarFormationRate don't follow the pattern of U_ units
         cgsvar = {'U_M': 'g', 'SOLAR_MASS': 'g', 'U_T': 's',
                   'SEC_PER_YEAR': 's', 'U_V': 'cm s**-1', 'U_L': 'cm', '[K]': 'K',
                   'solar masses / yr': 'g s**-1', 'BH smoothing': 'cm'}
@@ -602,7 +645,7 @@ class GadgetHDFSnap(SimSnap):
                 warnings.warn("Unable to find cosmological factors in HDF file; assuming mass is %s" % mass_unit)
 
         self._file_units_system = [units.Unit(x) for x in [
-            vel_unit, dist_unit, mass_unit, "K"]]
+            vel_unit*units.cm/units.s, dist_unit*units.cm, mass_unit*units.g, "K"]]
 
     @classmethod
     def _test_for_hdf5_key(cls, f):
@@ -624,8 +667,8 @@ class GadgetHDFSnap(SimSnap):
         if hasattr(h5py, "is_hdf5"):
             if h5py.is_hdf5(f):
                 return cls._test_for_hdf5_key(f)
-            elif h5py.is_hdf5(f+".0.hdf5"):
-                return cls._test_for_hdf5_key(f+".0.hdf5")
+            elif h5py.is_hdf5(f.with_suffix(".0.hdf5")):
+                return cls._test_for_hdf5_key(f.with_suffix(".0.hdf5"))
             else:
                 return False
         else:
@@ -634,44 +677,41 @@ class GadgetHDFSnap(SimSnap):
                     "It looks like you're trying to load HDF5 files, but python's HDF support (h5py module) is missing.", RuntimeWarning)
             return False
 
+    def _init_properties(self):
+        atr = self._get_hdf_header_attrs()
 
+        # expansion factor could be saved as redshift
+        try:
+            self.properties['a'] = atr['ExpansionFactor']
+        except KeyError:
+            self.properties['a'] = 1. / (1 + atr['Redshift'])
 
-@GadgetHDFSnap.decorator
-def do_properties(sim):
-    atr = sim._get_hdf_header_attrs()
+        # Gadget 4 stores parameters in a separate dictionary <sigh>. For older formats, this will point back to the same
+        # as the header attributes.
+        atr = self._get_hdf_parameter_attrs()
 
-    # expansion factor could be saved as redshift
-    try:
-        sim.properties['a'] = atr['ExpansionFactor']
-    except KeyError:
-        sim.properties['a'] = 1. / (1 + atr['Redshift'])
+        # not all omegas need to be specified in the attributes
+        try:
+            self.properties['omegaB0'] = atr['OmegaBaryon']
+        except KeyError:
+            pass
 
-    # Gadget 4 stores parameters in a separate dictionary <sigh>. For older formats, this will point back to the same
-    # as the header attributes.
-    atr = sim._get_hdf_parameter_attrs()
+        self.properties['omegaM0'] = atr['Omega0']
+        self.properties['omegaL0'] = atr['OmegaLambda']
+        self.properties['boxsize'] = atr['BoxSize'] * self.infer_original_units('cm')
+        self.properties['z'] = (1. / self.properties['a']) - 1
+        self.properties['h'] = atr['HubbleParam']
 
-    # not all omegas need to be specified in the attributes
-    try:
-        sim.properties['omegaB0'] = atr['OmegaBaryon']
-    except KeyError:
-        pass
+        # time unit might not be set in the attributes
+        if "Time_GYR" in atr:
+            self.properties['time'] = units.Gyr * atr['Time_GYR']
+        else:
+            from .. import analysis
+            self.properties['time'] = analysis.cosmology.age(self) * units.Gyr
 
-    sim.properties['omegaM0'] = atr['Omega0']
-    sim.properties['omegaL0'] = atr['OmegaLambda']
-    sim.properties['boxsize'] = atr['BoxSize'] * sim.infer_original_units('cm')
-    sim.properties['z'] = (1. / sim.properties['a']) - 1
-    sim.properties['h'] = atr['HubbleParam']
-
-    # time unit might not be set in the attributes
-    if "Time_GYR" in atr:
-        sim.properties['time'] = units.Gyr * atr['Time_GYR']
-    else:
-        from .. import analysis
-        sim.properties['time'] = analysis.cosmology.age(sim) * units.Gyr
-
-    for s,value in sim._get_hdf_header_attrs().items():
-        if s not in ['ExpansionFactor', 'Time_GYR', 'Time', 'Omega0', 'OmegaBaryon', 'OmegaLambda', 'BoxSize', 'HubbleParam']:
-            sim.properties[s] = value
+        for s,value in self._get_hdf_header_attrs().items():
+            if s not in ['ExpansionFactor', 'Time_GYR', 'Time', 'Omega0', 'OmegaBaryon', 'OmegaLambda', 'BoxSize', 'HubbleParam']:
+                self.properties[s] = value
 
 ###################
 # SubFindHDF class
@@ -679,16 +719,10 @@ def do_properties(sim):
 
 class SubFindHDFSnap(GadgetHDFSnap) :
     """
-    Class to read Gadget's SubFind HDF data
+    Reads the variant of Gadget HDF snapshots that include SubFind output inside the snapshot itself.
     """
-    _multifile_manager_class = SubfindHdfMultiFileManager
+    _multifile_manager_class = _SubfindHdfMultiFileManager
     _readable_hdf5_test_key = "FOF"
-
-    def __init__(self, filename) :
-        super().__init__(filename)
-
-    def halos(self) :
-        return halo.SubFindHDFHaloCatalogue(self)
 
 
 class EagleLikeHDFSnap(GadgetHDFSnap):
@@ -699,6 +733,7 @@ class EagleLikeHDFSnap(GadgetHDFSnap):
         """Load the Eagle FOF halos, or if subs is specified the Subhalos of the given FOF halo number.
 
         *subs* should be an integer specifying the parent FoF number"""
+        from .. import halo
         if subs:
             if not np.issubdtype(type(subs), np.integer):
                 raise ValueError("The subs argument must specify the group number")
@@ -706,16 +741,16 @@ class EagleLikeHDFSnap(GadgetHDFSnap):
             if len(parent_group)==0:
                 raise ValueError("No group found with id %d"%subs)
 
-            cat = halo.GrpCatalogue(parent_group,
+            cat = halo.number_array.HaloNumberCatalogue(parent_group,
                                      array="SubGroupNumber", ignore=np.max(self['SubGroupNumber']))
             cat._keep_subsnap_alive = parent_group # by default, HaloCatalogue only keeps a weakref (should this be changed?)
             return cat
         else:
-            return halo.GrpCatalogue(self, array="GroupNumber", ignore=np.max(self['GroupNumber']))
+            return halo.number_array.HaloNumberCatalogue(self, array="GroupNumber", ignore=np.max(self['GroupNumber']))
 
 ## Gadget has internal energy variable
-@GadgetHDFSnap.derived_quantity
-@SubFindHDFSnap.derived_quantity
+@GadgetHDFSnap.derived_array
+@SubFindHDFSnap.derived_array
 def u(self) :
     """Gas internal energy derived from snapshot variable or temperature"""
     try:
@@ -726,8 +761,8 @@ def u(self) :
 
     return u
 
-@GadgetHDFSnap.derived_quantity
-@SubFindHDFSnap.derived_quantity
+@GadgetHDFSnap.derived_array
+@SubFindHDFSnap.derived_array
 def p(sim) :
     """Calculate the pressure for gas particles, including polytropic equation of state gas"""
 
@@ -742,22 +777,22 @@ def p(sim) :
 
     return p
 
-@GadgetHDFSnap.derived_quantity
-@SubFindHDFSnap.derived_quantity
+@GadgetHDFSnap.derived_array
+@SubFindHDFSnap.derived_array
 def HII(sim) :
     """Number of HII ions per proton mass"""
 
     return sim.g["hydrogen"] - sim.g["HI"]
 
-@GadgetHDFSnap.derived_quantity
-@SubFindHDFSnap.derived_quantity
+@GadgetHDFSnap.derived_array
+@SubFindHDFSnap.derived_array
 def HeIII(sim) :
     """Number of HeIII ions per proton mass"""
 
     return sim.g["hetot"] - sim.g["HeII"] - sim.g["HeI"]
 
-@GadgetHDFSnap.derived_quantity
-@SubFindHDFSnap.derived_quantity
+@GadgetHDFSnap.derived_array
+@SubFindHDFSnap.derived_array
 def ne(sim) :
     """Number of electrons per proton mass, ignoring the contribution from He!"""
     ne = sim.g["HII"]  #+ sim["HeII"] + 2*sim["HeIII"]
@@ -765,49 +800,49 @@ def ne(sim) :
 
     return ne
 
-@GadgetHDFSnap.derived_quantity
-@SubFindHDFSnap.derived_quantity
+@GadgetHDFSnap.derived_array
+@SubFindHDFSnap.derived_array
 def rho_ne(sim) :
     """Electron number density per SPH particle, currently ignoring the contribution from He!"""
 
     return sim.g["ne"].in_units("m_p**-1") * sim.g["rho"].in_units("m_p cm**-3")
 
-@GadgetHDFSnap.derived_quantity
-@SubFindHDFSnap.derived_quantity
+@GadgetHDFSnap.derived_array
+@SubFindHDFSnap.derived_array
 def dm(sim) :
     """Dispersion measure per SPH particle currently ignoring n_e contribution from He """
 
     return sim.g["rho_ne"]
 
-@GadgetHDFSnap.derived_quantity
-@SubFindHDFSnap.derived_quantity
+@GadgetHDFSnap.derived_array
+@SubFindHDFSnap.derived_array
 def cosmodm(sim) :
     """Cosmological Dispersion measure per SPH particle includes (1+z) factor, currently ignoring n_e contribution from He """
 
     return sim.g["rho_ne"] * (1. + sim.g["redshift"])
-@GadgetHDFSnap.derived_quantity
-@SubFindHDFSnap.derived_quantity
+@GadgetHDFSnap.derived_array
+@SubFindHDFSnap.derived_array
 def redshift(sim) :
     """Redshift from LoS Velocity 'losvel' """
 
     return np.exp( sim['losvel'].in_units('c') ) - 1.
 
-@GadgetHDFSnap.derived_quantity
-@SubFindHDFSnap.derived_quantity
+@GadgetHDFSnap.derived_array
+@SubFindHDFSnap.derived_array
 def doppler_redshift(sim) :
     """Doppler Redshift from LoS Velocity 'losvel' using SR """
 
     return np.sqrt( (1. + sim['losvel'].in_units('c')) / (1. - sim['losvel'].in_units('c'))  ) - 1.
 
-@GadgetHDFSnap.derived_quantity
-@SubFindHDFSnap.derived_quantity
+@GadgetHDFSnap.derived_array
+@SubFindHDFSnap.derived_array
 def em(sim) :
     """Emission Measure (n_e^2) per particle to be integrated along LoS"""
 
     return sim.g["rho_ne"]*sim.g["rho_ne"]
 
-@GadgetHDFSnap.derived_quantity
-@SubFindHDFSnap.derived_quantity
+@GadgetHDFSnap.derived_array
+@SubFindHDFSnap.derived_array
 def halpha(sim) :
     """H alpha intensity (based on Emission Measure n_e^2) per particle to be integrated along LoS"""
 
@@ -826,8 +861,8 @@ def halpha(sim) :
 
     return alpha * sim["em"] # Flux erg cm^-3 s^-1 sr^-1
 
-@GadgetHDFSnap.derived_quantity
-@SubFindHDFSnap.derived_quantity
+@GadgetHDFSnap.derived_array
+@SubFindHDFSnap.derived_array
 def c_n_sq(sim) :
     """Turbulent amplitude C_N^2 for use in SM calculations (e.g. Eqn 20 of Macquart & Koay 2013 ApJ 776 2) """
 
@@ -839,20 +874,20 @@ def c_n_sq(sim) :
 
     return c_n_sq
 
-@GadgetHDFSnap.derived_quantity
-@SubFindHDFSnap.derived_quantity
+@GadgetHDFSnap.derived_array
+@SubFindHDFSnap.derived_array
 def hetot(self) :
     return self["He"]
 
-@GadgetHDFSnap.derived_quantity
-@SubFindHDFSnap.derived_quantity
+@GadgetHDFSnap.derived_array
+@SubFindHDFSnap.derived_array
 def hydrogen(self) :
     return self["H"]
 
 ## Need to use the ionisation fraction calculation here which gives ionisation fraction
 ## based on the gas temperature, density and redshift for a CLOUDY table
-@GadgetHDFSnap.derived_quantity
-@SubFindHDFSnap.derived_quantity
+@GadgetHDFSnap.derived_array
+@SubFindHDFSnap.derived_array
 def HI(sim) :
     """Fraction of Neutral Hydrogen HI use limited CLOUDY table"""
 
@@ -863,8 +898,8 @@ def HI(sim) :
 ## Need to use the ionisation fraction calculation here which gives ionisation fraction
 ## based on the gas temperature, density and redshift for a CLOUDY table, then applying
 ## selfshielding for the dense, star forming gas on the equation of state
-@GadgetHDFSnap.derived_quantity
-@SubFindHDFSnap.derived_quantity
+@GadgetHDFSnap.derived_array
+@SubFindHDFSnap.derived_array
 def HIeos(sim) :
     """Fraction of Neutral Hydrogen HI use limited CLOUDY table, assuming dense EoS gas is selfshielded"""
 
@@ -876,8 +911,8 @@ def HIeos(sim) :
 ## based on the gas temperature, density and redshift for a CLOUDY table, then applying
 ## selfshielding for the dense, star forming gas on the equation of state AND a further
 ## pressure based limit for
-@GadgetHDFSnap.derived_quantity
-@SubFindHDFSnap.derived_quantity
+@GadgetHDFSnap.derived_array
+@SubFindHDFSnap.derived_array
 def HID12(sim) :
     """Fraction of Neutral Hydrogen HI use limited CLOUDY table, using the Duffy +12a prescription for selfshielding"""
 
@@ -885,8 +920,8 @@ def HID12(sim) :
 
     return pynbody.analysis.hifrac.calculate(sim.g,ion='hi', selfshield='duffy12')
 
-@GadgetHDFSnap.derived_quantity
-@SubFindHDFSnap.derived_quantity
+@GadgetHDFSnap.derived_array
+@SubFindHDFSnap.derived_array
 def HeI(sim) :
     """Fraction of Helium HeI"""
 
@@ -894,8 +929,8 @@ def HeI(sim) :
 
     return pynbody.analysis.ionfrac.calculate(sim.g,ion='hei')
 
-@GadgetHDFSnap.derived_quantity
-@SubFindHDFSnap.derived_quantity
+@GadgetHDFSnap.derived_array
+@SubFindHDFSnap.derived_array
 def HeII(sim) :
     """Fraction of Helium HeII"""
 
@@ -903,8 +938,8 @@ def HeII(sim) :
 
     return pynbody.analysis.ionfrac.calculate(sim.g,ion='heii')
 
-@GadgetHDFSnap.derived_quantity
-@SubFindHDFSnap.derived_quantity
+@GadgetHDFSnap.derived_array
+@SubFindHDFSnap.derived_array
 def OI(sim) :
     """Fraction of Oxygen OI"""
 
@@ -912,8 +947,8 @@ def OI(sim) :
 
     return pynbody.analysis.ionfrac.calculate(sim.g,ion='oi')
 
-@GadgetHDFSnap.derived_quantity
-@SubFindHDFSnap.derived_quantity
+@GadgetHDFSnap.derived_array
+@SubFindHDFSnap.derived_array
 def OII(sim) :
     """Fraction of Oxygen OII"""
 
@@ -921,8 +956,8 @@ def OII(sim) :
 
     return pynbody.analysis.ionfrac.calculate(sim.g,ion='oii')
 
-@GadgetHDFSnap.derived_quantity
-@SubFindHDFSnap.derived_quantity
+@GadgetHDFSnap.derived_array
+@SubFindHDFSnap.derived_array
 def OVI(sim) :
     """Fraction of Oxygen OVI"""
 
@@ -930,8 +965,8 @@ def OVI(sim) :
 
     return pynbody.analysis.ionfrac.calculate(sim.g,ion='ovi')
 
-@GadgetHDFSnap.derived_quantity
-@SubFindHDFSnap.derived_quantity
+@GadgetHDFSnap.derived_array
+@SubFindHDFSnap.derived_array
 def CIV(sim) :
     """Fraction of Carbon CIV"""
 
@@ -939,8 +974,8 @@ def CIV(sim) :
 
     return pynbody.analysis.ionfrac.calculate(sim.g,ion='civ')
 
-@GadgetHDFSnap.derived_quantity
-@SubFindHDFSnap.derived_quantity
+@GadgetHDFSnap.derived_array
+@SubFindHDFSnap.derived_array
 def NV(sim) :
     """Fraction of Nitrogen NV"""
 
@@ -948,8 +983,8 @@ def NV(sim) :
 
     return pynbody.analysis.ionfrac.calculate(sim.g,ion='nv')
 
-@GadgetHDFSnap.derived_quantity
-@SubFindHDFSnap.derived_quantity
+@GadgetHDFSnap.derived_array
+@SubFindHDFSnap.derived_array
 def SIV(sim) :
     """Fraction of Silicon SiIV"""
 
@@ -957,8 +992,8 @@ def SIV(sim) :
 
     return pynbody.analysis.ionfrac.calculate(sim.g,ion='siiv')
 
-@GadgetHDFSnap.derived_quantity
-@SubFindHDFSnap.derived_quantity
+@GadgetHDFSnap.derived_array
+@SubFindHDFSnap.derived_array
 def MGII(sim) :
     """Fraction of Magnesium MgII"""
 
@@ -979,77 +1014,77 @@ XSOLS=4.0898522E-4
 XSOLCa=6.4355E-5
 XSOLFe=1.1032152E-3
 
-@GadgetHDFSnap.derived_quantity
-@SubFindHDFSnap.derived_quantity
+@GadgetHDFSnap.derived_array
+@SubFindHDFSnap.derived_array
 def feh(self) :
     minfe = np.amin(self['Fe'][np.where(self['Fe'] > 0)])
     self['Fe'][np.where(self['Fe'] == 0)]=minfe
     return np.log10(self['Fe']/self['H']) - np.log10(XSOLFe/XSOLH)
 
-@GadgetHDFSnap.derived_quantity
-@SubFindHDFSnap.derived_quantity
+@GadgetHDFSnap.derived_array
+@SubFindHDFSnap.derived_array
 def sixh(self) :
     minsi = np.amin(self['Si'][np.where(self['Si'] > 0)])
     self['Si'][np.where(self['Si'] == 0)]=minsi
     return np.log10(self['Si']/self['Si']) - np.log10(XSOLSi/XSOLH)
 
-@GadgetHDFSnap.derived_quantity
-@SubFindHDFSnap.derived_quantity
+@GadgetHDFSnap.derived_array
+@SubFindHDFSnap.derived_array
 def sxh(self) :
     minsx = np.amin(self['S'][np.where(self['S'] > 0)])
     self['S'][np.where(self['S'] == 0)]=minsx
     return np.log10(self['S']/self['S']) - np.log10(XSOLS/XSOLH)
 
-@GadgetHDFSnap.derived_quantity
-@SubFindHDFSnap.derived_quantity
+@GadgetHDFSnap.derived_array
+@SubFindHDFSnap.derived_array
 def mgxh(self) :
     minmg = np.amin(self['Mg'][np.where(self['Mg'] > 0)])
     self['Mg'][np.where(self['Mg'] == 0)]=minmg
     return np.log10(self['Mg']/self['Mg']) - np.log10(XSOLMg/XSOLH)
 
-@GadgetHDFSnap.derived_quantity
-@SubFindHDFSnap.derived_quantity
+@GadgetHDFSnap.derived_array
+@SubFindHDFSnap.derived_array
 def oxh(self) :
     minox = np.amin(self['O'][np.where(self['O'] > 0)])
     self['O'][np.where(self['O'] == 0)]=minox
     return np.log10(self['O']/self['H']) - np.log10(XSOLO/XSOLH)
 
-@GadgetHDFSnap.derived_quantity
-@SubFindHDFSnap.derived_quantity
+@GadgetHDFSnap.derived_array
+@SubFindHDFSnap.derived_array
 def nexh(self) :
     minne = np.amin(self['Ne'][np.where(self['Ne'] > 0)])
     self['Ne'][np.where(self['Ne'] == 0)]=minne
     return np.log10(self['Ne']/self['Ne']) - np.log10(XSOLNe/XSOLH)
 
-@SubFindHDFSnap.derived_quantity
+@SubFindHDFSnap.derived_array
 def hexh(self) :
     minhe = np.amin(self['He'][np.where(self['He'] > 0)])
     self['He'][np.where(self['He'] == 0)]=minhe
     return np.log10(self['He']/self['He']) - np.log10(XSOLHe/XSOLH)
 
-@GadgetHDFSnap.derived_quantity
-@SubFindHDFSnap.derived_quantity
+@GadgetHDFSnap.derived_array
+@SubFindHDFSnap.derived_array
 def cxh(self) :
     mincx = np.amin(self['C'][np.where(self['C'] > 0)])
     self['C'][np.where(self['C'] == 0)]=mincx
     return np.log10(self['C']/self['H']) - np.log10(XSOLC/XSOLH)
 
-@GadgetHDFSnap.derived_quantity
-@SubFindHDFSnap.derived_quantity
+@GadgetHDFSnap.derived_array
+@SubFindHDFSnap.derived_array
 def caxh(self) :
     mincax = np.amin(self['Ca'][np.where(self['Ca'] > 0)])
     self['Ca'][np.where(self['Ca'] == 0)]=mincax
     return np.log10(self['Ca']/self['H']) - np.log10(XSOLCa/XSOLH)
 
-@GadgetHDFSnap.derived_quantity
-@SubFindHDFSnap.derived_quantity
+@GadgetHDFSnap.derived_array
+@SubFindHDFSnap.derived_array
 def nxh(self) :
     minnx = np.amin(self['N'][np.where(self['N'] > 0)])
     self['N'][np.where(self['N'] == 0)]=minnx
     return np.log10(self['N']/self['H']) - np.log10(XSOLH/XSOLH)
 
-@GadgetHDFSnap.derived_quantity
-@SubFindHDFSnap.derived_quantity
+@GadgetHDFSnap.derived_array
+@SubFindHDFSnap.derived_array
 def ofe(self) :
     minox = np.amin(self['O'][np.where(self['O'] > 0)])
     self['O'][np.where(self['O'] == 0)]=minox
@@ -1057,8 +1092,8 @@ def ofe(self) :
     self['Fe'][np.where(self['Fe'] == 0)]=minfe
     return np.log10(self['O']/self['Fe']) - np.log10(XSOLO/XSOLFe)
 
-@GadgetHDFSnap.derived_quantity
-@SubFindHDFSnap.derived_quantity
+@GadgetHDFSnap.derived_array
+@SubFindHDFSnap.derived_array
 def mgfe(sim) :
     minmg = np.amin(sim['Mg'][np.where(sim['Mg'] > 0)])
     sim['Mg'][np.where(sim['Mg'] == 0)]=minmg
@@ -1066,8 +1101,8 @@ def mgfe(sim) :
     sim['Fe'][np.where(sim['Fe'] == 0)]=minfe
     return np.log10(sim['Mg']/sim['Fe']) - np.log10(XSOLMg/XSOLFe)
 
-@GadgetHDFSnap.derived_quantity
-@SubFindHDFSnap.derived_quantity
+@GadgetHDFSnap.derived_array
+@SubFindHDFSnap.derived_array
 def nefe(sim) :
     minne = np.amin(sim['Ne'][np.where(sim['Ne'] > 0)])
     sim['Ne'][np.where(sim['Ne'] == 0)]=minne
@@ -1075,8 +1110,8 @@ def nefe(sim) :
     sim['Fe'][np.where(sim['Fe'] == 0)]=minfe
     return np.log10(sim['Ne']/sim['Fe']) - np.log10(XSOLNe/XSOLFe)
 
-@GadgetHDFSnap.derived_quantity
-@SubFindHDFSnap.derived_quantity
+@GadgetHDFSnap.derived_array
+@SubFindHDFSnap.derived_array
 def sife(sim) :
     minsi = np.amin(sim['Si'][np.where(sim['Si'] > 0)])
     sim['Si'][np.where(sim['Si'] == 0)]=minsi
