@@ -1,63 +1,41 @@
 """
+Implements classes and functions for handling RAMSES files.
 
-ramses
-======
-
-Implements classes and functions for handling RAMSES files. AMR cells
-are loaded as particles. You rarely need to access this module
-directly as it will be invoked automatically via pynbody.load.
-
-
-For a complete demo on how to use RAMSES outputs with pynbody, look at
-the `ipython notebook demo
-<http://nbviewer.ipython.org/github/pynbody/pynbody/blob/master/examples/notebooks/pynbody_demo-ramses.ipynb>`_
+Note that AMR cells are loaded as particles, although this works surprisingly well for most analysis purposes.
+In particular SPH image generation routines automatically switch on an appropriate de-noising scheme to provide
+effective interpolation between cells.
 
 """
-
-  # for py2.5
-
 
 import csv
 import logging
 import os
 import re
-import time
 import warnings
 from pathlib import Path
 
 import numpy as np
 
-from .. import analysis, array, config, config_parser, family, units
+import pynbody.array.shared
+
+from .. import array, config_parser, family, units
+from ..analysis.cosmology import age
 from ..extern.cython_fortran_utils import FortranFile
 from . import SimSnap, namemapper
 
 logger = logging.getLogger('pynbody.snapshot.ramses')
 
-from collections import OrderedDict
-
 multiprocess_num = int(config_parser.get('ramses', "parallel-read"))
 multiprocess = (multiprocess_num > 1)
 
-issue_multiprocess_warning = False
-
 if multiprocess:
-    try:
-        import multiprocessing
+    import multiprocessing
 
-        import posix_ipc
-        remote_exec = array.shared_array_remote
-        remote_map = array.remote_map
-    except ImportError:
-        issue_multiprocess_warning = True
-        multiprocess = False
-
-if not multiprocess:
+    remote_exec = pynbody.array.shared.shared_array_remote
+    remote_map = pynbody.array.shared.remote_map
+else:
     def remote_exec(fn):
-        def q(*args):
-            t0 = time.time()
-            r = fn(*args)
-            return r
-        return q
+        return fn
 
     def remote_map(*args, **kwargs):
         return list(map(*args[1:], **kwargs))
@@ -66,10 +44,9 @@ if not multiprocess:
 _float_type = 'd'
 _int_type = 'i'
 
-
 def _timestep_id(basename):
     try:
-        return re.findall("output_([0-9]*)/*$", basename)[0]
+        return re.findall("output_([0-9]*)/*$", str(basename))[0]
     except IndexError:
         return None
 
@@ -122,7 +99,7 @@ def _cpui_count_particles_with_explicit_families(filename, family_field, family_
 @remote_exec
 def _cpui_load_particle_block(filename, arrays, offset, first_index, type_, family_mask):
     with FortranFile(filename) as f:
-        header = f.read_attrs(ramses_particle_header)
+        _header = f.read_attrs(ramses_particle_header)
         f.skip(offset)
         data = f.read_vector(type_)
         for fam_id, ar in enumerate(arrays):
@@ -380,7 +357,8 @@ positive_typemap = [family.get_family(str.strip(x)) for x in config_parser.get('
 
 negative_typemap = [family.get_family(str.strip(x)) for x in config_parser.get('ramses', 'type-mapping-negative').split(",")]
 
-def read_descriptor(fname):
+def _read_descriptor(fname):
+    """Read a RAMSES file descriptor and return a list of the variable names."""
     description = []
     name_mapping = namemapper.AdaptiveNameMapper('ramses-name-mapping')
     with open(fname) as fd:
@@ -395,16 +373,41 @@ def read_descriptor(fname):
 
 
 class RamsesSnap(SimSnap):
+    """Implements loading of Ramses snapshots.
+
+    Note that AMR cells are loaded as particles, although this works surprisingly well for most analysis purposes.
+    In particular SPH image generation routines automatically switch on an appropriate de-noising scheme to provide
+    effective interpolation between cells.
+    """
     reader_pool = None
 
-    def __init__(self, dirname, **kwargs):
-        """Initialize a RamsesSnap. Extra kwargs supported:
+    def __init__(self, dirname, cpus=None, maxlevel=None, with_gas=True, force_gas=False, times_are_proper=None):
+        """
+        Initialize a RamsesSnap.
 
-         *cpus* : a list of the CPU IDs to load. If not set, load all CPU's data.
-         *maxlevel* : the maximum refinement level to load. If not set, the deepest level is loaded.
-         *with_gas* : if False, never load any gas cells (particles only) - default is True
-         *force_gas* : if True, load the AMR cells as "gas particles" even if they don't actually contain gas in the run
-         """
+        Parameters
+        ----------
+        dirname : str
+            Directory containing the RAMSES output
+        cpus : list of int, optional
+            A list of the CPU IDs to load. If not set, load all
+            CPU's data.
+        maxlevel : int, optional
+            The maximum refinement level to load. If not set, the
+            deepest level is loaded.
+        with_gas : bool, optional
+            If False, never load any gas cells (particles only).
+            Default is True
+        force_gas : bool, optional
+            If True, load the AMR cells as "gas particles" even if
+            they don't actually contain gas in the run. Default is
+            False.
+        times_are_proper : bool, optional
+            If True, the times in the output are assumed to be proper
+            times. If False, they are assumed to be conformal. If
+            None (default), assume proper for non-cosmological simulations
+            and conformal for cosmological ones.
+        """
 
         global config
         super().__init__()
@@ -416,6 +419,12 @@ class RamsesSnap(SimSnap):
 
         self._timestep_id = _timestep_id(dirname)
         self._filename = dirname
+        if os.path.isdir(dirname):
+            self._dirname = dirname
+        else:
+            self._dirname = os.path.split(dirname)[0]
+        self._translate_array_name = namemapper.AdaptiveNameMapper('ramses-name-mapping')
+
         self._load_sink_data_to_temporary_store()
         self._load_infofile()
         self._load_namelistfile()
@@ -428,19 +437,23 @@ class RamsesSnap(SimSnap):
 
         self._ndim = self._info['ndim']
         self.ncpu = self._info['ncpu']
-        if 'cpus' in kwargs:
-            self._cpus = kwargs['cpus']
+        if cpus is not None:
+            self._cpus = cpus
         else:
             self._cpus = list(range(1, self.ncpu + 1))
-        self._maxlevel = kwargs.get('maxlevel', None)
+        self._maxlevel = maxlevel
 
         type_map = self._count_particles()
 
         has_gas = os.path.exists(
-            self._hydro_filename(1)) or kwargs.get('force_gas', False)
+            self._hydro_filename(1)
+        ) or force_gas
 
-        if not kwargs.get('with_gas',True):
+        if not with_gas:
             has_gas = False
+
+        if times_are_proper is not None:
+            self.times_are_proper = times_are_proper
 
         ngas = self._count_gas_cells() if has_gas else 0
 
@@ -463,10 +476,6 @@ class RamsesSnap(SimSnap):
             self._shared_arrays = True
             if (RamsesSnap.reader_pool is None):
                 RamsesSnap.reader_pool = multiprocessing.Pool(multiprocess_num)
-        elif issue_multiprocess_warning:
-            warnings.warn(
-                "RamsesSnap is configured to use multiple processes, but the posix_ipc module is missing. Reverting to single thread.",
-                RuntimeWarning)
 
     def _load_fluid_descriptors(self):
         types = ["hydro", "grav"]
@@ -475,7 +484,7 @@ class RamsesSnap(SimSnap):
 
         for desc_type, default_block, descriptor_fname in zip(types, default_blocks, descriptors_fnames):
             try:
-                block = read_descriptor(descriptor_fname)
+                block = _read_descriptor(descriptor_fname)
             except (FileNotFoundError, OSError):
                 block = default_block
 
@@ -485,11 +494,12 @@ class RamsesSnap(SimSnap):
         self._rt_blocks = []
         self._rt_blocks_3d = set()
         try:
-            f = open(os.path.join(self._filename, f"info_rt_{self._timestep_id}.txt"))
+            with open(os.path.join(self._filename, f"info_rt_{self._timestep_id}.txt")) as f:
+                lines = f.readlines()
         except OSError:
             return
 
-        self._load_info_from_specified_file(f)
+        self._load_info_from_specified_file(lines)
 
         for group in range(self._info['nGroups']):
             for block in rt_blocks:
@@ -500,10 +510,10 @@ class RamsesSnap(SimSnap):
         for block in self._rt_blocks:
             self._rt_blocks_3d.add(self._array_name_1D_to_ND(block) or block)
 
-    def _load_info_from_specified_file(self, f):
-        for l in f:
-            if '=' in l:
-                name, val = list(map(str.strip, l.split('=')))
+    def _load_info_from_specified_file(self, lines):
+        for line in lines:
+            if '=' in line:
+                name, val = list(map(str.strip, line.split('=')))
                 try:
                     if '.' in val:
                         self._info[name] = float(val)
@@ -514,25 +524,26 @@ class RamsesSnap(SimSnap):
 
     def _load_infofile(self):
         self._info = {}
-        f = open(os.path.join(self._filename, f"info_{self._timestep_id}.txt"))
-
-        self._load_info_from_specified_file(f)
+        info_fname = os.path.join(self._filename, f"info_{self._timestep_id}.txt")
+        header_fname = os.path.join(self._filename, f"header_{self._timestep_id}.txt")
+        with open(info_fname) as f:
+            self._load_info_from_specified_file(f)
         try:
-            f = open(os.path.join(self._filename, f"header_{self._timestep_id}.txt"))
             # most of this file is unhelpful, but depending on the ramses
             # version, there may be information on the particle fields present
-            for l in f:
-                if "level" in l:
-                    self._info['particle-blocks'] = l.split()
+            with open(header_fname) as f:
+                for line in f:
+                    if "level" in line:
+                        self._info['particle-blocks'] = line.split()
         except OSError:
             warnings.warn(
                 "No header file found -- no particle block information available")
 
     def _load_namelist_from_specified_file(self, f):
-        for l in f:
-            l = l.split("!")[0]  # remove fortran comments
-            if '=' in l:
-                name, val = map(str.strip, l.split('='))
+        for line in f:
+            line = line.split("!")[0]  # remove fortran comments
+            if '=' in line:
+                name, val = map(str.strip, line.split('='))
                 if val == ".true.":
                     self._namelist[name] = True
                 elif val == ".false.":
@@ -552,11 +563,11 @@ class RamsesSnap(SimSnap):
         namelist_file = os.path.join(self._filename, "namelist.txt")
 
         if os.path.exists(namelist_file):
-            f = open(namelist_file)
-            try:
-                self._load_namelist_from_specified_file(f)
-            except ValueError:
-                warnings.warn("Namelist found but unable to read.")
+            with open(namelist_file) as f:
+                try:
+                    self._load_namelist_from_specified_file(f)
+                except ValueError:
+                    warnings.warn("Namelist found but unable to read.")
 
         else:
             warnings.warn("No namelist file found.")
@@ -572,10 +583,9 @@ class RamsesSnap(SimSnap):
         with open(os.path.join(self._filename, "part_file_descriptor.txt")) as f:
             self._particle_blocks = []
             self._particle_types = []
-            self._translate_array_name = namemapper.AdaptiveNameMapper('ramses-name-mapping')
-            for l in f:
-                if not l.startswith("#"):
-                    ivar, name, dtype = list(map(str.strip,l.split(",")))
+            for line in f:
+                if not line.startswith("#"):
+                    ivar, name, dtype = list(map(str.strip,line.split(",")))
                     self._particle_blocks.append(self._translate_array_name(name, reverse=True))
                     self._particle_types.append(dtype)
             self._particle_blocks_are_explictly_known = True
@@ -622,7 +632,7 @@ class RamsesSnap(SimSnap):
             return self._count_particles_using_explicit_families()
         else:
             ndm, nstar = self._count_particles_using_implicit_families()
-            return OrderedDict([(family.dm, ndm), (family.star, nstar)])
+            return {family.dm: ndm, family.star: nstar}
 
     def _has_particle_file(self):
         """Check whether the output has a particle file available"""
@@ -634,7 +644,7 @@ class RamsesSnap(SimSnap):
     def _count_particles_using_explicit_families(self):
         """Returns an ordered dictionary of family types based on the new explicit RAMSES particle file format"""
         if not self._has_particle_file():
-            return OrderedDict()
+            return {}
         family_block = self._particle_blocks.index('family')
         family_dtype = self._particle_types[family_block]
         self._particle_family_ids_on_disk = []
@@ -684,7 +694,7 @@ class RamsesSnap(SimSnap):
         for fid in self._particle_family_ids_on_disk:
             fid[:] = ramses_id_to_internal_id[fid]
 
-        return_d = OrderedDict()
+        return_d = {}
         self._particle_file_start_indices = [ [] for x in results]
         for internal_family_id in range(256):
             if aggregate_counts_remapped[internal_family_id]>0:
@@ -733,7 +743,8 @@ class RamsesSnap(SimSnap):
         column_names[0] = column_names[0][1:].strip()
         dimensions[0] = dimensions[0][1:].strip()
 
-        self._sink_column_names = column_names
+        self._sink_column_names = [self._translate_array_name(c, reverse=True)
+                                   for c in column_names]
         self._sink_dimensions = dimensions
         self._sink_data = data
 
@@ -852,7 +863,6 @@ class RamsesSnap(SimSnap):
             yield from _cpui_level_iterator(*self._cpui_level_iterator_args(cpu))
 
     def _load_gas_pos(self):
-        i0 = 0
         self.gas['pos'].set_default_units()
         smooth = self.gas['smooth']
         smooth.set_default_units()
@@ -869,8 +879,6 @@ class RamsesSnap(SimSnap):
                    self._cpui_level_iterator_args())
 
     def _load_gas_vars(self, mode=_gv_load_hydro):
-        i1 = 0
-
         dims = []
 
         for i in [self._hydro_blocks, self._grav_blocks, self._rt_blocks][mode]:
@@ -883,7 +891,6 @@ class RamsesSnap(SimSnap):
             dims.append(self.gas[i])
             self.gas[i].set_default_units()
 
-
         if not os.path.exists(self._hydro_filename(1)):
             #Case where force_gas = True, make sure rho is non-zero and such that mass=1.
             # This does not keep track of units for mass or rho since their value is enforced.
@@ -893,10 +900,7 @@ class RamsesSnap(SimSnap):
             self.gas['rho'].set_default_units()
 
 
-
-        nvar = len(dims)
-
-        grid_info_iter = self._level_iterator()
+        _grid_info_iter = self._level_iterator()
 
         logger.info("Loading %s files", ['hydro', 'grav', 'rt'][mode])
 
@@ -952,7 +956,7 @@ class RamsesSnap(SimSnap):
                        [_type] * len(self._cpus),
                        self._particle_family_ids_on_disk
                        )
-        except:
+        except Exception:
             warnings.warn("Exception encountered while reading %r; is there an incompatibility in your Ramses configuration?"%blockname)
             del self[blockname]
             raise
@@ -1014,6 +1018,14 @@ class RamsesSnap(SimSnap):
             keys_ND.add(self._array_name_1D_to_ND(key) or key)
         return list(keys_ND)
 
+    @property
+    def is_cosmological(self):
+        return not self._not_cosmological()
+
+    @property
+    def is_not_cosmological(self):
+        return self._not_cosmological()
+
     def _not_cosmological(self):
         not_cosmological = True
 
@@ -1032,24 +1044,60 @@ class RamsesSnap(SimSnap):
         self.star['tform_raw'] = self.star['tform']
         self.star['tform_raw'].units = self._file_units_system[1]
 
-        if self._is_using_proper_time:
-            t0 = analysis.cosmology.age(self, z=0.0, unit="Gyr")
-            birth_time = t0 + self.s["tform_raw"].in_units("Gyr") / self.properties["a"] ** 2
-            birth_time[birth_time > t0] = t0 - 1e-7
-            self.star['tform'] = birth_time
-        elif not self._not_cosmological():
+        if self.is_cosmological:
             # Only attempt tform conversion for cosmological runs. The built-in tforms for isolated runs
             # are actually meaningful (issue 554)
             from ..analysis import ramses_util
 
             # Replace the tform array by its usual meaning using the birth files
-            ramses_util.get_tform(self)
+            ramses_util.get_tform(self, times_are_proper=self.times_are_proper)
 
-    def _read_proper_time(self):
-        try:
-            self._is_using_proper_time = config_parser.getboolean("ramses", "proper_time")
-        except:
-            self._is_using_proper_time = False
+
+    @property
+    def times_are_proper(self) -> bool:
+        """Best guess for whether the times in the output (e.g. star formation times) are proper times"""
+        if hasattr(self, "_is_using_proper_time"):
+            # Already set, skipping
+            pass
+        elif config_parser.has_option("ramses", "proper_time"):
+            self.times_are_proper = config_parser.getboolean(
+                "ramses", "proper_time"
+            )
+        elif self.is_not_cosmological:
+            self.times_are_proper = True
+        else:
+            # If we detect an RT file, assume proper time
+            iout = self._timestep_id
+            rt_file_candidates = (
+                os.path.join(
+                    self._dirname,
+                    f"rt_{iout}.out{icpu:05d}",
+                )
+                for icpu in self._cpus
+            )
+            for rt_file in rt_file_candidates:
+                if not os.path.exists(rt_file):
+                    continue
+                choice = "proper"
+                reason = f"one RT file was detected ({rt_file})"
+                self.times_are_proper = True
+                break
+            else:
+                choice = "conformal"
+                reason = "no RT file was found"
+                self.times_are_proper = False
+            warnings.warn(
+                f"Assumed times to be in {choice} units because {reason}. "
+                "If this is incorrect, pass the `times_are_proper` keyword "
+                "argument when loading the dataset, or set the option `proper_time` "
+                "in your .pynbodyrc."
+            )
+
+        return self._is_using_proper_time
+
+    @times_are_proper.setter
+    def times_are_proper(self, value):
+        self._is_using_proper_time = value
 
     def _convert_metal_name(self):
         # Name of ramses metallicity has no 's' at the end, contrary to tipsy and gadget
@@ -1068,14 +1116,13 @@ class RamsesSnap(SimSnap):
         elif fam is not family.gas and fam is not None:
 
             # Deal with tform for stars that require extra conversion
-            if array_name == 'tform' or array_name == 'tform_raw':
+            if array_name in ('tform', 'tform_raw'):
                 if 'tform' in self._particle_blocks: # Only attempt these conversion if tform is actually on disc (Issue #689)
                     self._load_particle_block('tform')
-                    self._read_proper_time()
                     self._convert_tform()
 
             # Deal with metals for stars that have extra name mapping
-            elif array_name == 'metals' or array_name == 'metal':
+            elif array_name in ('metals', 'metal'):
                 if 'metal' in self._particle_blocks:
                     self._load_particle_block('metal')
                     self._convert_metal_name()
@@ -1108,6 +1155,14 @@ class RamsesSnap(SimSnap):
                 self._load_gas_vars(_gv_load_rt)
                 for u_block in self._rt_blocks_3d:
                     self[fam][u_block].units = self._rt_unit
+            elif array_name == 'mass':
+                # Very special case. If 'mass' has been created already for particle blocks (or sink blocks), and the
+                # user requests it across the whole simulation, the SimSnap framework will issue this request rather
+                # than a request to load across all families (handled below). If we raise an exception, the user
+                # will see a confusing message about mass being unavailable for gas. So instead we redirect this to
+                # a derived array, which will result in a full mass array being returned (good) and a warning about
+                # conjoining derived and non-derived arrays (logical, at least).
+                self._derive_array('mass', fam)
             else:
                 raise OSError("No such array on disk")
         elif fam is None and array_name in ['pos', 'vel']:
@@ -1128,9 +1183,15 @@ class RamsesSnap(SimSnap):
                 self._load_particle_block(name)
 
         elif fam is None and array_name == 'mass':
+            # mass has to be handled separately because it is present as a particle block
+            # but not as a gas block -- there it will be derived. It is also present as a sink block,
+            # which is again handled differently. So somehow we need to bring all of this together in
+            # a way which doesn't baffle users.
+
             self._create_array('mass')
             self._load_particle_block('mass')
             self['mass'].set_default_units()
+
             if len(self.gas) > 0:
                 gasmass = mass(self.gas)
                 gasmass.convert_units(self['mass'].units)
@@ -1149,7 +1210,7 @@ class RamsesSnap(SimSnap):
 
 
 @RamsesSnap.decorator
-def translate_info(sim):
+def _translate_info(sim):
 
     if sim._info['H0']>1e-3:
         sim.properties['a'] = sim._info['aexp']
@@ -1165,26 +1226,28 @@ def translate_info(sim):
 
     sim.properties['boxsize'] = sim._info['boxlen'] * l_unit
 
-    if sim._not_cosmological():
+    if sim.is_not_cosmological:
         sim.properties['time'] = sim._info['time'] * t_unit
     else:
-        sim.properties['time'] = analysis.cosmology.age(
-            sim) * units.Unit('Gyr')
+        sim.properties["time"] = age(sim, unit="Gyr") * units.Unit("Gyr")
 
     sim._file_units_system = [d_unit, t_unit, l_unit]
 
 
-@RamsesSnap.derived_quantity
+@RamsesSnap.derived_array
 def mass(sim):
+    """Gas mass estimated from cell size and density"""
     return sim['rho'] * sim['smooth'] ** 3
 
 
-@RamsesSnap.derived_quantity
+@RamsesSnap.derived_array
 def temp(sim):
-    """ Gas temperature derived from pressure and density """
+    """Gas temperature derived from pressure and density """
+
     # Has to be redefined and rederived here from Ramses native variables
     # to avoid running into circular dependencies with the traditional derived definition
     # Now uses the self-consistent molecular weight field from pynbody (issue 598)
+
     from ..derived import mu
     mu_est = array.SimArray(np.ones(len(sim)), units="1")
     for i in range(5):
