@@ -1,4 +1,5 @@
 import numpy as np
+import pytest
 
 import pynbody
 
@@ -105,6 +106,63 @@ def test_spherical_coordinates_after_coordinate_transform():
         test_spherical_coordinates_arrays()
 
     # And translation + rotation
-    with pynbody.transformation.translate(f, pynbody.array.SimArray([-100, 140, 200], units='kpc')):
-        with f.rotate_x(173):
-            test_spherical_coordinates_arrays()
+    with f.translate(pynbody.array.SimArray([-100, 140, 200], units='kpc')).rotate_x(173):
+        test_spherical_coordinates_arrays()
+
+class ExampleSnap(pynbody.snapshot.SimSnap):
+    pass
+
+@ExampleSnap.derived_array
+def _test_quantity(sim):
+    return sim['input']+2
+
+def test_derived_array_update():
+    f = pynbody.new(10, class_=ExampleSnap)
+    f['input'] = np.arange(0,10)
+    assert (f['_test_quantity']==np.arange(2,12)).all()
+    f['input']*=2
+    assert (f['_test_quantity']==np.arange(2,22,2)).all()
+
+def test_derived_family_array_update():
+    f = pynbody.new(dm=10,star=10, class_=ExampleSnap)
+    f.dm['input'] = np.arange(0,10)
+    assert (f.dm['_test_quantity']==np.arange(2,12)).all()
+    f.dm['input']*=2
+    assert (f.dm['_test_quantity']==np.arange(2,22,2)).all()
+
+def test_derived_family_array_with_nonderived_partner_update():
+    f = pynbody.new(dm=10,star=10, gas=10, class_=ExampleSnap)
+
+    f.star['_test_quantity'] = np.arange(-10,0)
+    f['input'] = np.arange(0,30)
+
+    assert (f.dm['_test_quantity']==np.arange(2,12)).all()
+    assert (f.star['_test_quantity']==np.arange(-10,0)).all()
+
+    assert f.dm['_test_quantity'].derived
+    assert not f.star['_test_quantity'].derived
+
+    f['input']*=2
+
+    assert (f.dm['_test_quantity']==np.arange(2,22,2)).all()
+    assert (f.star['_test_quantity']==np.arange(-10,0)).all()
+
+
+def test_deprecated_derived_quantity():
+    with pytest.warns(DeprecationWarning):
+        @ExampleSnap.derived_quantity
+        def another_test_quantity(sim):
+            return sim['input'] + 2
+
+    with pytest.warns(DeprecationWarning):
+        @ExampleSnap.stable_derived_quantity
+        def another_test_quantity_stable(sim):
+            return sim['input'] + 4
+
+    f = pynbody.new(10, class_=ExampleSnap)
+    f['input'] = np.arange(0,10)
+    assert (f['another_test_quantity']==np.arange(2,12)).all()
+    assert (f['another_test_quantity_stable']==np.arange(4,14)).all()
+    f['input']+=2
+    assert (f['another_test_quantity']==np.arange(4,14)).all()
+    assert (f['another_test_quantity_stable']==np.arange(4,14)).all()
