@@ -1,33 +1,34 @@
 import numpy as np
 import numpy.testing as npt
+import pytest
 
 import pynbody
 
 
 def test_gravity():
-    f = pynbody.load("testdata/g15784.lr.01024")
+    f = pynbody.load("testdata/gasoline_ahf/g15784.lr.01024")
     h = f.halos()
-    pynbody.analysis.angmom.faceon(h[1])
+    pynbody.analysis.angmom.faceon(h[0])
     pro = pynbody.analysis.profile.Profile(
-        h[1], type='equaln', nbins=50, rmin='100 pc', rmax='50 kpc')
+        h[0], type='equaln', nbins=50, rmin='100 pc', rmax='50 kpc')
 
-    v_circ_correct = np.array([  57.19634349,  103.04331454,  132.12411594,  155.65275799,
-        175.52850379,  193.26102536,  209.52418648,  224.3799979 ,
-        237.98666437,  250.73544548,  262.7691209 ,  273.84754088,
-        283.90399116,  293.11826051,  301.35145207,  308.54834075,
-        314.80692238,  320.03311211,  324.21239628,  327.41041566,
-        329.65615964,  330.93911797,  331.26282888,  330.58067058,
-        328.98176969,  326.54392634,  323.59673278,  321.05292019,
-        318.70939336,  316.17958584,  312.99775752,  308.6537309 ,
-        303.85807775,  299.75311891,  296.75429822,  290.89777773,
-        283.85313613,  277.87270183,  271.31168303,  267.90725824,
-        263.91617477,  259.83215657,  256.09476119,  251.42881956,
-        247.00831007,  244.89573797,  240.35616032,  238.19548506,
-        233.91552211,  229.68799203])
+    v_circ_correct = np.array([ 57.19445406, 103.07628964, 132.11223368, 155.58932808,
+          175.48202525, 193.19929044, 209.50654738, 224.36527055,
+          237.93659247, 250.68636441, 262.72949256, 273.84865161,
+          283.90032809, 293.10420183, 301.32998659, 308.540641  ,
+          314.80948167, 320.0281338 , 324.2028914 , 327.40124838,
+          329.64625769, 330.92733193, 331.25336332, 330.57151552,
+          328.97485235, 326.54886075, 323.60357994, 321.03602331,
+          318.69193219, 316.16834582, 312.98912943, 308.66044926,
+          303.84896218, 299.75583821, 296.77125812, 290.91251525,
+          283.84275867, 277.87033036, 271.31481248, 267.90817792,
+          263.89271744, 259.83899922, 256.07340296, 251.42403094,
+          247.00347206, 244.89404984, 240.3383157 , 238.17681295,
+          233.90977942, 229.66882597])
 
     v_circ = pro['v_circ'].in_units('km s^-1')
 
-    npt.assert_allclose(v_circ, v_circ_correct,atol=1e-5)
+    npt.assert_allclose(v_circ, v_circ_correct, rtol=1e-2, atol=0)
 
 def test_gravity_float():
     f = pynbody.new(100)
@@ -40,9 +41,8 @@ def test_gravity_float():
     f['mass'] = np.ones(100,dtype=np.float32)
     pynbody.gravity.calc.all_direct(f)
 
-
 def test_eps_retrieval_str():
-    f = pynbody.load("testdata/test_g2_snap.0")
+    f = pynbody.load("testdata/gadget2/test_g2_snap.0")
     f.properties['eps'] = "0.3 kpc"
     pynbody.gravity.calc.all_direct(f)
     true_phi_10 = np.array([-0.06696571, -0.07087147, -0.07049192,
@@ -52,7 +52,7 @@ def test_eps_retrieval_str():
 
 
 def test_eps_retrieval_unit():
-    f = pynbody.load("testdata/test_g2_snap.0")
+    f = pynbody.load("testdata/gadget2/test_g2_snap.0")
     f.properties['eps'] = 0.3 * pynbody.units.kpc
     pynbody.gravity.calc.all_direct(f)
     true_phi_10 = np.array([-0.06696571, -0.07087147, -0.07049192,
@@ -62,7 +62,7 @@ def test_eps_retrieval_unit():
 
 
 def test_eps_retrieval_number():
-    f = pynbody.load("testdata/test_g2_snap.0")
+    f = pynbody.load("testdata/gadget2/test_g2_snap.0")
     f.properties['eps'] = 0.3
     pynbody.gravity.calc.all_direct(f)
     true_phi_10 = np.array([-0.06696571, -0.07087147, -0.07049192,
