@@ -1,3 +1,5 @@
+import threading
+import time
 import weakref
 
 import numpy as np
@@ -176,7 +178,7 @@ def test_copy():
     # copied in from IndexedSimArrays
 
 
-def test_mean_by_mass():
+def test_mean_by_mass_units():
     f['pos'].units = 'kpc'
     f['mass'].units = 'Msol'
 
@@ -253,6 +255,51 @@ def test_nd_array_slicing():
     assert not f._array_name_implies_ND_slice('rho_x')
 
 def test_index_list():
-    f = pynbody.load("testdata/g15784.lr.01024")
+    f = pynbody.load("testdata/gasoline_ahf/g15784.lr.01024")
     h = f.halos()
     index_list = h[1].get_index_list(f)
+
+def test_mean_by_mass_value():
+    f = pynbody.new(dm=2000)
+    f['mass'] = np.ones(2000)
+    f['mass'][:500] = 2
+
+    f['test_array'] = np.arange(2000)
+
+    f = f[np.arange(1000)] # test on a subsnap for generality
+
+    np.testing.assert_allclose(f.mean_by_mass('test_array'),
+                               (f['test_array'][:500]*2./3 + f['test_array'][500:1000]*1./3).mean())
+
+
+class SlowLoadingSnapshot(pynbody.snapshot.simsnap.SimSnap):
+    def __init__(self, *args, **kwargs):
+        super().__init__(*args, **kwargs)
+        self._has_entered = False
+        self._has_exited = False
+        self._filename = "test"
+
+    def _load_array(self, array_name, fam=None):
+        if self._has_entered:
+            raise RuntimeError("Entered twice")
+        if self._has_exited:
+            raise RuntimeError("Entered after exiting")
+        self._has_entered=True
+        time.sleep(0.1)
+        self[array_name] = np.arange(len(self))
+        self._has_entered=False
+        self._has_exited=True
+
+def test_race_condition():
+    """There was a race condition where if multiple threads tried to load the same array at once, strange errors
+    would result. This has been fixed by a lock on deriving/loading data."""
+    f = SlowLoadingSnapshot()
+
+    def load_array():
+        f['anyarray'] # noqa
+
+    threads = [threading.Thread(target=load_array) for i in range(5)]
+    for t in threads:
+        t.start()
+    for t in threads:
+        t.join()
