@@ -8,21 +8,19 @@ import pytest
 import pynbody
 
 
-def setup_module():
-    global f, original
-
-    f = pynbody.new(dm=1000)
-    f['pos'] = np.random.normal(scale=1.0, size=f['pos'].shape)
-    f['vel'] = np.random.normal(scale=1.0, size=f['vel'].shape)
-    f['mass'] = np.random.uniform(1.0, 10.0, size=f['mass'].shape)
-
-    original = copy.deepcopy(f)
+@pytest.fixture
+def test_simulation_with_copy():
+    s = pynbody.new(dm=1000)
+    s['pos'] = np.random.normal(scale=1.0, size=s['pos'].shape)
+    s['vel'] = np.random.normal(scale=1.0, size=s['vel'].shape)
+    s['mass'] = np.random.uniform(1.0, 10.0, size=s['mass'].shape)
+    return s, copy.deepcopy(s)
 
 
-def test_translate():
-    global f, original
+def test_translate(test_simulation_with_copy):
+    f, original = test_simulation_with_copy
 
-    with pynbody.transformation.translate(f, [1, 0, 0]):
+    with f.translate([1, 0, 0]):
         npt.assert_almost_equal(f['pos'], original['pos'] + [1, 0, 0])
 
     # check moved back
@@ -30,7 +28,7 @@ def test_translate():
 
     # try again with with abnormal exit
     try:
-        with pynbody.transformation.translate(f, [1, 0, 0]):
+        with f.translate([1, 0, 0]):
             npt.assert_almost_equal(f['pos'], original['pos'] + [1, 0, 0])
             raise RuntimeError
     except RuntimeError:
@@ -39,10 +37,10 @@ def test_translate():
     npt.assert_almost_equal(f['pos'], original['pos'])
 
 
-def test_v_translate():
-    global f, original
+def test_v_translate(test_simulation_with_copy):
+    f, original = test_simulation_with_copy
 
-    with pynbody.transformation.v_translate(f, [1, 0, 0]):
+    with f.offset_velocity([1, 0, 0]):
         npt.assert_almost_equal(f['vel'], original['vel'] + [1, 0, 0])
 
     # check moved back
@@ -50,7 +48,7 @@ def test_v_translate():
 
     # try again with with abnormal exit
     try:
-        with pynbody.transformation.v_translate(f, [1, 0, 0]):
+        with f.offset_velocity([1, 0, 0]):
             npt.assert_almost_equal(f['vel'], original['vel'] + [1, 0, 0])
             raise RuntimeError
     except RuntimeError:
@@ -59,10 +57,10 @@ def test_v_translate():
     npt.assert_almost_equal(f['vel'], original['vel'])
 
 
-def test_vp_translate():
-    global f, original
+def test_vp_translate(test_simulation_with_copy):
+    f, original = test_simulation_with_copy
 
-    with pynbody.transformation.xv_translate(f, [1, 0, 0], [2, 0, 0]):
+    with f.translate([1, 0, 0]).offset_velocity([2,0,0]):
         npt.assert_almost_equal(f['vel'], original['vel'] + [2, 0, 0])
         npt.assert_almost_equal(f['pos'], original['pos'] + [1, 0, 0])
 
@@ -72,7 +70,7 @@ def test_vp_translate():
 
     # try again with with abnormal exit
     try:
-        with pynbody.transformation.xv_translate(f, [1, 0, 0], [2, 0, 0]):
+        with f.translate([1, 0, 0]).offset_velocity([2,0,0]):
             npt.assert_almost_equal(f['vel'], original['vel'] + [2, 0, 0])
             npt.assert_almost_equal(f['pos'], original['pos'] + [1, 0, 0])
             raise RuntimeError
@@ -83,8 +81,8 @@ def test_vp_translate():
     npt.assert_almost_equal(f['pos'], original['pos'])
 
 
-def test_rotate():
-    global f, original
+def test_rotate(test_simulation_with_copy):
+    f, original = test_simulation_with_copy
 
     with f.rotate_x(90):
         npt.assert_almost_equal(f['y'], -original['z'])
@@ -110,28 +108,71 @@ def test_family_rotate():
     npt.assert_almost_equal(f['pos'][:, 1], 1.0)
     npt.assert_almost_equal(f.bh['vel'][:, 1], 1.0)
 
-def test_chaining():
-    with pynbody.transformation.translate(f.rotate_x(90), [0, 1, 0]):
+def test_chaining(test_simulation_with_copy):
+    f, original = test_simulation_with_copy
+
+    with f.rotate_x(90).translate([0, 1, 0]):
         npt.assert_almost_equal(f['y'], 1.0 - original['z'])
         npt.assert_almost_equal(f['z'], original['y'])
 
     npt.assert_almost_equal(f['pos'], original['pos'])
 
 
-def test_halo_managers():
-    with pynbody.analysis.angmom.sideon(f, disk_size=1, cen_size=1):
+def test_halo_managers(test_simulation_with_copy):
+    f, original = test_simulation_with_copy
+
+    with pynbody.analysis.angmom.sideon(f, disk_size=1.0):
         pass
 
     npt.assert_almost_equal(f['pos'], original['pos'])
 
+def test_repr(test_simulation_with_copy):
+    f, _ = test_simulation_with_copy
+
+    with f.rotate_x(45).translate([0, 1, 0]).offset_velocity([0, 0, 1]) as tx:
+        assert str(tx) == "rotate_x(45), translate, offset_velocity"
+        assert repr(tx) == "<Transformation rotate_x(45), translate, offset_velocity>"
+
+def test_null(test_simulation_with_copy):
+    f, original = test_simulation_with_copy
+
+    with pynbody.transformation.NullTransformation(f) as tx:
+        npt.assert_almost_equal(f['pos'], original['pos'])
+        assert tx.sim is not None
+
+    with tx.rotate_x(90) as tx2:
+        assert tx2.next_transformation is None
+        assert str(tx2) == "rotate_x(90)"
+        npt.assert_almost_equal(f['y'], -original['z'])
+
+
+    npt.assert_almost_equal(f['pos'], original['pos'])
 
 def test_weakref():
-    global f
+    f = pynbody.new(dm=10)
+
     tx1 = f.rotate_y(90)
-    tx2 = pynbody.transformation.translate(f.rotate_x(90), [0, 1, 0])
+    tx2 = tx1.rotate_x(90).translate([0, 1, 0])
     assert tx1.sim is not None
     assert tx2.sim is not None
     del f
     gc.collect()
     assert tx1.sim is None
     assert tx2.sim is None
+
+@pytest.mark.parametrize("family", [True, False])
+def test_derived_3d_array(family):
+    """Test for a bug where transformations would try to rotate derived arrays, raising an error"""
+    f = pynbody.new(dm=5, gas=5)
+
+    @pynbody.derived_array
+    def test_derived_3d_array_transformation(sim):
+        return np.random.normal(size=(len(sim), 3))
+
+    if family:
+        f = f.dm
+
+    _ = f['test_derived_3d_array_transformation']
+
+    with f.ancestor.rotate_y(90):
+        _ = f['test_derived_3d_array_transformation']
